### 2024-09-26 #7292 @editorialbot
[PRE REVIEW]: EKF_CAL: Extended Kalman Filter-based Calibration and Localization
**Submitting author:** @JHartzer (<a href="http://orcid.org/0000-0002-3051-2213">Jacob Hartzer</a>)
**Repository:** https://github.com/unmannedlab/ekf_cal/
**Branch with paper.md** (empty if default branch):
**Version:** 0.5
