### 2024-09-26 #7292 @kyleniemeyer
Hi @JHartzer, we will use this pre-review issue to find an editor and reviewers, and resolve any issues.

Although I can see that your software package meets most of our requirements (documentation, installation instructions, etc.), I'm not sure about a test suite. At minimum, we need instructions on manually testing the full functionality of the software; normally, we want to see—and reviewers will request—an automated test suite for unit and functional testing. (And at best, hooked into continuous integration.)
