+++ eval/tab_fiducial.py
@@ -21,8 +21,8 @@ from bokeh.layouts import layout
 from bokeh.models import Band, Spacer, TabPanel
 from bokeh.plotting import ColumnDataSource, figure
 import numpy as np
-from utilities import calculate_alpha, get_colors, plot_update_timing
 from scipy.spatial.transform import Rotation
+from utilities import calculate_alpha, get_colors, plot_update_timing
 
 
 class tab_fiducial:
