+++ docs/parameters.md
@@ -1,49 +1,11 @@
-Getting Started {#tutorial}
+Parameter Definitions {#parameters}
 ============
-
-
-# Dependencies
-The ekf_cal package has the following hard dependencies that are required for all compilations:
-- [OpenCV](https://opencv.org/)
-- [Eigen 3](https://eigen.tuxfamily.org/index.php?title=Main_Page)
-
-The following dependencies are for building the ROS node and simulation, respectively
-- [ROS2](https://docs.ros.org/en/rolling/index.html)
-- [yaml-cpp](https://github.com/jbeder/yaml-cpp)
-
-The following soft dependencies useful for development and documentation
-- [Doxygen](https://www.doxygen.nl/index.html)
-- [Google Test](https://google.github.io/googletest/)
-
-These can be installed by running `rosdep` in the base directory of this repo
-```
-rosdep install --from-paths src -y --ignore-src
-```
-
-# Build
-
-Building can be done simply with the following command:
-
-```
-colcon build --symlink-install --packages-select ekf_cal
-```
-
-## Docker
-Alternatively, a Dockerfile is provided, which can be used either inside a VS Code [devcontainer](https://code.visualstudio.com/docs/devcontainers/containers), or a standalone container.
-
-# Input Files
-
-This repository offers two main ways to utilize the Kalman filter framework: a simulation and ROS2 node.
-Both the simulation and ROS node are configurable and runnable using identically formatted YAML
-files. The basic structure of the YAML files is a set of ROS parameters, lists that define the
-sensors to use, and the sensor definitions themselves. The top-level of the YAML is the definition
-of the node name `/EkfCalNode` and a ROS2-specific implementation for `ros__parameters`. Structuring
-the yaml in this way allows the node to utilize these parameters. The `sim_params` section includes
-parameters that are only relevant to the simulation.
 
 ## Main parameters
 
-```YAML
+The basic structure of the YAML files is a set of ROS parameters, lists that define the sensors to use, and the sensor definitions themselves. The top-level of the YAML is the definition of the node name `/EkfCalNode` and a ROS2-specific implementation for `ros__parameters`. Structuring the yaml in this way allows the node to utilize these parameters. The `sim_params` section includes parameters that are only relevant to the simulation. The main simulation and node parameters are as follows
+
+```{.py}
 /EkfCalNode:
     ros__parameters:
         debug_log_level: 2                  # Debug_Log_Level:
@@ -96,7 +58,7 @@ must read from a known list to the declare and read unknown lists. Therefore, th
 must exist, but can be empty. These define the parameter sections for each of their respective
 categories.
 
-```YAML
+```{.py}
 /EkfCalNode:
     ros__parameters:
 
@@ -122,7 +84,7 @@ categories.
 
 The following is an example of an IMU input configuration.
 
-```YAML
+```{.py}
 /EkfCalNode:
     ros__parameters:
 
@@ -165,7 +127,7 @@ The following is an example of an IMU input configuration.
 
 The following is an example of an camera input configuration.
 
-```YAML
+```{.py}
 /EkfCalNode:
     ros__parameters:
 
@@ -207,7 +169,7 @@ The following is an example of an camera input configuration.
 
 The following is an example of a tracker input configuration.
 
-```YAML
+```{.py}
 /EkfCalNode:
     ros__parameters:
 
@@ -242,7 +204,7 @@ The following is an example of a tracker input configuration.
 
 The following is an example of a fiducial input configuration.
 
-```YAML
+```{.py}
 /EkfCalNode:
     ros__parameters:
 
@@ -278,7 +240,7 @@ The following is an example of a fiducial input configuration.
 
[…]
