+++ docs/main.md
@@ -1,20 +1,124 @@
 # ekf_cal
 [![Documentation](https://github.com/unmannedlab/ekf_cal/actions/workflows/documentation.yaml/badge.svg)](https://github.com/unmannedlab/ekf_cal/actions/workflows/documentation.yaml)
 
-Extended Kalman Filter Calibration and Localization: ekf_cal is a package focused on the simulation and development of a multi-sensor online calibration kalman filter. It combines the architecture of a Multi-State Constraint Kalman Filter (MSCKF) with a multi-IMU calibration filter to provide intrinsic and extrinsic estimates for the following:
+Extended Kalman Filter Calibration and Localization: ekf_cal is a package focused on the simulation and development of a multi-sensor online calibration kalman filter. It combines the architecture of a Multi-State Constraint Kalman Filter (MSCKF) with a multi-IMU calibration filter to provide intrinsic and extrinsic estimates for the following sensors:
 - [IMU](docs/filters/imu.md)
 - [GPS](docs/filters/gps.md)
 - [Cameras](docs/filters/camera.md)
 - [Fiducials](docs/filters/fiducial.md)
 
-<img src="images/setup.svg" alt="setup" width="70%"/>
+<img src="images/setup.svg" alt="setup" width="80%"/>
+
+## Quick Start
+
+### Clone the Repository
+
+This guide assumes you have the `ekf_cal` repository in a colcon workspace.
+```
+mkdir ekf_cal_ws/
+mkdir ekf_cal_ws/src/
+cd ekf_cal_ws/src/
+git clone git@github.com:unmannedlab/ekf_cal.git
+cd ../
+```
+
+### Dependencies
+The ekf_cal package has the following hard dependencies that are required for all compilations:
+- [OpenCV](https://opencv.org/)
+- [Eigen 3](https://eigen.tuxfamily.org/index.php?title=Main_Page)
+
+The following dependencies are for building the ROS node and simulation, respectively
+- [ROS2](https://docs.ros.org/en/rolling/index.html)
+- [yaml-cpp](https://github.com/jbeder/yaml-cpp)
+
+The following soft dependencies useful for development and documentation
+- [Doxygen](https://www.doxygen.nl/index.html)
+- [Google Test](https://google.github.io/googletest/)
+
+These can be installed by running [rosdep](https://wiki.ros.org/rosdep) in the base directory of the colcon workspace (e.g. `ekf_cal_ws`)
+```
+rosdep install --from-paths src -y --ignore-src
+```
+
+### Build
+
+Building can be done simply with the following command:
+
+```
+colcon build --symlink-install --packages-select ekf_cal --event-handlers console_cohesion+ --cmake-args -DCMAKE_C_FLAGS='--coverage' -DCMAKE_CXX_FLAGS='--coverage'
+```
+
+#### Docker
+Alternatively, a Dockerfile is provided, which can be used either inside a VS Code [devcontainer](https://code.visualstudio.com/docs/devcontainers/containers), or a standalone container.
+
+### Input Files
+
+This repository offers two main ways to utilize the Kalman filter framework: a simulation and ROS2 node. Both the simulation and ROS node are configurable and runnable using identically formatted YAML files. Further documentation can be found on the [Parameters](docs/parameters.md) page.
+
+### Simulation
+
+Simulations can be run using a YAML configuration file that extends the base configuration file with additional parameters. See the example [example.yaml](../../../config/example.yaml). Multiple simulations can be run in parallel using the [run.py](../../../eval/run.py). An example using a single input is given below
+
+```
+python3 eval/run.py config/example.yaml
+```
+
+The results of a run can be plotted using [report.py](../../../eval/report.py)
+```
+python3 eval/report.py config/example.yaml
+```
+
[…]
