### 2024-09-27 320ec27
Add cloc and flamegraph instructions
