+++ docs/flamegraph/run_perf.sh
@@ -1,4 +1,4 @@
-../../../../build/ekf_cal/sim perf.yaml perf/ & 
+../../../../build/ekf_cal/sim perf.yaml perf/ &
 perf record -F 99 -p $! --call-graph dwarf -- sleep 60
 perf script > sim.perf
 ./stackcollapse-perf.pl sim.perf > sim.folded 
