+++ docs/software.md
@@ -1,6 +1,7 @@
 Software {#software}
 ============
 
+These software evaluation pages are not automatically generated (yet) and require some manual commands to produce the relevant information. See the mainpage for more information.
 - @subpage cloc
 - @subpage flamegraph
 - @subpage code_coverage
