+++ docs/filters.md
@@ -1,6 +1,8 @@
 Filter Design {#filter}
 ============
 
+`ekf_cal` combines the architecture of a Multi-State Constraint Kalman Filter (MSCKF) with a multi-sensor calibration filter to provide intrinsic and extrinsic estimates for the following sensors:
+
 - @subpage imu
 - @subpage gps
 - @subpage camera
