+++ src/ekf/update/test/imu_updater_test.cpp
@@ -116,7 +116,7 @@ TEST(test_imu_updater, update) {
   EXPECT_EQ(state.body_state.pos_b_in_l[1], 4);
   EXPECT_EQ(state.body_state.pos_b_in_l[2], 4);
 
-  EXPECT_NEAR(state.body_state.vel_b_in_l[0], 10.4139, 1e-3);
+  EXPECT_NEAR(state.body_state.vel_b_in_l[0], 10.4139, 1e-2);
   EXPECT_EQ(state.body_state.vel_b_in_l[1], 1);
   EXPECT_EQ(state.body_state.vel_b_in_l[2], 1);
 }
@@ -188,11 +188,11 @@ TEST(test_imu_updater, imu_prediction_update) {
     ekf, time, acceleration, acceleration_cov, angular_rate, angular_rate_cov, use_for_prediction);
 
   state = ekf->m_state;
-  EXPECT_EQ(state.body_state.pos_b_in_l[0], 3);
+  EXPECT_EQ(state.body_state.pos_b_in_l[0], 8);
   EXPECT_EQ(state.body_state.pos_b_in_l[1], 3);
   EXPECT_EQ(state.body_state.pos_b_in_l[2], 3);
 
-  EXPECT_EQ(state.body_state.vel_b_in_l[0], 1);
+  EXPECT_EQ(state.body_state.vel_b_in_l[0], 11);
   EXPECT_EQ(state.body_state.vel_b_in_l[1], 1);
   EXPECT_EQ(state.body_state.vel_b_in_l[2], 1);
 
@@ -201,11 +201,11 @@ TEST(test_imu_updater, imu_prediction_update) {
     ekf, time, acceleration, acceleration_cov, angular_rate, angular_rate_cov, use_for_prediction);
 
   state = ekf->m_state;
-  EXPECT_EQ(state.body_state.pos_b_in_l[0], 4);
+  EXPECT_EQ(state.body_state.pos_b_in_l[0], 24);
   EXPECT_EQ(state.body_state.pos_b_in_l[1], 4);
   EXPECT_EQ(state.body_state.pos_b_in_l[2], 4);
 
-  EXPECT_EQ(state.body_state.vel_b_in_l[0], 11);
+  EXPECT_EQ(state.body_state.vel_b_in_l[0], 21);
   EXPECT_EQ(state.body_state.vel_b_in_l[1], 1);
   EXPECT_EQ(state.body_state.vel_b_in_l[2], 1);
 }
