### 2024-09-28 a8ef1c3
Add main page acceleration plots
