+++ eval/theme-dark.yaml
@@ -1,0 +1,51 @@
+attrs:
+    Plot:
+        background_fill_color: '#20262B'
+        # border_fill_color: '#15191C'
+        border_fill_color: '#1C1D1F'
+        outline_line_color: '#E0E0E0'
+        outline_line_alpha: 0.25
+    Grid:
+        grid_line_color: '#E0E0E0'
+        grid_line_alpha: 0.25
+    Axis:
+        major_tick_line_alpha: 0
+        major_tick_line_color: '#E0E0E0'
+        minor_tick_line_alpha: 0
+        minor_tick_line_color: '#E0E0E0'
+        axis_line_alpha: 0
+        axis_line_color: '#E0E0E0'
+        major_label_text_color: '#E0E0E0'
+        major_label_text_font: 'Helvetica'
+        major_label_text_font_size: '1.025em'
+        axis_label_standoff: 10
+        axis_label_text_color: '#E0E0E0'
+        axis_label_text_font: 'Helvetica'
+        axis_label_text_font_size: '1.25em'
+        axis_label_text_font_style: 'normal'
+    Legend:
+        spacing: 8
+        glyph_width: 15
+        label_standoff: 8
+        label_text_color: '#E0E0E0'
+        label_text_font: 'Helvetica'
+        label_text_font_size: '1.025em'
+        border_line_alpha: 0
+        background_fill_alpha: 0.25
+        background_fill_color: '#20262B'
+    BaseColorBar:
+        title_text_color: '#E0E0E0'
+        title_text_font: 'Helvetica'
+        title_text_font_size: '1.025em'
+        title_text_font_style: 'normal'
+        major_label_text_color: '#E0E0E0'
+        major_label_text_font: 'Helvetica'
+        major_label_text_font_size: '1.025em'
+        # background_fill_color: '#15191C'
+        background_fill_color: '#15191C'
+        major_tick_line_alpha: 0
+        bar_line_alpha: 0
+    Title:
+        text_color: '#E0E0E0'
+        text_font: 'Helvetica'
+        text_font_size: '1.15em'
