### 2024-09-28 #7292 @JHartzer
Hello @kyleniemeyer! Thank you for the initial feedback. I do have a test suite (that is not yet automated) for code testing and coverage. I've updated the main README and documentation landing to hopefully make that a little more clear. I can also take an action to work on a CI pipeline.

As far as  reviewer suggestions, here is a list of possibilities from similar authors and a keyword search:
- AKuederle
- stevemacenski
- abougouffa
- mhu48
- Mechazo11
