+++ src/utility/math_helper.cpp
@@ -421,3 +421,15 @@ double limit_matrix_condition(Eigen::MatrixXd & mat)
   }
   return condition;
 }
+
+Eigen::MatrixXd QR_r(Eigen::MatrixXd A, Eigen::MatrixXd B)
+{
+  Eigen::MatrixXd vert_cat(A.rows() + B.rows(), A.cols());
+
+  vert_cat << A, B;
+  Eigen::HouseholderQR<Eigen::MatrixXd> QR(vert_cat);
+  Eigen::MatrixXd R = QR.matrixQR().block(0, 0, A.cols(), A.cols()).triangularView<Eigen::Upper>();
+  R = R.diagonal().cwiseSign().asDiagonal() * R;
+
+  return R;
+}
