### 2024-10-01 c3ccbe9
Use common EKF updater function
