+++ src/trackers/sim/sim_fiducial_tracker.cpp
@@ -125,19 +125,28 @@ std::shared_ptr<SimFiducialTrackerMessage> SimFiducialTracker::GenerateMessage(
     Eigen::Vector3d pos_f_in_c_true =
       rot_b_to_c * (rot_g_to_b * (pos_f_in_g_true - pos_b_in_g) - pos_c_in_b_true);
 
+    Eigen::Quaterniond ang_f_to_c_true =
+      ang_c_to_b_true.inverse() * ang_b_to_g.inverse() * ang_f_to_g_true;
+
     BoardDetection board_detection;
     board_detection.frame_id = frame_id;
-    board_detection.t_vec_f_in_c[0] = m_rng.NormRand(pos_f_in_c_true[0], m_t_vec_error[0]);
-    board_detection.t_vec_f_in_c[1] = m_rng.NormRand(pos_f_in_c_true[1], m_t_vec_error[1]);
-    board_detection.t_vec_f_in_c[2] = m_rng.NormRand(pos_f_in_c_true[2], m_t_vec_error[2]);
+    if (!m_no_errors) {
+      board_detection.t_vec_f_in_c[0] = m_rng.NormRand(pos_f_in_c_true[0], m_t_vec_error[0]);
+      board_detection.t_vec_f_in_c[1] = m_rng.NormRand(pos_f_in_c_true[1], m_t_vec_error[1]);
+      board_detection.t_vec_f_in_c[2] = m_rng.NormRand(pos_f_in_c_true[2], m_t_vec_error[2]);
 
-    Eigen::Vector3d ang_f_to_c_error_rpy;
-    ang_f_to_c_error_rpy(0) = m_rng.NormRand(0.0, m_r_vec_error[0]);
-    ang_f_to_c_error_rpy(1) = m_rng.NormRand(0.0, m_r_vec_error[1]);
-    ang_f_to_c_error_rpy(2) = m_rng.NormRand(0.0, m_r_vec_error[2]);
-    Eigen::Quaterniond ang_f_to_c = EigVecToQuat(ang_f_to_c_error_rpy) *
-      ang_c_to_b_true.inverse() * ang_b_to_g.inverse() * ang_f_to_g_true;
-    board_detection.r_vec_f_to_c = QuatToRodrigues(ang_f_to_c);
+      Eigen::Vector3d ang_f_to_c_error_rpy;
+      ang_f_to_c_error_rpy(0) = m_rng.NormRand(0.0, m_r_vec_error[0]);
+      ang_f_to_c_error_rpy(1) = m_rng.NormRand(0.0, m_r_vec_error[1]);
+      ang_f_to_c_error_rpy(2) = m_rng.NormRand(0.0, m_r_vec_error[2]);
+      Eigen::Quaterniond ang_f_to_c = EigVecToQuat(ang_f_to_c_error_rpy) * ang_f_to_c_true;
+      board_detection.r_vec_f_to_c = QuatToRodrigues(ang_f_to_c);
+    } else {
+      board_detection.t_vec_f_in_c[0] = pos_f_in_c_true[0];
+      board_detection.t_vec_f_in_c[1] = pos_f_in_c_true[1];
+      board_detection.t_vec_f_in_c[2] = pos_f_in_c_true[2];
+      board_detection.r_vec_f_to_c = QuatToRodrigues(ang_f_to_c_true);
+    }
 
     m_board_track.push_back(board_detection);
   } else if (m_board_track.size() < m_min_track_length) {
