+++ config/example.yaml
@@ -65,8 +65,8 @@
             pos_errors: [1.0, 1.0, 1.0]     # Error in achieving translational vectors
             ang_errors: [0.1, 0.1, 0.1]     # Error in achieving rotational vectors
 
-            pos_l_in_g_err: [1.0, 1.0, 1.0] # Local frame position error
-            ang_l_to_g_err: 0.1             # Local frame heading error
+            pos_l_in_g_error: [1.0, 1.0, 1.0] # Local frame position error
+            ang_l_to_g_error: 0.1             # Local frame heading error
 
         imu_list:       # List of imu. Can be empty but list must exist
             - "imu_1"
