### 2024-10-02 da9a0cd
Use common updater for stationary IMU
