+++ src/ekf/update/imu_updater.cpp
@@ -294,31 +294,16 @@ bool ImuUpdater::ZeroAccelerationUpdate(
   Eigen::Matrix3d angular_rate_covariance)
 {
   /// @todo: Need a cohesive method to handle stationary rotations about the gravity axis
-  return false;
-  ekf->ProcessModel(time);
+  // return false;
 
-  unsigned int sub_size{3}, meas_size{3};
-  unsigned int sub_ex_index{0}, sub_in_index{0};
-  unsigned int ex_index{0}, in_index{0};
+  unsigned int meas_size{3};
+  unsigned int index_extrinsic = ekf->m_state.imu_states[imu_id].index_extrinsic;
+  unsigned int index_intrinsic = ekf->m_state.imu_states[imu_id].index_intrinsic;
 
   auto t_start = std::chrono::high_resolution_clock::now();
 
-  if (ekf->m_state.imu_states[imu_id].GetIsExtrinsic()) {
-    ex_index = ekf->m_state.imu_states[imu_id].index;
-    in_index += 3;
-    sub_ex_index = 3;
-    sub_size += 3;
-  }
-
-  if (ekf->m_state.imu_states[imu_id].GetIsIntrinsic()) {
-    in_index += ekf->m_state.imu_states[imu_id].index;
-    sub_in_index = sub_ex_index + 3;
-    sub_size += 6;
-    meas_size += 3;
-  }
-
-  Eigen::MatrixXd H = Eigen::MatrixXd::Zero(meas_size, sub_size);
-  H.block<3, 3>(0, 0) = -ekf->m_state.imu_states[imu_id].ang_i_to_b.inverse().toRotationMatrix() *
+  Eigen::MatrixXd H = Eigen::MatrixXd::Zero(meas_size, ekf->GetStateSize());
+  H.block<3, 3>(0, 9) = -ekf->m_state.imu_states[imu_id].ang_i_to_b.inverse().toRotationMatrix() *
     SkewSymmetric(ekf->m_state.body_state.ang_b_to_l.inverse() * g_gravity);
 
   Eigen::MatrixXd R = Eigen::MatrixXd::Zero(meas_size, meas_size);
@@ -327,38 +312,29 @@ bool ImuUpdater::ZeroAccelerationUpdate(
   Eigen::Vector3d bias_a = ekf->m_state.imu_states[imu_id].acc_bias;
   Eigen::Vector3d bias_g = ekf->m_state.imu_states[imu_id].omg_bias;
 
-  Eigen::VectorXd z = Eigen::VectorXd::Zero(meas_size);
-  z.segment<3>(0) = -(acceleration - bias_a -
+  Eigen::VectorXd resid = Eigen::VectorXd::Zero(meas_size);
+  resid.segment<3>(0) = -(acceleration - bias_a -
     ekf->m_state.imu_states[imu_id].ang_i_to_b.inverse() *
     ekf->m_state.body_state.ang_b_to_l.inverse() * g_gravity);
 
-  Eigen::MatrixXd sub_cov = Eigen::MatrixXd::Zero(sub_size, sub_size);
-  sub_cov.block<3, 3>(0, 0) = ekf->m_cov.block<3, 3>(6, 6);
-
   if (ekf->m_state.imu_states[imu_id].GetIsExtrinsic()) {
-    H.block<3, 3>(0, 0) = -SkewSymmetric(
+    H.block<3, 3>(0, index_extrinsic) = -SkewSymmetric(
       ekf->m_state.imu_states[imu_id].ang_i_to_b.inverse() *
       ekf->m_state.body_state.ang_b_to_l.inverse() * g_gravity
     );
[…]
