### 2024-10-04 767b494
Prevent stationary updates from rotating heading
