### 2024-10-04 716f7e3
Prevent stationary rotation about the vertical axis
