+++ src/ekf/update/msckf_updater.cpp
@@ -76,8 +76,6 @@ bool MsckfUpdater::TriangulateFeature(
 
   AugState aug_state_0 = ekf->GetAugState(
     m_id, feature_track.track[0].frame_id, feature_track.track[0].frame_time);
-  CamState cam_state = ekf->m_state.cam_states[m_id];
-  Intrinsics intrinsics = ekf->m_state.cam_states[m_id].intrinsics;
 
   // 3D Cartesian Triangulation
   Eigen::Matrix3d A = Eigen::Matrix3d::Zero();
@@ -85,11 +83,10 @@ bool MsckfUpdater::TriangulateFeature(
 
   const Eigen::Vector3d pos_b0_in_l = aug_state_0.pos_b_in_l;
   const Eigen::Matrix3d rot_b0_to_l = aug_state_0.ang_b_to_l.toRotationMatrix();
-  const Eigen::Vector3d pos_c_in_b = cam_state.pos_c_in_b;
-  const Eigen::Matrix3d ang_c_to_b = cam_state.ang_c_to_b.toRotationMatrix();
+  const Eigen::Matrix3d rot_c_to_b = m_ang_c_to_b.toRotationMatrix();
 
   const Eigen::Matrix3d rot_l_to_b0 = rot_b0_to_l.transpose();
-  const Eigen::Matrix3d rot_b0_to_c0 = ang_c_to_b.transpose();
+  const Eigen::Matrix3d rot_b0_to_c0 = rot_c_to_b.transpose();
 
   for (unsigned int i = 0; i < feature_track.track.size(); ++i) {
     AugState aug_state_i = ekf->GetAugState(
@@ -99,17 +96,17 @@ bool MsckfUpdater::TriangulateFeature(
     const Eigen::Matrix3d rot_bi_to_l = aug_state_i.ang_b_to_l.toRotationMatrix();
 
     // Convert current pos relative to anchor
-    Eigen::Matrix3d rot_ci_to_c0 = rot_b0_to_c0 * rot_l_to_b0 * rot_bi_to_l * ang_c_to_b;
+    Eigen::Matrix3d rot_ci_to_c0 = rot_b0_to_c0 * rot_l_to_b0 * rot_bi_to_l * rot_c_to_b;
     Eigen::Vector3d pos_ci_in_c0 = rot_b0_to_c0 * rot_l_to_b0 * (
-      (rot_bi_to_l * pos_c_in_b + pos_bi_in_l) -
-      (rot_b0_to_l * pos_c_in_b + pos_b0_in_l));
+      (rot_bi_to_l * m_pos_c_in_b + pos_bi_in_l) -
+      (rot_b0_to_l * m_pos_c_in_b + pos_b0_in_l));
 
     // Get the UV coordinate normal
     Eigen::Vector3d b_i;
-    b_i(0) = (feature_track.track[i].key_point.pt.x - intrinsics.width / 2) /
-      (intrinsics.f_x / intrinsics.pixel_size);
-    b_i(1) = (feature_track.track[i].key_point.pt.y - intrinsics.height / 2) /
-      (intrinsics.f_y / intrinsics.pixel_size);
+    b_i(0) = (feature_track.track[i].key_point.pt.x - m_intrinsics.width / 2) /
+      (m_intrinsics.f_x / m_intrinsics.pixel_size);
+    b_i(1) = (feature_track.track[i].key_point.pt.y - m_intrinsics.height / 2) /
+      (m_intrinsics.f_y / m_intrinsics.pixel_size);
     b_i(2) = 1;
 
     // Rotate and normalize
@@ -133,7 +130,7 @@ bool MsckfUpdater::TriangulateFeature(
[…]
