+++ src/ekf/update/fiducial_updater.cpp
@@ -86,7 +86,13 @@ void FiducialUpdater::UpdateEKF(
   std::vector<Eigen::Vector3d> pos_f_in_l_vec;
   std::vector<Eigen::Quaterniond> ang_f_to_l_vec;
 
-  CamState cam_state = ekf->m_state.cam_states[m_camera_id];
+  if (!ekf->GetUseFirstEstimateJacobian() || m_is_first_estimate) {
+    m_pos_c_in_b = ekf->m_state.cam_states[m_camera_id].pos_c_in_b;
+    m_ang_c_to_b = ekf->m_state.cam_states[m_camera_id].ang_c_to_b;
+    m_pos_f_in_l = ekf->m_state.fid_states[m_id].pos_f_in_l;
+    m_ang_f_to_l = ekf->m_state.fid_states[m_id].ang_f_to_l;
+    m_is_first_estimate = false;
+  }
 
   /// @todo: Wrap this in a function
   for (auto board_detection : board_track) {
@@ -94,17 +100,16 @@ void FiducialUpdater::UpdateEKF(
 
     const Eigen::Vector3d pos_bi_in_g = aug_state_i.pos_b_in_l;
     const Eigen::Matrix3d rot_bi_to_l = aug_state_i.ang_b_to_l.toRotationMatrix();
-    const Eigen::Vector3d pos_ci_in_bi = cam_state.pos_c_in_b;
-    const Eigen::Matrix3d rot_c_to_b = cam_state.ang_c_to_b.toRotationMatrix();
+    const Eigen::Matrix3d rot_c_to_b = m_ang_c_to_b.toRotationMatrix();
 
     Eigen::Vector3d pos_f_in_c;
     CvVectorToEigen(board_detection.t_vec_f_in_c, pos_f_in_c);
     Eigen::Vector3d pos_f_in_l =
-      rot_bi_to_l * ((rot_c_to_b * pos_f_in_c) + pos_ci_in_bi) + pos_bi_in_g;
+      rot_bi_to_l * ((rot_c_to_b * pos_f_in_c) + m_pos_c_in_b) + pos_bi_in_g;
     pos_f_in_l_vec.push_back(pos_f_in_l);
 
     Eigen::Quaterniond ang_f_to_c = RodriguesToQuat(board_detection.r_vec_f_to_c);
-    Eigen::Quaterniond ang_f_to_l = aug_state_i.ang_b_to_l * cam_state.ang_c_to_b * ang_f_to_c;
+    Eigen::Quaterniond ang_f_to_l = aug_state_i.ang_b_to_l * m_ang_c_to_b * ang_f_to_c;
     ang_f_to_l_vec.push_back(ang_f_to_l);
     pos_weights.push_back(1.0);
     ang_weights.push_back(1.0);
@@ -121,31 +126,7 @@ void FiducialUpdater::UpdateEKF(
     m_board_logger.RateLimitedLog(data_msg.str(), time);
   }
 
-  Eigen::Vector3d pos_f_in_l_est = ekf->m_state.fid_states[m_id].pos_f_in_l;
-  Eigen::Quaterniond ang_f_to_l_est = ekf->m_state.fid_states[m_id].ang_f_to_l;
-
-  // Eigen::Vector3d pos_f_in_l_est = average_vectors(pos_f_in_l_vec, pos_weights);
-  // Eigen::Quaterniond ang_f_to_l_est = average_quaternions(ang_f_to_l_vec, ang_weights);
-
-  // /// Project fiducial onto 3-sigma error bound
-  // Eigen::Vector3d f_pos_delta = pos_f_in_l_est - ekf->m_state.fid_states[m_id].pos_f_in_l;
[…]
