### 2024-10-07 1ac9279
Incorporate GPS uncertainty into update
