+++ src/sensors/sim/sim_imu.cpp
@@ -104,11 +104,9 @@ std::vector<std::shared_ptr<SimImuMessage>> SimIMU::GenerateMessages()
       sim_imu_msg->angular_rate += m_rng.VecNormRand(gyr_bias_true, m_omg_error);
     }
 
-    Eigen::Vector3d accSigmas(m_acc_error[0], m_acc_error[1], m_acc_error[2]);
-    Eigen::Vector3d omgSigmas(m_omg_error[0], m_omg_error[1], m_omg_error[2]);
+    sim_imu_msg->acceleration_covariance = m_acc_error.asDiagonal();
+    sim_imu_msg->angular_rate_covariance = m_omg_error.asDiagonal();
 
-    sim_imu_msg->acceleration_covariance = accSigmas.asDiagonal() * 10.0;
-    sim_imu_msg->angular_rate_covariance = omgSigmas.asDiagonal() * 10.0;
     messages.push_back(sim_imu_msg);
   }
   return messages;
