### 2024-10-08 5cdab82
Stabilize example input
