### 2024-10-10 c9f3394
Do single update as predictor after being stationary
