### 2024-10-17 b011966
First pass at separate IMU filter
