### 2024-10-18 1459321
Fix additional body state index bugs
