+++ src/ekf/update/gps_updater.cpp
@@ -65,7 +65,7 @@ Eigen::MatrixXd GpsUpdater::GetMeasurementJacobian(std::shared_ptr<EKF> ekf)
 
   Eigen::MatrixXd measurement_jacobian = Eigen::MatrixXd::Zero(3, ekf->GetStateSize());
   measurement_jacobian.block<3, 3>(0, 0) = Eigen::Matrix3d::Identity(3, 3);
-  measurement_jacobian.block<3, 3>(0, 9) = -ang_b_to_g.toRotationMatrix() *
+  measurement_jacobian.block<3, 3>(0, 6) = -ang_b_to_g.toRotationMatrix() *
     SkewSymmetric(m_pos_a_in_b) * quaternion_jacobian(ang_b_to_g);
 
   if (m_is_extrinsic) {
