### 2024-10-18 e3c4c9b
Update yaml naming and fully remove use _for_prediction
