+++ config/regression/pred_imu_cam.yaml
@@ -2,7 +2,7 @@
     ros__parameters:
         debug_log_level: 2
         data_logging_on: true
-        body_data_rate: 10.0
+        data_log_rate: 10.0
         process_noise:
             pos: 1.0e-3
             vel: 1.0e-4
@@ -37,7 +37,6 @@
 
         imu:
             imu_1:
-                use_for_prediction: true
                 is_extrinsic: false
                 is_intrinsic: false
                 rate: 400.0
