### 2024-10-18 #7292 @kyleniemeyer
@JHartzer please do work on the CI pipeline; generally this isn't too challenging to set up with GitHub Actions.

@pdebuyl could you edit this?
