### 2024-10-19 76fff8b
Remove prediction regression cases
