+++ config/regression/pred_imu_rotation.yaml
@@ -1,64 +1,0 @@
-/EkfCalNode:
-    ros__parameters:
-        debug_log_level: 2
-        data_logging_on: true
-        data_log_rate: 10.0
-        process_noise:
-            pos: 1.0e-3
-            vel: 1.0e-4
-            acc: 1.0e-1
-            ang_pos: 1.0e-3
-            ang_vel: 1.0e-1
-            ang_acc: 1.0e-5
-        sim_params:
-            seed: 0.0
-            use_seed: true
-            max_time: 60.0
-            truth_type: "cyclic"
-            pos_frequency: [0.0, 0.0, 0.0]
-            ang_frequency: [0.2, 0.3, 0.4]
-            pos_offset: [0.0, 0.0, 0.0]
-            no_errors: false
-            number_of_runs: 10
-            run_number: 0
-
-        imu_list:
-            - "imu_1"
-
-        camera_list: []
-
-        tracker_list: []
-
-        fiducial_list: []
-
-        gps_list: []
-
-        imu:
-            imu_1:
-                is_extrinsic: false
-                is_intrinsic: false
-                rate: 400.0
-                topic: "/imu_1"
-                variance: [0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01]
-                pos_i_in_b: [0.0, 0.0, 0.0]
-                ang_i_to_b: [1.0, 0.0, 0.0, 0.0]
-                acc_bias: [0.0, 0.0, 0.0]
-                omg_bias: [0.0, 0.0, 0.0]
-                sim_params:
-                    data_log_rate: 10.0
-                    time_bias_error: 0.0
-                    time_error: 1.0e-6
-                    pos_error: [0.0, 0.0, 0.0]
-                    ang_error: [0.0, 0.0, 0.0]
-                    acc_error: [1.0e-2, 1.0e-2, 1.0e-2]
-                    omg_error: [1.0e-2, 1.0e-2, 1.0e-2]
-                    acc_bias_error: [1.0e-1, 1.0e-1, 1.0e-1]
-                    omg_bias_error: [1.0e-1, 1.0e-1, 1.0e-1]
-
-        camera: {}
-
-        tracker: {}
-
-        fiducial: {}
-
-        gps: {}
