+++ config/regression/pred_imu_cam.yaml
@@ -1,90 +1,0 @@
-/EkfCalNode:
-    ros__parameters:
-        debug_log_level: 2
-        data_logging_on: true
-        data_log_rate: 10.0
-        process_noise:
-            pos: 1.0e-3
-            vel: 1.0e-4
-            acc: 1.0e-1
-            ang_pos: 1.0e-3
-            ang_vel: 1.0e-1
-            ang_acc: 1.0e-5
-        sim_params:
-            seed: 0.0
-            use_seed: true
-            max_time: 60.0
-            truth_type: "cyclic"
-            pos_frequency: [0.3, 0.5, 0.7]
-            ang_frequency: [0.2, 0.3, 0.4]
-            pos_offset: [0.0, 0.0, 0.0]
-            no_errors: false
-            number_of_runs: 10
-            run_number: 0
-
-        imu_list:
-            - "imu_1"
-
-        camera_list:
-            - "cam_1"
-
-        tracker_list:
-            - "orb"
-
-        fiducial_list: []
-
-        gps_list: []
-
-        imu:
-            imu_1:
-                is_extrinsic: false
-                is_intrinsic: false
-                rate: 400.0
-                topic: "/imu_1"
-                variance: [0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01]
-                pos_i_in_b: [0.0, 0.0, 0.0]
-                ang_i_to_b: [1.0, 0.0, 0.0, 0.0]
-                acc_bias: [0.0, 0.0, 0.0]
-                omg_bias: [0.0, 0.0, 0.0]
-                sim_params:
-                    no_errors: true
-                    data_log_rate: 10.0
-                    time_bias_error: 0.0
-                    time_error: 1.0e-6
-                    pos_error: [0.0, 0.0, 0.0]
-                    ang_error: [0.0, 0.0, 0.0]
-                    acc_error: [1.0e-2, 1.0e-2, 1.0e-2]
-                    omg_error: [1.0e-2, 1.0e-2, 1.0e-2]
-                    acc_bias_error: [1.0e-1, 1.0e-1, 1.0e-1]
-                    omg_bias_error: [1.0e-1, 1.0e-1, 1.0e-1]
-
-        camera:
-            cam_1:
-                rate: 20.0
-                data_log_rate: 10.0
-                topic:  "/cam_1"
-                pos_c_in_b: [0.0, 0.0, 0.0]
-                ang_c_to_b: [-0.5, 0.5, -0.5, 0.5]
-                variance: [0.1, 0.1, 0.1, 0.1, 0.1, 0.1]
-                tracker: "orb"
-                sim_params:
-                    time_bias_error: 0.0
-                    time_error: 1.0e-6
-                    pos_error: [0.0, 0.0, 0.0]
-                    ang_error: [0.0, 0.0, 0.0]
-
-        tracker:
-            orb:
-                feature_detector: 4
-                descriptor_extractor: 0
-                descriptor_matcher: 0
-                detector_threshold: 10.0
-                pixel_error: 0.5
-                sim_params:
-                    no_errors: true
-                    feature_count: 1e6
-                    room_size: 10.0
-
-        fiducial: {}
-
-        gps: {}

[…]
