### 2024-10-19 fee851f
Automatically run unit tests
