+++ .github/workflows/documentation.yaml
@@ -1,46 +1,0 @@
-name: Documentation
-
-on:
-  push:
-    branches:
-      - main
-
-jobs:
-  build-documentation:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v4
-        with:
-          submodules: 'true'
-      - name: Count Lines of Code
-        uses: djdefi/cloc-action@6
-        with:
-          options: --exclude-lang=CSS,CSV,HTML,JavaScript,make,SVG,TeX,YAML --md --report-file=cloc.md
-      - name: Format and move cloc.md
-        run: |
-          echo -e 'Count Lines of Code {#cloc}\n============\n' > docs/software/cloc.md
-          cat cloc.md | tail -n +4 >> docs/software/cloc.md
-          sed -i 's/--------|--------|--------|--------|--------/| | | | | |/' docs/software/cloc.md
-          rm cloc.md
-      - name: Run Doxygen
-        uses: mattnotmitt/doxygen-action@1.9.5
-        with:
-          doxyfile-path: .doxyfile
-      - name: Upload Artifact
-        uses: actions/upload-pages-artifact@v1
-        with:
-          path: docs/doxygen/html
-  deploy-documentation:
-    needs: build-documentation
-    permissions:
-      pages: write
-      id-token: write
-    environment:
-      name: github-pages
-      url: ${{ steps.deployment.outputs.page_url }}
-    runs-on: ubuntu-latest
-    steps:
-      - name: Deploy to GitHub Pages
-        id: deployment
-        uses: actions/deploy-pages@v2
