### 2024-10-19 #7292 @JHartzer
The unit tests are now being performed automatically with a github workflow!
