### 2024-10-20 47d68f1
Add option for camera extrinsic calibration
