### 2024-10-22 d0d541f
Remove quaternion inverse jacobian
