+++ src/utility/test/math_helper_test.cpp
@@ -281,12 +281,6 @@ TEST(test_MathHelper, quaternion_jacobian) {
   EXPECT_TRUE(EXPECT_EIGEN_NEAR(jac, Eigen::Matrix3d::Identity(3, 3), 1e-6));
 }
 
-TEST(test_MathHelper, quaternion_jacobian_inv) {
-  Eigen::Quaterniond quat{1, 0, 0, 0};
-  Eigen::Matrix3d jac = quaternion_jacobian_inv(quat);
-  EXPECT_TRUE(EXPECT_EIGEN_NEAR(jac, Eigen::Matrix3d::Identity(3, 3), 1e-6));
-}
-
 TEST(test_MathHelper, kabsch_2d) {
   std::vector<Eigen::Vector3d> points_src;
   points_src.push_back(Eigen::Vector3d{0, 0, 0});
