+++ src/utility/math_helper.cpp
@@ -278,21 +278,6 @@ Eigen::MatrixXd quaternion_jacobian(Eigen::Quaterniond quat)
   return jacobian;
 }
 
-Eigen::MatrixXd quaternion_jacobian_inv(Eigen::Quaterniond quat)
-{
-  Eigen::Vector3d rot_vec = QuatToRotVec(quat);
-  Eigen::Matrix3d skew_mat = SkewSymmetric(rot_vec);
-  double vec_norm = std::max(rot_vec.norm(), 1e-9);
-  double coeff_one = 0.5;
-  double coeff_two =
-    std::pow(vec_norm, -2) - (1 + std::cos(vec_norm)) / (2 * vec_norm * std::sin(vec_norm));
-  Eigen::Matrix3d jacobian =
-    Eigen::Matrix3d::Identity(3, 3) +
-    coeff_one * skew_mat +
-    coeff_two * skew_mat * skew_mat;
-  return jacobian;
-}
-
 double sign(double val)
 {
   return (0.0 < val) - (val < 0.0);
