### 2024-10-22 4959061
AugmentCovariance bugfix
