### 2024-10-22 c14973b
Fix error in stationary jacobian
