### 2024-10-22 85b21d3
Code cleanup
