### 2024-10-24 0ef6fef
Use public workflow badge in README and documentation
