### 2024-10-24 #7292 @JHartzer
Thanks for the catch! That should be resolved now. It was a classic case of 'it works on my machine' using a private test repo badge.
