### 2024-10-25 d1824f8
Only perform stationary updates before motion
