+++ src/ekf/update/imu_updater.hpp
@@ -105,7 +105,7 @@ private:
   bool m_is_extrinsic;
   bool m_is_intrinsic;
   bool m_is_first_estimate{true};
-  bool m_was_stationary{false};
+  bool m_initial_motion_detected{false};
 
   DataLogger m_data_logger;
 };
