### 2024-10-25 b4fb719
Fix bug in square root form QR decomposition
