### 2024-10-26 7b9d201
Restore nullspace projection
