### 2024-10-26 fab50a4
Add camera extrinsic flags to MSCKF and fiducial updaters
