+++ src/trackers/feature_tracker.cpp
@@ -36,6 +36,7 @@ FeatureTracker::FeatureTracker(FeatureTracker::Parameters params)
 : Tracker(params),
   m_msckf_updater(
     params.camera_id,
+    params.is_cam_extrinsic,
     params.output_directory,
     params.data_logging_on,
     params.data_log_rate,
