### 2024-10-26 1beeac9
Fiducial measurement covariance bugfix
