+++ src/application/sim/simulation.cpp
@@ -191,7 +191,7 @@ int main(int argc, char * argv[])
   ekf_params.motion_detection_chi_squared =
     ros_params["motion_detection_chi_squared"].as<double>(0.1);
   ekf_params.imu_noise_scale_factor = ros_params["imu_noise_scale_factor"].as<double>(100.0);
-  ekf_params.use_root_covariance = ros_params["use_root_covariance"].as<bool>(false);
+  ekf_params.use_root_covariance = ros_params["use_root_covariance"].as<bool>(true);
   ekf_params.use_first_estimate_jacobian =
     ros_params["use_first_estimate_jacobian"].as<bool>(false);
   auto ekf = std::make_shared<EKF>(ekf_params);
