+++ config/regression/fiducial.yaml
@@ -4,12 +4,9 @@
         data_logging_on: true
         data_log_rate: 10.0
         process_noise:
-            pos: 1.0e-3
-            vel: 1.0e-4
-            acc: 1.0e-1
-            ang_pos: 1.0e-3
-            ang_vel: 1.0e-1
-            ang_acc: 1.0e-5
+            pos: 1.0e-1
+            vel: 1.0e-1
+            ang_pos: 1.0e-1
         sim_params:
             seed: 0.0
             use_seed: true
@@ -24,7 +21,7 @@
             pos_offset: [0.0, 0.0, 0.0]
             ang_offset: [0.0, 0.0, 0.0]
             ang_amplitude: 0.1
-            pos_amplitude: 1.0
+            pos_amplitude: 0.5
 
         imu_list:
             - "imu_1"
@@ -62,8 +59,8 @@
                     ang_error: [0.0, 0.0, 0.0]
                     acc_error: [1.0e-2, 1.0e-2, 1.0e-2]
                     omg_error: [1.0e-2, 1.0e-2, 1.0e-2]
-                    acc_bias_error: [1.0e-1, 1.0e-1, 1.0e-1]
-                    omg_bias_error: [1.0e-1, 1.0e-1, 1.0e-1]
+                    acc_bias_error: [0.0, 0.0, 0.0]
+                    omg_bias_error: [0.0, 0.0, 0.0]
 
         camera:
             cam_1:
@@ -103,9 +100,10 @@
                 min_track_length: 0
                 max_track_length: 1
                 data_log_rate: 10.0
+                is_extrinsic: false
                 sim_params:
-                    pos_error: [1.0e-1, 1.0e-1, 1.0e-1]
-                    ang_error: [1.0e-1, 1.0e-1, 1.0e-1]
+                    pos_error: [0.0, 0.0, 0.0]
+                    ang_error: [0.0, 0.0, 0.0]
                     t_vec_error: [1.0e-2, 1.0e-2, 1.0e-2]
                     r_vec_error: [1.0e-2, 1.0e-2, 1.0e-2]
 
