### 2024-10-26 5898a66
Utilize zero acceleration in prediction step
