### 2024-10-27 cb602f7
GPS plotting updates
