### 2024-10-27 446227d
IMU angular-only update
