### 2024-10-28 7b27a16
Fix covariance augmenting when using root form
