+++ src/ekf/update/imu_updater.cpp
@@ -136,6 +136,35 @@ void ImuUpdater::UpdateEKF(
   ekf->LogBodyStateIfNeeded(t_execution.count());
 }
 
+Eigen::MatrixXd ImuUpdater::GetMeasurementJacobian(
+  std::shared_ptr<EKF> ekf,
+  unsigned int imu_id)
+{
+  unsigned int meas_size = m_is_intrinsic ? 6 : 3;
+  Eigen::MatrixXd H = Eigen::MatrixXd::Zero(meas_size, ekf->GetStateSize());
+
+
+  Eigen::Quaterniond ang_i_to_b = ekf->m_state.imu_states[imu_id].ang_i_to_b;
+  Eigen::Quaterniond ang_b_to_l = ekf->m_state.body_state.ang_b_to_l;
+
+  H.block<3, 3>(0, 6) = -SkewSymmetric(ang_i_to_b.inverse() * ang_b_to_l.inverse() * g_gravity) *
+    quaternion_jacobian(ang_b_to_l).transpose();
+
+  if (m_is_extrinsic) {
+    unsigned int index_extrinsic = ekf->m_state.imu_states[imu_id].index_extrinsic;
+    H.block<3, 3>(0, index_extrinsic) = -SkewSymmetric(
+      ang_i_to_b.inverse() * ang_b_to_l.inverse() * g_gravity
+    );
+  }
+
+  if (m_is_intrinsic) {
+    unsigned int index_intrinsic = ekf->m_state.imu_states[imu_id].index_intrinsic;
+    H.block<3, 3>(0, index_intrinsic + 0) = -Eigen::Matrix3d::Identity();
+    H.block<3, 3>(3, index_intrinsic + 3) = -Eigen::Matrix3d::Identity();
+  }
+
+  return H;
+}
 
 bool ImuUpdater::ZeroAccelerationUpdate(
   std::shared_ptr<EKF> ekf,
@@ -152,18 +181,17 @@ bool ImuUpdater::ZeroAccelerationUpdate(
     return false;
   }
 
-  unsigned int index_extrinsic = ekf->m_state.imu_states[imu_id].index_extrinsic;
-  unsigned int index_intrinsic = ekf->m_state.imu_states[imu_id].index_intrinsic;
   unsigned int meas_size = m_is_intrinsic ? 6 : 3;
   Eigen::Quaterniond ang_i_to_b = ekf->m_state.imu_states[imu_id].ang_i_to_b;
   Eigen::Quaterniond ang_b_to_l = ekf->m_state.body_state.ang_b_to_l;
 
-  Eigen::MatrixXd H = Eigen::MatrixXd::Zero(meas_size, ekf->GetStateSize());
-  H.block<3, 3>(0, 6) = -SkewSymmetric(ang_i_to_b.inverse() * ang_b_to_l.inverse() * g_gravity) *
-    quaternion_jacobian(ang_b_to_l).transpose();
+  Eigen::MatrixXd H = GetMeasurementJacobian(ekf, imu_id);
 
   Eigen::MatrixXd R = Eigen::MatrixXd::Zero(meas_size, meas_size);
   R.block<3, 3>(0, 0) = acceleration_covariance;
+  if (m_is_intrinsic) {
+    R.block<3, 3>(3, 3) = angular_rate_covariance;
+  }
 
   Eigen::Vector3d bias_a = ekf->m_state.imu_states[imu_id].acc_bias;
   Eigen::Vector3d bias_g = ekf->m_state.imu_states[imu_id].omg_bias;
@@ -172,18 +200,8 @@ bool ImuUpdater::ZeroAccelerationUpdate(
   resid.segment<3>(0) = -(acceleration - bias_a -
     ang_i_to_b.inverse() *
[…]
