### 2024-10-29 f6e6043
Use augmented state index for Fiducial MSCKF
