+++ src/ekf/update/fiducial_updater.cpp
@@ -95,27 +95,23 @@ void FiducialUpdater::UpdateEKF(
   for (auto board_detection : board_track) {
     AugState aug_state_i = ekf->GetAugState(m_camera_id, board_detection.frame_id, time);
 
-    const Eigen::Vector3d pos_bi_in_g = aug_state_i.pos_b_in_l;
+    const Eigen::Vector3d pos_bi_in_l = aug_state_i.pos_b_in_l;
     const Eigen::Matrix3d rot_bi_to_l = aug_state_i.ang_b_to_l.toRotationMatrix();
     const Eigen::Matrix3d rot_c_to_b = m_ang_c_to_b.toRotationMatrix();
 
     Eigen::Vector3d pos_f_in_c;
     CvVectorToEigen(board_detection.t_vec_f_in_c, pos_f_in_c);
     Eigen::Vector3d pos_f_in_l =
-      rot_bi_to_l * ((rot_c_to_b * pos_f_in_c) + m_pos_c_in_b) + pos_bi_in_g;
+      rot_bi_to_l * ((rot_c_to_b * pos_f_in_c) + m_pos_c_in_b) + pos_bi_in_l;
 
     Eigen::Quaterniond ang_f_to_c = RodriguesToQuat(board_detection.r_vec_f_to_c);
     Eigen::Quaterniond ang_f_to_l = aug_state_i.ang_b_to_l * m_ang_c_to_b * ang_f_to_c;
 
     std::stringstream data_msg;
     data_msg << std::setprecision(3) << time;
-    data_msg << ",0," << pos_f_in_l[0];
-    data_msg << "," << pos_f_in_l[1];
-    data_msg << "," << pos_f_in_l[2];
-    data_msg << "," << ang_f_to_l.w();
-    data_msg << "," << ang_f_to_l.x();
-    data_msg << "," << ang_f_to_l.y();
-    data_msg << "," << ang_f_to_l.z();
+    data_msg << ",0";
+    data_msg << VectorToCommaString(pos_f_in_l);
+    data_msg << QuaternionToCommaString(ang_f_to_l);
     m_board_logger.RateLimitedLog(data_msg.str(), time);
   }
 
@@ -143,17 +139,18 @@ void FiducialUpdater::UpdateEKF(
     Eigen::Matrix3d rot_bi_to_l = aug_state_i.ang_b_to_l.toRotationMatrix();
     Eigen::Matrix3d rot_bi_to_c = rot_c_to_b.transpose();
     Eigen::Matrix3d rot_l_to_bi = rot_bi_to_l.transpose();
-    Eigen::Matrix3d rot_l_to_ci = rot_bi_to_c * rot_bi_to_l.transpose();
+    Eigen::Matrix3d rot_l_to_ci = rot_bi_to_c * rot_l_to_bi;
+    Eigen::Matrix3d rot_l_to_c = rot_bi_to_c * rot_l_to_bi;
     Eigen::Quaterniond ang_l_to_ci(rot_l_to_ci);
 
-    Eigen::Vector3d pos_bi_in_g = aug_state_i.pos_b_in_l;
+    Eigen::Vector3d pos_bi_in_l = aug_state_i.pos_b_in_l;
 
     Eigen::Vector3d pos_predicted, pos_measured, pos_residual;
     Eigen::Quaterniond ang_predicted, ang_measured, ang_residual;
 
     // Project the current feature into the current frame of reference
-    Eigen::Vector3d pos_f_in_bi = rot_bi_to_l.transpose() * (m_pos_f_in_l - pos_bi_in_g);
[…]
