### 2024-10-31 c189e76
Minor coordinate naming updates
