+++ src/ekf/update/msckf_updater.cpp
@@ -68,13 +68,13 @@ MsckfUpdater::MsckfUpdater(
 }
 
 bool MsckfUpdater::TriangulateFeature(
+  double time,
   std::shared_ptr<EKF> ekf,
   FeatureTrack & feature_track,
   Eigen::Vector3d & pos_f_in_l)
 {
   /// @todo: Need to continue debugging the triangulated features
   pos_f_in_l = feature_track.true_feature_position;
-  return true;
 
   AugState aug_state_0 = ekf->GetAugState(
     m_id, feature_track.track[0].frame_id, feature_track.track[0].frame_time);
@@ -124,15 +124,22 @@ bool MsckfUpdater::TriangulateFeature(
   // Solve linear triangulation for 3D cartesian estimate of feature pos
   Eigen::Vector3d pos_f_in_c0 = A.colPivHouseholderQr().solve(b);
 
+  Eigen::Vector3d pos_f_in_l_tri;
   if (pos_f_in_c0.z() < m_min_feat_dist || m_max_feat_dist < pos_f_in_c0.z()) {
     std::stringstream err_msg;
     err_msg << "MSCKF triangulated point out of bounds. r = " << pos_f_in_l.norm();
     m_logger->Log(LogLevel::INFO, err_msg.str());
-    pos_f_in_l = Eigen::Vector3d::Zero();
-    return false;
+    pos_f_in_l_tri = Eigen::Vector3d::Zero();
+    // return false;
   }
 
-  pos_f_in_l = rot_b0_to_l * (m_ang_c_to_b * pos_f_in_c0 + m_pos_c_in_b) + pos_b0_in_l;
+  pos_f_in_l_tri = rot_b0_to_l * (m_ang_c_to_b * pos_f_in_c0 + m_pos_c_in_b) + pos_b0_in_l;
+
+  std::stringstream msg;
+  msg << std::setprecision(3) << time;
+  msg << "," << std::to_string(feature_track.track[0].key_point.class_id);
+  msg << VectorToCommaString(pos_f_in_l_tri);
+  m_triangulation_logger.RateLimitedLog(msg.str(), time);
 
   return true;
 }
@@ -241,16 +248,10 @@ void MsckfUpdater::UpdateEKF(
 
     // Get triangulated estimate of feature pos
     Eigen::Vector3d pos_f_in_l;
-    bool triangulation_successful = TriangulateFeature(ekf, feature_track, pos_f_in_l);
+    bool triangulation_successful = TriangulateFeature(time, ekf, feature_track, pos_f_in_l);
     if (!triangulation_successful) {
       continue;
     }
-
-    std::stringstream msg;
-    msg << std::setprecision(3) << time;
-    msg << "," << std::to_string(feature_track.track[0].key_point.class_id);
-    msg << "," << VectorToCommaString(pos_f_in_l);
-    m_triangulation_logger.RateLimitedLog(msg.str(), time);
 
     unsigned int aug_state_size = ekf->GetAugStateSize();
     unsigned int aug_state_start = ekf->GetAugStateStart();
@@ -271,16 +272,17 @@ void MsckfUpdater::UpdateEKF(
       // Project the current feature into the current frame of reference
       Eigen::Vector3d pos_f_in_bi = rot_bi_to_l.transpose() * (pos_f_in_l - pos_bi_in_l);
       Eigen::Vector3d pos_f_in_ci = rot_ci_to_b.transpose() * (pos_f_in_bi - m_pos_c_in_b);
-      Eigen::Vector2d xz_predicted;
-      xz_predicted(0) = pos_f_in_ci(0) / pos_f_in_ci(2);
[…]
