### 2024-11-04 300471e
Fix off-diagonal covariance propagation for non-root form
