+++ src/ekf/types.hpp
@@ -269,6 +269,7 @@ typedef struct FeaturePoint
 typedef struct FeatureTrack
 {
   std::vector<FeaturePoint> track;        ///< @brief Vector of tracked feature keypoints
+  /// @todo: Remove true feature position once MSCKF complete
   Eigen::Vector3d true_feature_position;  ///< @brief True feature position (sim only)
 } FeatureTrack;
 
