### 2024-11-05 d19b6ea
Fix off-diagonal covariance propagation for root form
