### 2024-11-07 e89c5f4
Fix left nullspace projection
