+++ src/utility/test/math_helper_test.cpp
@@ -121,25 +121,30 @@ TEST(test_MathHelper, RemoveFromMatrix)
 }
 
 TEST(test_MathHelper, ApplyLeftNullspace) {
-  Eigen::MatrixXd H_f(2, 2);
-  H_f << 1, 0, 0, 1;
-  Eigen::MatrixXd H_x(5, 5);
+  Eigen::MatrixXd H_f(6, 3);
+  H_f <<
+    1, 0, -1,
+    0, 1, -1,
+    1, 0, -2,
+    0, 1, -3,
+    1, 0, -4,
+    0, 1, -5;
+  Eigen::MatrixXd H_x(6, 5);
   H_x <<
     1, 0, 0, 0, 0,
     0, 1, 0, 0, 0,
     0, 0, 1, 0, 0,
     0, 0, 0, 1, 0,
-    0, 0, 0, 0, 1;
-  Eigen::VectorXd res(5);
-  res << 1, 2, 3, 4, 5;
+    0, 0, 0, 0, 1,
+    1, 1, 1, 1, 1;
+  Eigen::VectorXd res(6);
+  res << 1, 2, 3, 4, 5, 6;
+
   ApplyLeftNullspace(H_f, H_x, res);
 
+  EXPECT_EQ(res.size(), 3U);
   EXPECT_EQ(H_x.rows(), 3U);
   EXPECT_EQ(H_x.cols(), 5U);
-  EXPECT_EQ(res.size(), 3U);
-  EXPECT_EQ(res(0), 3);
-  EXPECT_EQ(res(1), 4);
-  EXPECT_EQ(res(2), 5);
 }
 
 TEST(test_MathHelper, CompressMeasurements) {
