### 2024-11-07 73e726d
Remove unused helper functions
