+++ src/utility/test/math_helper_test.cpp
@@ -35,21 +35,6 @@ TEST(test_MathHelper, SkewSymmetric) {
   EXPECT_EQ(outMat(2, 2), 0.0);
 }
 
-TEST(test_MathHelper, MinBoundDiagonal)
-{
-  Eigen::MatrixXd mat2 = Eigen::MatrixXd::Ones(2, 2);
-  MinBoundDiagonal(mat2, 1);
-  EXPECT_EQ(mat2, Eigen::MatrixXd::Ones(2, 2));
-
-  Eigen::MatrixXd mat3 = Eigen::MatrixXd::Zero(3, 3);
-  MinBoundDiagonal(mat3, 1);
-  EXPECT_EQ(mat3, Eigen::MatrixXd::Identity(3, 3));
-
-  Eigen::MatrixXd mat4 = Eigen::MatrixXd::Zero(4, 4);
-  MinBoundDiagonal(mat4, 1);
-  EXPECT_EQ(mat4, Eigen::MatrixXd::Identity(4, 4));
-}
-
 TEST(test_MathHelper, MinBoundVector)
 {
   Eigen::VectorXd vec2 = Eigen::VectorXd::Ones(2);
@@ -63,15 +48,6 @@ TEST(test_MathHelper, MinBoundVector)
   Eigen::VectorXd vec4 = Eigen::VectorXd::Zero(4);
   MinBoundVector(vec4, 1);
   EXPECT_EQ(vec4, Eigen::VectorXd::Ones(4));
-}
-
-TEST(test_MathHelper, MaxBoundDiagonal)
-{
-  Eigen::MatrixXd eye = Eigen::Matrix3d::Identity();
-  Eigen::MatrixXd bounded_mat = Eigen::Matrix3d::Identity() * 2;
-  MaxBoundDiagonal(bounded_mat, 1.0);
-
-  EXPECT_TRUE(EXPECT_EIGEN_NEAR(bounded_mat, eye, 1e-6));
 }
 
 TEST(test_MathHelper, RemoveFromMatrix)
@@ -178,76 +154,6 @@ TEST(test_MathHelper, CompressMeasurements) {
   EXPECT_NEAR(residual2(0), 2.0, 1e-6);
 }
 
-TEST(test_MathHelper, average_quaternions) {
-  std::vector<Eigen::Quaterniond> quaternions_1;
-  quaternions_1.push_back(Eigen::Quaterniond(1.0, 0.0, 0.0, 0.0));
-  quaternions_1.push_back(Eigen::Quaterniond(1.0, 0.0, 0.0, 0.0));
-  quaternions_1.push_back(Eigen::Quaterniond(1.0, 0.0, 0.0, 0.0));
-  quaternions_1.push_back(Eigen::Quaterniond(1.0, 0.0, 0.0, 0.0));
-  quaternions_1.push_back(Eigen::Quaterniond(1.0, 0.0, 0.0, 0.0));
-
-  std::vector<double> weights_1;
-  weights_1.push_back(1.0);
-  weights_1.push_back(1.0);
-  weights_1.push_back(1.0);
-  weights_1.push_back(1.0);
-  weights_1.push_back(1.0);
-
-  Eigen::Quaterniond average_quaternion_1 = average_quaternions(quaternions_1, weights_1);
-  EXPECT_EQ(average_quaternion_1.w(), 1.0);
-  EXPECT_EQ(average_quaternion_1.x(), 0.0);
-  EXPECT_EQ(average_quaternion_1.y(), 0.0);
-  EXPECT_EQ(average_quaternion_1.z(), 0.0);
-
-  std::vector<Eigen::Quaterniond> quaternions_2;
-  quaternions_2.push_back(Eigen::Quaterniond(0.9990482, 0.0436194, 0, 0));
[…]

[…]
