### 2024-11-09 79cd237
Don't pass sensor ID through to trackers via function call
