+++ src/trackers/sim/sim_fiducial_tracker.cpp
@@ -36,40 +36,40 @@ SimFiducialTracker::SimFiducialTracker(
   m_min_track_length = params.fiducial_params.min_track_length;
   m_max_track_length = params.fiducial_params.max_track_length;
 
-  Eigen::Vector3d pos_f_in_g_true;
-  Eigen::Quaterniond ang_f_to_g_true;
+  Eigen::Vector3d pos_f_in_l_true;
+  Eigen::Quaterniond ang_f_to_l_true;
   if (m_no_errors) {
-    pos_f_in_g_true = params.fiducial_params.pos_f_in_l;
-    ang_f_to_g_true = params.fiducial_params.ang_f_to_l;
+    pos_f_in_l_true = params.fiducial_params.pos_f_in_l;
+    ang_f_to_l_true = params.fiducial_params.ang_f_to_l;
   } else {
-    pos_f_in_g_true = m_rng.VecNormRand(params.fiducial_params.pos_f_in_l, params.pos_error);
-    ang_f_to_g_true = m_rng.QuatNormRand(params.fiducial_params.ang_f_to_l, params.ang_error);
+    pos_f_in_l_true = m_rng.VecNormRand(params.fiducial_params.pos_f_in_l, params.pos_error);
+    ang_f_to_l_true = m_rng.QuatNormRand(params.fiducial_params.ang_f_to_l, params.ang_error);
   }
-  truth_engine->SetBoardPosition(m_id, pos_f_in_g_true);
-  truth_engine->SetBoardOrientation(m_id, ang_f_to_g_true);
+  truth_engine->SetBoardPosition(m_id, pos_f_in_l_true);
+  truth_engine->SetBoardOrientation(m_id, ang_f_to_l_true);
 }
 
 bool SimFiducialTracker::IsBoardVisible(double time)
 {
-  Eigen::Vector3d pos_b_in_g = m_truth->GetBodyPosition(time);
-  Eigen::Quaterniond ang_b_to_g = m_truth->GetBodyAngularPosition(time);
+  Eigen::Vector3d pos_b_in_l = m_truth->GetBodyPosition(time);
+  Eigen::Quaterniond ang_b_to_l = m_truth->GetBodyAngularPosition(time);
   Eigen::Vector3d pos_c_in_b = m_truth->GetCameraPosition(m_camera_id);
   Eigen::Quaterniond ang_c_to_b = m_truth->GetCameraAngularPosition(m_camera_id);
   Intrinsics intrinsics = m_truth->GetCameraIntrinsics(m_camera_id);
-  Eigen::Matrix3d rot_g_to_c = (ang_b_to_g * ang_c_to_b).toRotationMatrix().transpose();
-  cv::Mat ang_g_to_c_cv(3, 3, cv::DataType<double>::type);
-  EigenMatrixToCv(rot_g_to_c, ang_g_to_c_cv);
+  Eigen::Matrix3d rot_l_to_c = (ang_b_to_l * ang_c_to_b).toRotationMatrix().transpose();
+  cv::Mat ang_l_to_c_cv(3, 3, cv::DataType<double>::type);
+  EigenMatrixToCv(rot_l_to_c, ang_l_to_c_cv);
 
   // Creating Rodrigues rotation vector
   cv::Mat r_vec(3, 1, cv::DataType<double>::type);
-  cv::Rodrigues(ang_g_to_c_cv, r_vec);
+  cv::Rodrigues(ang_l_to_c_cv, r_vec);
 
-  Eigen::Vector3d pos_g_in_c = rot_g_to_c * (-(pos_b_in_g + ang_b_to_g * pos_c_in_b));
+  Eigen::Vector3d pos_l_in_c = rot_l_to_c * (-(pos_b_in_l + ang_b_to_l * pos_c_in_b));
 
[…]
