+++ src/ekf/update/gps_updater.cpp
@@ -60,16 +60,16 @@ GpsUpdater::GpsUpdater(
 
 Eigen::MatrixXd GpsUpdater::GetMeasurementJacobian(std::shared_ptr<EKF> ekf)
 {
-  Eigen::Quaterniond ang_b_to_g = ekf->m_state.body_state.ang_b_to_l;
+  Eigen::Quaterniond ang_b_to_l = ekf->m_state.body_state.ang_b_to_l;
 
   Eigen::MatrixXd measurement_jacobian = Eigen::MatrixXd::Zero(3, ekf->GetStateSize());
   measurement_jacobian.block<3, 3>(0, 0) = Eigen::Matrix3d::Identity(3, 3);
-  measurement_jacobian.block<3, 3>(0, 6) = -ang_b_to_g.toRotationMatrix() *
-    SkewSymmetric(m_pos_a_in_b) * quaternion_jacobian(ang_b_to_g);
+  measurement_jacobian.block<3, 3>(0, 6) = -ang_b_to_l.toRotationMatrix() *
+    SkewSymmetric(m_pos_a_in_b) * quaternion_jacobian(ang_b_to_l);
 
   if (m_is_extrinsic) {
     unsigned int gps_index = ekf->m_state.gps_states[m_id].index;
-    measurement_jacobian.block<3, 3>(0, gps_index) = ang_b_to_g.toRotationMatrix();
+    measurement_jacobian.block<3, 3>(0, gps_index) = ang_b_to_l.toRotationMatrix();
   }
   return measurement_jacobian;
 }
@@ -102,10 +102,10 @@ void GpsUpdater::UpdateEKF(
     Eigen::Vector3d gps_enu = lla_to_enu(gps_lla, reference_lla);
     pos_a_in_l = enu_to_local(gps_enu, ang_l_to_g);
 
-    Eigen::Vector3d pos_b_in_g = ekf->m_state.body_state.pos_b_in_l;
+    Eigen::Vector3d pos_b_in_l = ekf->m_state.body_state.pos_b_in_l;
     Eigen::Quaterniond ang_b_to_l = ekf->m_state.body_state.ang_b_to_l;
 
-    Eigen::Vector3d pos_a_in_l_hat = pos_b_in_g + ang_b_to_l * m_pos_a_in_b;
+    Eigen::Vector3d pos_a_in_l_hat = pos_b_in_l + ang_b_to_l * m_pos_a_in_b;
     residual = pos_a_in_l - pos_a_in_l_hat;
     Eigen::MatrixXd jacobian = GetMeasurementJacobian(ekf);
     KalmanUpdate(ekf, jacobian, residual, pos_covariance);
