### 2024-11-10 e55a4e9
Add MSCKF input for tracking
