### 2024-11-10 72bce83
Default to HD camera
