### 2024-11-10 31e5211
Generate new features when insufficient are visible
