+++ src/trackers/sim/sim_feature_tracker.cpp
@@ -45,13 +45,12 @@ SimFeatureTracker::SimFeatureTracker(
   m_truth = truthEngine;
 }
 
-std::vector<cv::KeyPoint> SimFeatureTracker::VisibleKeypoints(double time)
+std::vector<cv::KeyPoint> SimFeatureTracker::GetVisibleKeypoints(double time)
 {
   Eigen::Vector3d pos_b_in_l = m_truth->GetBodyPosition(time);
   Eigen::Quaterniond ang_b_to_l = m_truth->GetBodyAngularPosition(time);
   Eigen::Vector3d pos_c_in_b = m_truth->GetCameraPosition(m_camera_id);
   Eigen::Quaterniond ang_c_to_b = m_truth->GetCameraAngularPosition(m_camera_id);
-  Intrinsics intrinsics = m_truth->GetCameraIntrinsics(m_camera_id);
   Eigen::Matrix3d rot_c_to_l = (ang_b_to_l * ang_c_to_b).toRotationMatrix();
   Eigen::Matrix3d rot_l_to_c = rot_c_to_l.transpose();
 
@@ -72,6 +71,7 @@ std::vector<cv::KeyPoint> SimFeatureTracker::VisibleKeypoints(double time)
   t_vec.at<double>(2) = pos_l_in_c[2];
 
   // Create intrinsic matrices
+  Intrinsics intrinsics = m_truth->GetCameraIntrinsics(m_camera_id);
   cv::Mat camera_matrix = intrinsics.ToCameraMatrix();
   cv::Mat distortion = intrinsics.ToDistortionVector();
 
@@ -80,10 +80,31 @@ std::vector<cv::KeyPoint> SimFeatureTracker::VisibleKeypoints(double time)
 
   std::vector<cv::Point3d> feature_points = m_truth->GetFeatures();
   cv::projectPoints(feature_points, r_vec, t_vec, camera_matrix, distortion, projected_points);
+  std::vector<cv::KeyPoint> projected_features =
+    FilterInvisiblePoints(feature_points, projected_points, rot_c_to_l, pos_c_in_l, intrinsics);
 
+  if (m_feature_count > projected_features.size()) {
+    unsigned int new_feature_count = m_feature_count - projected_features.size();
+    feature_points = m_truth->GenerateVisibleFeatures(time, m_camera_id, new_feature_count, m_rng);
+    cv::projectPoints(feature_points, r_vec, t_vec, camera_matrix, distortion, projected_points);
+    projected_features =
+      FilterInvisiblePoints(feature_points, projected_points, rot_c_to_l, pos_c_in_l, intrinsics);
+  }
+
+  return projected_features;
+}
+
+std::vector<cv::KeyPoint> SimFeatureTracker::FilterInvisiblePoints(
+  std::vector<cv::Point3d> feature_points,
+  std::vector<cv::Point2d> projected_points,
+  Eigen::Matrix3d rot_c_to_l,
+  Eigen::Vector3d pos_c_in_l,
+  Intrinsics intrinsics
+)
+{
   // Convert to feature points
-  std::vector<cv::KeyPoint> projected_features;
   Eigen::Vector3d cam_plane_vec = rot_c_to_l * Eigen::Vector3d(0, 0, 1);
+  std::vector<cv::KeyPoint> projected_features;
   for (unsigned int i = 0; i < projected_points.size(); ++i) {
     cv::Point3d point_cv = feature_points[i];
     Eigen::Vector3d point_eig(
@@ -105,13 +126,12 @@ std::vector<cv::KeyPoint> SimFeatureTracker::VisibleKeypoints(double time)
         feat.pt.x = projected_points[i].x;
         feat.pt.y = projected_points[i].y;
       } else {
[…]
