+++ src/infrastructure/sim/truth_engine.hpp
@@ -286,7 +286,7 @@ private:
   std::vector<cv::Point3d> m_feature_points;
   std::map<unsigned int, cv::Point3d> m_feature_points_map;
   std::map<unsigned int, Eigen::Vector3d> m_gps_pos;
-  Eigen::Vector3d m_lla_reference;
+  Eigen::Vector3d m_lla_reference{0, 0, 0};
   double m_heading;
   double m_room_size{4};
   int m_grid_size{15};
