### 2024-11-13 18b4913
Resolve test with uninitialized heading
