### 2024-11-13 31a7a45
Use default camera intrinsics in regression set
