### 2024-11-14 cc9a09e
Update defaults and example
