+++ docs/parameters.md
@@ -8,49 +8,73 @@ The basic structure of the YAML files is a set of ROS parameters, lists that def
 ```{.py}
 /EkfCalNode:
     ros__parameters:
-        debug_log_level: 2                  # Debug_Log_Level:
-                                            #   1: FATAL
-                                            #   2: ERROR
-                                            #   3: WARN
-                                            #   4: INFO
-                                            #   5: DEBUG
-        data_logging_on: true               # Flag to enable data logging
-        data_log_rate: 10.0                 # EKF body data logging rate
+        debug_log_level: 2                    # Debug_Log_Level:
+                                              #   1: FATAL
+                                              #   2: ERROR
+                                              #   3: WARN
+                                              #   4: INFO
+                                              #   5: DEBUG
+        data_logging_on: true                 # Flag to enable data logging
+        data_log_rate: 10.0                   # EKF body data logging rate
+        augmenting_type: 0                    # Camera state augmentation type
+                                              #   0: All camera frames are augmented
+                                              #   1: Only the primary camera frames are augmented
+                                              #   2: Time-based frame augmentation
+                                              #   3: Linear Error-based frame augmentation
+        augmenting_delta_time: 0.             # Maximum time between augments
+        augmenting_pos_error: 0.1             # Maximum position error between augments
+        augmenting_ang_error: 0.1             # Maximum angular error between augments
+        process_noise:
+            pos: 1.0e-0                       # Position process noise
+            vel: 1.0e-1                       # Velocity process noise
+            ang_pos: 1.0e-0                   # Angular Position process noise
+
+        pos_l_in_g: [0.0, 0.0, 0.0]           # Local frame position in global frame
+        ang_l_to_g: 0.0                       # Local frame heading in global frame
+        gps_init_type: 0                      # GPS Initialization Type
+                                              #   0: CONSTANT
+                                              #   1: BASELINE_DIST
+                                              #   2: ERROR_THRESHOLD
+        gps_init_baseline_dist: 100.0         # Baseline distance threshold to initialize
+        gps_init_pos_thresh: 1.0              # Local frame position error threshold
+        gps_init_ang_thresh: 0.1              # Local frame heading error threshold
+        use_root_covariance: true             # Flag to utilize the square root form of covariance
+
         sim_params:
-            seed: 0.0                       # Seed to provide to random number generator
-            use_seed: true                  # Flag to use seed (required for deterministic runs)
-            max_time: 10.0                  # Maximum simulation time
-            number_of_runs: 10              # Number of runs for Multi-Threaded Simulation
-            run_number: 0                   # Start run number (also used to initialize seed)
-            stationary_time: 5.0            # Time to be stationary before motion begins
-            truth_type: "cyclic"            # TruthType:
-                                            #   "Cyclic"
-                                            #   "Spline"
+            seed: 0.0                         # Seed to provide to random number generator
+            use_seed: true                    # Flag to use seed (required for deterministic runs)
+            max_time: 10.0                    # Maximum simulation time
+            number_of_runs: 10                # Number of runs for Multi-Threaded Simulation
+            run_number: 0                     # Start run number (also used to initialize seed)
+            stationary_time: 0.1              # Time to be stationary before motion begins
+            truth_type: "cyclic"              # TruthType:
[…]
