+++ config/example.yaml
@@ -1,70 +1,69 @@
 /EkfCalNode:
     ros__parameters:
-        debug_log_level: 2                  # Debug_Log_Level:
-                                            #   1: FATAL
-                                            #   2: ERROR
-                                            #   3: WARN
-                                            #   4: INFO
-                                            #   5: DEBUG
-        data_logging_on: true               # Flag to enable data logging
-        data_log_rate: 10.0                # EKF body data logging rate
-        augmenting_type: 3                  # Camera state augmentation type
-                                            #   0: All camera frames are augmented
-                                            #   1: Only the primary camera frames are augmented
-                                            #   2: Time-based frame augmentation
-                                            #   3: Linear Error-based frame augmentation
-        augmenting_delta_time: 0.           # Maximum time between augments
-        augmenting_pos_error: 0.1           # Maximum position error between augments
-        augmenting_ang_error: 0.1           # Maximum angular error between augments
+        debug_log_level: 2                    # Debug_Log_Level:
+                                              #   1: FATAL
+                                              #   2: ERROR
+                                              #   3: WARN
+                                              #   4: INFO
+                                              #   5: DEBUG
+        data_logging_on: true                 # Flag to enable data logging
+        data_log_rate: 10.0                   # EKF body data logging rate
+        augmenting_type: 0                    # Camera state augmentation type
+                                              #   0: All camera frames are augmented
+                                              #   1: Only the primary camera frames are augmented
+                                              #   2: Time-based frame augmentation
+                                              #   3: Linear Error-based frame augmentation
+        augmenting_delta_time: 0.             # Maximum time between augments
+        augmenting_pos_error: 0.1             # Maximum position error between augments
+        augmenting_ang_error: 0.1             # Maximum angular error between augments
         process_noise:
-            pos: 1.0e-3                     # Position
-            vel: 1.0e-4                     # Velocity
-            acc: 1.0e-2                     # Acceleration
-            ang_pos: 1.0e-3                 # Angular Position
-            ang_vel: 1.0e-2                 # Angular Velocity
-            ang_acc: 1.0e-5                 # Angular Acceleration
+            pos: 1.0e-0                       # Position process noise
+            vel: 1.0e-1                       # Velocity process noise
+            ang_pos: 1.0e-0                   # Angular Position process noise
 
-        pos_l_in_g: [0.0, 0.0, 0.0]     # Local frame position in global frame
-        ang_l_to_g: 0.0                 # Local frame heading in global frame
-        gps_init_type: 0                # GPS Initialization Type
-                                        #   0: CONSTANT
-                                        #   1: BASELINE_DIST
-                                        #   2: ERROR_THRESHOLD
-        gps_init_baseline_dist: 100.0   # Baseline distance threshold to initialize
-        gps_init_pos_thresh: 1.0        # Local frame position error threshold
-        gps_init_ang_thresh: 0.1        # Local frame heading error threshold
-        use_root_covariance: false      # Flag to utilize the square root form of covariance
+        pos_l_in_g: [0.0, 0.0, 0.0]           # Local frame position in global frame
+        ang_l_to_g: 0.0                       # Local frame heading in global frame
+        gps_init_type: 0                      # GPS Initialization Type
+                                              #   0: CONSTANT
+                                              #   1: BASELINE_DIST
+                                              #   2: ERROR_THRESHOLD
[…]
