### 2024-11-15 7aeb113
Utilize primary camera ID for state augmentation
