### 2024-11-16 cd74413
Minor fixes to primary augmentation interpolation
