### 2024-11-16 0282a3a
Minor feature tracker bugfixes
