### 2024-11-17 776ee96
Fixes to error and time based augmenting
