### 2024-11-17 f1f28f8
Add augmented states plots
