### 2024-11-20 3ea42f5
Remove redundant use_seed parameter
