+++ src/sensors/sim/sim_sensor.hpp
@@ -37,7 +37,6 @@ public:
   {
     bool no_errors {false};        ///< @brief Perfect measurements flag
     double time_error {0.0};       ///< @brief Time offset error
-    double time_bias_error {0.0};  ///< @brief Time offset bias error
     SimRNG rng;                    ///< @brief Random number generator
   } Parameters;
 
@@ -63,7 +62,6 @@ public:
 protected:
   bool m_no_errors {false};              ///< @brief Flag to remove measurement errors
   double m_time_error {0.0};            ///< @brief Time offset error
-  double m_time_bias_error{0.0};         ///< @brief Time offset bias error
   SimRNG m_rng;                          ///< @brief Random number generator
   std::shared_ptr<TruthEngine> m_truth;  ///< @brief Truth engine pointer
 };
