### 2024-11-20 5d1ea05
Improve error based augmentation
