### 2024-11-21 1bd4aa8
Fix bug in augmented delta position
