### 2024-11-22 4a76928
Unit test corrections
