### 2024-11-22 a306c03
Remove data_logging_on input
