+++ src/ekf/update/test/imu_updater_test.cpp
@@ -36,7 +36,6 @@ TEST(test_imu_updater, update) {
 
   unsigned int imu_id{0};
   std::string log_file_directory{""};
-  bool data_logging_on {true};
 
   ImuState imu_state_1, imu_state_2, imu_state_3;
   imu_state_1.SetIsIntrinsic(true);
@@ -53,7 +52,7 @@ TEST(test_imu_updater, update) {
   ekf->RegisterIMU(2, imu_state_3, imu_covariance_3);
 
   auto logger = std::make_shared<DebugLogger>(LogLevel::DEBUG, "");
-  ImuUpdater imu_updater(imu_id, true, true, log_file_directory, data_logging_on, 0.0, logger);
+  ImuUpdater imu_updater(imu_id, true, true, log_file_directory, 0.0, logger);
 
   Eigen::Vector3d acceleration = g_gravity;
   Eigen::Matrix3d acceleration_cov = Eigen::Matrix3d::Identity() * 1e-3;
@@ -117,7 +116,6 @@ TEST(test_imu_updater, imu_prediction_update) {
 
   unsigned int imu_id{0};
   std::string log_file_directory{""};
-  bool data_logging_on {true};
 
   ImuState imu_state;
   imu_state.SetIsExtrinsic(true);
@@ -126,7 +124,7 @@ TEST(test_imu_updater, imu_prediction_update) {
   ekf->RegisterIMU(imu_id, imu_state, imu_cov);
 
   auto logger = std::make_shared<DebugLogger>(LogLevel::DEBUG, "");
-  ImuUpdater imu_updater(imu_id, true, true, log_file_directory, data_logging_on, 0.0, logger);
+  ImuUpdater imu_updater(imu_id, true, true, log_file_directory, 0.0, logger);
 
   Eigen::Vector3d acceleration = g_gravity;
   Eigen::Matrix3d acceleration_cov = Eigen::Matrix3d::Identity() * 1e-3;
@@ -198,7 +196,6 @@ TEST(test_imu_updater, non_initialized_time) {
 
   unsigned int imu_id{0};
   std::string log_file_directory{""};
-  bool data_logging_on {true};
 
   ImuState imu_state;
   imu_state.SetIsExtrinsic(true);
@@ -207,7 +204,7 @@ TEST(test_imu_updater, non_initialized_time) {
   ekf->RegisterIMU(imu_id, imu_state, imu_cov);
 
   auto logger = std::make_shared<DebugLogger>(LogLevel::DEBUG, "");
-  ImuUpdater imu_updater(imu_id, true, true, log_file_directory, data_logging_on, 0.0, logger);
+  ImuUpdater imu_updater(imu_id, true, true, log_file_directory, 0.0, logger);
 
   Eigen::Vector3d acceleration = g_gravity;
   Eigen::Matrix3d acceleration_cov = Eigen::Matrix3d::Identity() * 1e-3;
