### 2024-11-22 ac30b88
Remove duplicate GPS variables
