### 2024-11-22 2f9de83
Align node with up-to-date inputs
