+++ src/application/ros/node/ekf_cal_node.cpp
@@ -55,78 +55,84 @@ EkfCalNode::EkfCalNode()
 : Node("EkfCalNode")
 {
   // Declare EKF Parameters
-  this->declare_parameter("debug_log_level", 0);
-  this->declare_parameter("data_logging_on", false);
-  this->declare_parameter("data_log_rate", 0.0);
-  this->declare_parameter("augmenting_type", 0);
-  this->declare_parameter("augmenting_delta_time", 0.0);
-  this->declare_parameter("augmenting_pos_error", 0.0);
-  this->declare_parameter("augmenting_ang_error", 0.0);
-  this->declare_parameter("process_noise.pos", 1.0e-2);
-  this->declare_parameter("process_noise.vel", 1.0e-2);
-  this->declare_parameter("process_noise.acc", 1.0e-2);
-  this->declare_parameter("process_noise.ang_pos", 1.0e-2);
-  this->declare_parameter("process_noise.ang_vel", 1.0e-2);
-  this->declare_parameter("process_noise.ang_acc", 1.0e-2);
-  this->declare_parameter("pos_l_in_g", std::vector<double>{0, 0, 0});
-  this->declare_parameter("ang_l_to_g", 0.0);
-  this->declare_parameter("init_type", 0);
-  this->declare_parameter("init_pos_thresh", 1.0);
-  this->declare_parameter("init_ang_thresh", 1.0);
-  this->declare_parameter("init_baseline_dist", 1.0);
-  this->declare_parameter("motion_detection_chi_squared", 1.0);
-  this->declare_parameter("imu_noise_scale_factor", 100.0);
+  declare_parameter("debug_log_level", 0);
+  declare_parameter("data_log_rate", 0.0);
+  declare_parameter("augmenting_type", 0);
+  declare_parameter("augmenting_delta_time", 0.0);
+  declare_parameter("augmenting_pos_error", 0.0);
+  declare_parameter("augmenting_ang_error", 0.0);
+  declare_parameter("process_noise.pos", 1.0e-2);
+  declare_parameter("process_noise.vel", 1.0e-2);
+  declare_parameter("process_noise.ang_pos", 1.0e-2);
+  declare_parameter("pos_b_in_l", std::vector<double>{0, 0, 0});
+  declare_parameter("ang_b_to_l", std::vector<double>{0, 0, 0, 0});
+  declare_parameter("pos_l_in_g", std::vector<double>{0, 0, 0});
+  declare_parameter("ang_l_to_g", 0.0);
+  declare_parameter("gps_init_type", 0);
+  declare_parameter("gps_init_baseline_dist", 1.0);
+  declare_parameter("gps_init_pos_thresh", 1.0);
+  declare_parameter("gps_init_ang_thresh", 1.0);
+  declare_parameter("motion_detection_chi_squared", 1.0);
+  declare_parameter("imu_noise_scale_factor", 100.0);
+  declare_parameter("use_root_covariance", false);
+  declare_parameter("use_first_estimate_jacobian", false);
 
   // Declare Sensor Lists
-  this->declare_parameter("imu_list", std::vector<std::string>{});
-  this->declare_parameter("camera_list", std::vector<std::string>{});
-  this->declare_parameter("tracker_list", std::vector<std::string>{});
-  this->declare_parameter("fiducial_list", std::vector<std::string>{});
-  this->declare_parameter("gps_list", std::vector<std::string>{});
+  declare_parameter("imu_list", std::vector<std::string>{});
+  declare_parameter("camera_list", std::vector<std::string>{});
+  declare_parameter("tracker_list", std::vector<std::string>{});
[…]
