### 2024-11-25 d1e68d5
Make sim_params ros yaml compliant
