+++ src/infrastructure/sim/truth_engine_spline.cpp
@@ -112,8 +112,8 @@ Eigen::Vector3d TruthEngineSpline::GetBodyAngularAcceleration(double time)
 }
 
 TruthEngineSpline::TruthEngineSpline(
-  std::vector<std::vector<double>> positions,
-  std::vector<std::vector<double>> angles,
+  std::vector<double> positions,
+  std::vector<double> angles,
   std::vector<double> position_errors,
   std::vector<double> angle_errors,
   double stationary_time,
@@ -123,25 +123,22 @@ TruthEngineSpline::TruthEngineSpline(
 )
 : TruthEngine(max_time, logger)
 {
-  unsigned int spline_size = std::min(positions.size(), angles.size());
+  unsigned int spline_size = std::floor(std::min(positions.size() / 3.0, angles.size() / 3.0));
 
   Eigen::Vector3d pos_errors = StdToEigVec(position_errors);
   Eigen::Vector3d ang_errors = StdToEigVec(angle_errors);
 
   Eigen::MatrixXd pos_mat(3, spline_size);
   Eigen::MatrixXd ang_mat(3, spline_size);
-  for (unsigned int row_index = 0; row_index < spline_size; ++row_index) {
-    Eigen::Vector3d pos;
-    Eigen::Vector3d ang;
-    if (row_index) {
-      pos = rng.VecNormRand(StdToEigVec(positions[row_index]), pos_errors);
-      ang = rng.VecNormRand(StdToEigVec(angles[row_index]), ang_errors);
-    } else {
-      pos = StdToEigVec(positions[row_index]);
-      ang = StdToEigVec(angles[row_index]);
+  for (unsigned int index = 0; index < spline_size; ++index) {
+    Eigen::Vector3d pos {positions[3 * index], positions[3 * index + 1], positions[3 * index + 2]};
+    Eigen::Vector3d ang {angles[3 * index], angles[3 * index + 1], angles[3 * index + 2]};
+    if (index) {
+      pos = rng.VecNormRand(pos, pos_errors);
+      ang = rng.VecNormRand(ang, ang_errors);
     }
-    pos_mat.col(row_index) << pos[0], pos[1], pos[2];
-    ang_mat.col(row_index) << ang[0], ang[1], ang[2];
+    pos_mat.col(index) << pos[0], pos[1], pos[2];
+    ang_mat.col(index) << ang[0], ang[1], ang[2];
   }
 
   Eigen::MatrixXd derivatives(3, 1);
