+++ config/regression/spline.yaml
@@ -16,15 +16,18 @@
             run_number: 0
             stationary_time: 0.0
             truth_type: "spline"
-            positions:
-                - [0.0, 0.0, 0.0]
-                - [0.1, 0.1, 0.1]
-                - [-0.1, -0.1, -0.1]
-                - [0.0, 0.0, 0.0]
-            angles:
-                - [0.00, 0.00, 0.00]
-                - [0.01, 0.01, 0.01]
-                - [0.00, 0.00, 0.00]
+            positions: [
+                0.0, 0.0, 0.0,
+                0.1, 0.1, 0.1,
+                -0.1, -0.1, -0.1,
+                0.0, 0.0, 0.0
+            ]
+            angles: [
+                0.00, 0.00, 0.00,
+                0.01, 0.01, 0.01,
+                -0.01, -0.01, -0.01,
+                0.00, 0.00, 0.00
+            ]
 
         imu_list:
             - "imu_1"
