+++ config/regression/gps.yaml
@@ -13,18 +13,20 @@
             run_number: 0
             stationary_time: 1.0
             truth_type: "spline"
-            positions:
-                - [00.0, 00.0, 00.0]
-                - [10.0, 05.0, 10.0]
-                - [20.0, 20.0, 25.0]
-                - [35.0, 30.0, 30.0]
-                - [40.0, 40.0, 40.0]
-            angles:
-                - [0.0, 0.0, 0.0]
-                - [0.5, 0.0, 0.0]
-                - [0.0, 0.5, 0.0]
-                - [0.0, 0.0, 0.5]
-                - [0.0, 0.0, 0.0]
+            positions: [
+                00.0, 00.0, 00.0,
+                10.0, 05.0, 10.0,
+                20.0, 20.0, 25.0,
+                35.0, 30.0, 30.0,
+                40.0, 40.0, 40.0
+            ]
+            angles: [
+                0.0, 0.0, 0.0,
+                0.5, 0.0, 0.0,
+                0.0, 0.5, 0.0,
+                0.0, 0.0, 0.5,
+                0.0, 0.0, 0.0
+            ]
             pos_l_in_g: [0.0, 0.0, 0.0]
             ang_l_to_g: 0.0
             pos_l_in_g_error: [1.0e-9, 1.0e-9, 1.0e-9]
