### 2024-12-16 6754d55
Restore unit tests
