+++ src/ekf/test/types_test.cpp
@@ -140,32 +140,35 @@ TEST(test_ekf_types, state_plus_equals_vector) {
   State left_state;
   left_state.body_state.pos_b_in_l = Eigen::Vector3d::Ones() * 1.0;
   left_state.body_state.vel_b_in_l = Eigen::Vector3d::Ones() * 2.0;
+  left_state.body_state.acc_b_in_l = Eigen::Vector3d::Ones() * 3.0;
   left_state.body_state.ang_b_to_l = quat;
+  left_state.body_state.ang_vel_b_in_l = Eigen::Vector3d::Ones() * 4.0;
+  left_state.body_state.ang_acc_b_in_l = Eigen::Vector3d::Ones() * 5.0;
 
   ImuState imu_state;
   imu_state.SetIsIntrinsic(true);
   imu_state.SetIsExtrinsic(true);
-  imu_state.pos_i_in_b = Eigen::Vector3d::Ones() * 3.0;
+  imu_state.pos_i_in_b = Eigen::Vector3d::Ones() * 6.0;
   imu_state.ang_i_to_b = quat;
-  imu_state.acc_bias = Eigen::Vector3d::Ones() * 4.0;
-  imu_state.omg_bias = Eigen::Vector3d::Ones() * 5.0;
+  imu_state.acc_bias = Eigen::Vector3d::Ones() * 7.0;
+  imu_state.omg_bias = Eigen::Vector3d::Ones() * 8.0;
 
   GpsState gps_state;
   gps_state.SetIsExtrinsic(true);
-  gps_state.pos_a_in_b = Eigen::Vector3d::Ones() * 6.0;
+  gps_state.pos_a_in_b = Eigen::Vector3d::Ones() * 9.0;
 
   CamState cam_state;
   cam_state.SetIsExtrinsic(true);
-  cam_state.pos_c_in_b = Eigen::Vector3d::Ones() * 7.0;
+  cam_state.pos_c_in_b = Eigen::Vector3d::Ones() * 10.0;
   cam_state.ang_c_to_b = quat;
 
   FidState fid_state;
   fid_state.SetIsExtrinsic(true);
-  fid_state.pos_f_in_l = Eigen::Vector3d::Ones() * 8.0;
+  fid_state.pos_f_in_l = Eigen::Vector3d::Ones() * 11.0;
   fid_state.ang_f_to_l = quat;
 
   AugState aug_state;
-  aug_state.pos_b_in_l = Eigen::Vector3d::Ones() * 9.0;
+  aug_state.pos_b_in_l = Eigen::Vector3d::Ones() * 12.0;
   aug_state.ang_b_to_l = quat;
 
   left_state.imu_states[1] = imu_state;
@@ -178,23 +181,26 @@ TEST(test_ekf_types, state_plus_equals_vector) {
   // Body States
   state_vector.segment<3>(0) = Eigen::Vector3d::Ones() * 1.0;
   state_vector.segment<3>(3) = Eigen::Vector3d::Ones() * 2.0;
-  state_vector.segment<3>(6) = Eigen::Vector3d::Zero();
+  state_vector.segment<3>(6) = Eigen::Vector3d::Ones() * 3.0;
+  state_vector.segment<3>(9) = Eigen::Vector3d::Zero();
+  state_vector.segment<3>(12) = Eigen::Vector3d::Ones() * 4.0;
+  state_vector.segment<3>(15) = Eigen::Vector3d::Ones() * 5.0;
   // IMU States
[…]
