### 2025-01-04 4c3ab53
Command camera data log rate
