+++ config/regression/fiducial.yaml
@@ -70,6 +70,7 @@
                 fiducial: "charuco"
                 pos_stability: 1.0e-9
                 ang_stability: 1.0e-9
+                data_log_rate: 10.0
                 sim_params:
                     time_bias_error: 1.0e-9
                     time_error: 1.0e-9
@@ -88,7 +89,6 @@
                 variance: [0.1, 0.1, 0.1, 0.1, 0.1, 0.1]
                 min_track_length: 0
                 max_track_length: 1
-                data_log_rate: 10.0
                 is_extrinsic: false
                 sim_params:
                     pos_error: [0.0, 0.0, 0.0]
