### 2025-01-04 31c2d7f
Resolve warnings and unit tests
