### 2025-01-04 c311f55
Add coriolis term to IMU prediction
