+++ src/ekf/update/msckf_updater.cpp
@@ -324,6 +324,7 @@ void MsckfUpdater::UpdateEKF(
         H_c.block<2, g_aug_state_size>(2 * i, aug_state_i.index) = H_aug_0;
 
         if (aug_state_i.index + 2 * g_aug_state_size > state_size) {
+          // If measurement is after last augmented state, correlate with the current body state
           H_c.block<2, 3>(2 * i, 0) = H_aug_1.block<2, 3>(0, 0);
           H_c.block<2, 3>(2 * i, 6) = H_aug_1.block<2, 3>(0, 3);
         } else {
