### 2025-01-04 f497c2e
Add missing diagonal term to cov augmentation
