### 2025-01-05 68eee4b
Fix broken covariance augmentation
