### 2025-01-05 03782b6
Re-enable higher order body data logging and plotting
