### 2025-01-05 69daf28
Compare calibrated values to truth in report
