+++ eval/tab_msckf.py
@@ -23,20 +23,22 @@ from bokeh.plotting import ColumnDataSource, figure
 
 import numpy as np
 
-from utilities import calculate_alpha, get_colors, plot_update_timing
+from utilities import calculate_alpha, get_colors, interpolate_error, interpolate_quat_error, \
+    plot_update_timing
 
 
 class tab_msckf:
 
-    def __init__(self, mskcf_dfs, tri_dfs, feat_dfs, args):
-        self.mskcf_dfs = mskcf_dfs
+    def __init__(self, msckf_dfs, tri_dfs, feat_dfs, body_truth_dfs, args):
+        self.msckf_dfs = msckf_dfs
         self.tri_dfs = tri_dfs
         self.feat_dfs = feat_dfs
+        self.body_truth_dfs = body_truth_dfs
 
-        self.alpha = calculate_alpha(len(self.mskcf_dfs))
+        self.alpha = calculate_alpha(len(self.msckf_dfs))
         self.colors = get_colors(args)
 
-    def plot_camera_pos(self):
+    def plot_cam_pos(self):
         """Plot camera position offsets."""
         fig = figure(
             width=800,
@@ -44,29 +46,29 @@ class tab_msckf:
             x_axis_label='Time [s]',
             y_axis_label='Position [m]',
             title='Camera Position')
-        for mskcf_df in self.mskcf_dfs:
-            t_cam = mskcf_df['time']
+        for msckf_df in self.msckf_dfs:
+            t_cam = msckf_df['time']
             fig.line(
                 t_cam,
-                mskcf_df['cam_pos_0'],
+                msckf_df['cam_pos_0'],
                 alpha=self.alpha,
                 color=self.colors[0],
                 legend_label='X')
             fig.line(
                 t_cam,
-                mskcf_df['cam_pos_1'],
+                msckf_df['cam_pos_1'],
                 alpha=self.alpha,
                 color=self.colors[1],
                 legend_label='Y')
             fig.line(
                 t_cam,
-                mskcf_df['cam_pos_2'],
+                msckf_df['cam_pos_2'],
                 alpha=self.alpha,
                 color=self.colors[2],
                 legend_label='Z')
         return fig
 
-    def plot_camera_ang(self):
+    def plot_cam_ang(self):
         """Plot camera angular offsets."""
         fig = figure(
             width=800,
@@ -74,26 +76,113 @@ class tab_msckf:
             x_axis_label='Time [s]',
             y_axis_label='Orientation',
             title='Camera Orientation')
-        for mskcf_df in self.mskcf_dfs:
-            t_cam = mskcf_df['time']
+        for msckf_df in self.msckf_dfs:
+            t_cam = msckf_df['time']
             fig.line(
                 t_cam,
-                mskcf_df['cam_ang_pos_0'],
+                msckf_df['cam_ang_pos_0'],
                 alpha=self.alpha,
                 color=self.colors[0],
                 legend_label='X')
             fig.line(
                 t_cam,
-                mskcf_df['cam_ang_pos_1'],
+                msckf_df['cam_ang_pos_1'],
                 alpha=self.alpha,
                 color=self.colors[1],
                 legend_label='Y')
             fig.line(
                 t_cam,
-                mskcf_df['cam_ang_pos_2'],
+                msckf_df['cam_ang_pos_2'],
                 alpha=self.alpha,
                 color=self.colors[2],
                 legend_label='Z')
+        return fig
+
+    def plot_cam_pos_err(self):
+        """Plot camera extrinsic position errors."""
+        fig = figure(
[…]
