+++ requirements.txt
@@ -1,0 +1,6 @@
+bokeh>=3.6
+matplotlib>=3.5
+numpy>=1.24
+pandas>=2.2
+PyYAML>=6.0
+scipy>=1.8
