+++ src/sensors/sim/sim_gps.cpp
@@ -60,8 +60,8 @@ std::vector<std::shared_ptr<SimGpsMessage>> SimGPS::GenerateMessages()
     Eigen::Vector3d pos_b_in_l = m_truth->GetBodyPosition(measurement_time);
     Eigen::Quaterniond ang_b_to_l = m_truth->GetBodyAngularPosition(measurement_time);
     Eigen::Vector3d pos_a_in_b = m_truth->GetGpsPosition(m_id);
-    Eigen::Vector3d pos_l_in_g = m_truth->GetLocalPosition();
-    double ang_l_to_g = m_truth->GetLocalHeading();
+    Eigen::Vector3d pos_e_in_g = m_truth->GetLocalPosition();
+    double ang_l_to_e = m_truth->GetLocalHeading();
 
     Eigen::Vector3d pos_a_in_l = pos_b_in_l + ang_b_to_l * pos_a_in_b;
     if (!m_no_errors) {
@@ -70,9 +70,9 @@ std::vector<std::shared_ptr<SimGpsMessage>> SimGPS::GenerateMessages()
       pos_a_in_l(2) += m_rng.NormRand(0, m_lla_error(2));
     }
 
-    Eigen::Vector3d antenna_enu = local_to_enu(pos_a_in_l, ang_l_to_g);
+    Eigen::Vector3d antenna_enu = local_to_enu(pos_a_in_l, ang_l_to_e);
 
-    sim_gps_msg->gps_lla = enu_to_lla(antenna_enu, pos_l_in_g);
+    sim_gps_msg->gps_lla = enu_to_lla(antenna_enu, pos_e_in_g);
     sim_gps_msg->pos_covariance = m_lla_error.cwiseProduct(m_lla_error).asDiagonal();
 
     messages.push_back(sim_gps_msg);
