### 2025-01-06 899bef4
Minor MSCKF covariance bug
