### 2025-01-07 1288c6b
Stable extrinsic calibration
