### 2025-01-08 347b73a
Plot angle errors in mrad
