### 2025-01-09 28a85be
Fiducial plotting updates
