### 2025-01-09 ba0a34e
Default to use root covariance
