### 2025-01-09 1f58aa0
Side Quest: Fiducial Updates
