+++ eval/report.py
@@ -87,14 +87,12 @@ def plot_sim_results(config_sets, args):
             tabs.append(tab_msckf(mskcf_dfs, tri_dfs, feat_dfs, body_truth_dfs, args).get_tab())
 
         fiducial_dfs_dict = find_and_read_data_frames(data_dirs, 'fiducial')
-        board_dfs_dict = find_and_read_data_frames(data_dirs, 'board')
         board_truth_dfs_dict = find_and_read_data_frames(data_dirs, 'board_truth')
         for key in sorted(fiducial_dfs_dict.keys()):
             fiducial_dfs = fiducial_dfs_dict[key]
-            tri_dfs = board_dfs_dict[key]
             board_dfs = board_truth_dfs_dict[0]
-            tabs.append(
-                tab_fiducial(fiducial_dfs, tri_dfs, board_dfs, body_truth_dfs, args).get_tab())
+            body_truth_dfs = body_truth_dfs_dict[0]
+            tabs.append(tab_fiducial(fiducial_dfs, board_dfs, body_truth_dfs, args).get_tab())
 
         gps_dfs_dict = find_and_read_data_frames(data_dirs, 'gps')
         for key in sorted(gps_dfs_dict.keys()):
