+++ src/ekf/update/fiducial_updater.cpp
@@ -46,25 +46,21 @@ FiducialUpdater::FiducialUpdater(
 )
 : Updater(fiducial_id, logger),
   m_is_cam_extrinsic(is_cam_extrinsic),
-  m_fiducial_logger(log_file_directory, "fiducial_" + std::to_string(fiducial_id) + ".csv"),
-  m_board_logger(log_file_directory, "board_" + std::to_string(fiducial_id) + ".csv"),
+  m_fiducial_logger(log_file_directory, "fiducial_" + std::to_string(camera_id) + ".csv"),
   m_camera_id(camera_id)
 {
   std::stringstream header;
-  header << "time";
-  header << ",track_size";
+  header << "time,board";
+  header << EnumerateHeader("board_pos", 3);
+  header << EnumerateHeader("board_ang", 4);
   header << EnumerateHeader("cam_pos", 3);
-  header << EnumerateHeader("cam_ang", 4);
+  header << EnumerateHeader("cam_ang_pos", 4);
   header << EnumerateHeader("residual", g_fid_measurement_size);
   if (m_is_cam_extrinsic) {header << EnumerateHeader("cam_cov", g_cam_extrinsic_state_size);}
   header << EnumerateHeader("duration", 1);
 
   m_fiducial_logger.DefineHeader(header.str());
   if (data_log_rate) {m_fiducial_logger.EnableLogging();}
-
-  m_board_logger.DefineHeader("time,board,pos_x,pos_y,pos_z,quat_w,quat_x,quat_y,quat_z");
-  if (data_log_rate) {m_board_logger.EnableLogging();}
-  m_board_logger.SetLogRate(data_log_rate);
 }
 
 void FiducialUpdater::UpdateEKF(
@@ -73,10 +69,6 @@ void FiducialUpdater::UpdateEKF(
 {
   m_logger->Log(
     LogLevel::DEBUG, "Called Fiducial Update for camera ID: " + std::to_string(m_camera_id));
-
-  if (board_track.size() == 0) {
-    return;
-  }
 
   ekf->PredictModel(time);
 
@@ -90,129 +82,85 @@ void FiducialUpdater::UpdateEKF(
     m_is_first_estimate = false;
   }
 
-  /// @todo: Wrap this in a function
-  for (auto board_detection : board_track) {
-    AugState aug_state_i = ekf->GetAugState(m_camera_id, board_detection.frame_id, time);
-
-    const Eigen::Vector3d pos_bi_in_l = aug_state_i.pos_b_in_l;
-    const Eigen::Matrix3d rot_bi_to_l = aug_state_i.ang_b_to_l.toRotationMatrix();
-    const Eigen::Matrix3d rot_c_to_b = m_ang_c_to_b.toRotationMatrix();
-
-    Eigen::Vector3d pos_f_in_c;
-    CvVectorToEigen(board_detection.t_vec_f_in_c, pos_f_in_c);
-    Eigen::Vector3d pos_f_in_l =
-      rot_bi_to_l * ((rot_c_to_b * pos_f_in_c) + m_pos_c_in_b) + pos_bi_in_l;
-
-    Eigen::Quaterniond ang_f_to_c = RodriguesToQuat(board_detection.r_vec_f_to_c);
-    Eigen::Quaterniond ang_f_to_l = aug_state_i.ang_b_to_l * m_ang_c_to_b * ang_f_to_c;
-
-    std::stringstream data_msg;
-    data_msg << std::setprecision(3) << time;
-    data_msg << ",0";
[…]
