+++ eval/tab_fiducial.py
@@ -28,9 +28,8 @@ from utilities import calculate_alpha, get_colors, interpolate_error, interpolat
 
 class tab_fiducial:
 
-    def __init__(self, fiducial_dfs, board_dfs, board_truth_dfs, body_truth_dfs, args):
+    def __init__(self, fiducial_dfs, board_truth_dfs, body_truth_dfs, args):
         self.fiducial_dfs = fiducial_dfs
-        self.board_dfs = board_dfs
         self.board_truth_dfs = board_truth_dfs
         self.body_truth_dfs = body_truth_dfs
 
@@ -45,21 +44,21 @@ class tab_fiducial:
             x_axis_label='Time [s]',
             y_axis_label='Position [m]',
             title='Camera Position')
-        for mskcf_df in self.fiducial_dfs:
-            t_cam = mskcf_df['time']
+        for fiducial_df in self.fiducial_dfs:
+            t_cam = fiducial_df['time']
             fig.line(
                 t_cam,
-                mskcf_df['cam_pos_0'],
+                fiducial_df['cam_pos_0'],
                 alpha=self.alpha,
                 color=self.colors[0])
             fig.line(
                 t_cam,
-                mskcf_df['cam_pos_1'],
+                fiducial_df['cam_pos_1'],
                 alpha=self.alpha,
                 color=self.colors[1])
             fig.line(
                 t_cam,
-                mskcf_df['cam_pos_2'],
+                fiducial_df['cam_pos_2'],
                 alpha=self.alpha,
                 color=self.colors[2])
         return fig
@@ -72,21 +71,21 @@ class tab_fiducial:
             x_axis_label='Time [s]',
             y_axis_label='Orientation',
             title='Camera Orientation')
-        for mskcf_df in self.fiducial_dfs:
-            t_cam = mskcf_df['time']
+        for fiducial_df in self.fiducial_dfs:
+            t_cam = fiducial_df['time']
             fig.line(
                 t_cam,
-                mskcf_df['cam_ang_pos_0'],
+                fiducial_df['cam_ang_pos_0'],
                 alpha=self.alpha,
                 color=self.colors[0])
             fig.line(
                 t_cam,
-                mskcf_df['cam_ang_pos_1'],
+                fiducial_df['cam_ang_pos_1'],
                 alpha=self.alpha,
                 color=self.colors[1])
             fig.line(
                 t_cam,
-                mskcf_df['cam_ang_pos_2'],
+                fiducial_df['cam_ang_pos_2'],
                 alpha=self.alpha,
                 color=self.colors[2])
         return fig
@@ -186,11 +185,11 @@ class tab_fiducial:
             x_axis_label='Time [s]',
             y_axis_label='Position Covariance [m]',
             title='Camera Position Covariance')
-        for mskcf_df in self.fiducial_dfs:
-            t_cam = mskcf_df['time']
-            cam_cov_0 = mskcf_df['cam_cov_0']
-            cam_cov_1 = mskcf_df['cam_cov_1']
-            cam_cov_2 = mskcf_df['cam_cov_2']
+        for fiducial_df in self.fiducial_dfs:
+            t_cam = fiducial_df['time']
+            cam_cov_0 = fiducial_df['cam_cov_0']
+            cam_cov_1 = fiducial_df['cam_cov_1']
+            cam_cov_2 = fiducial_df['cam_cov_2']
             fig.line(
                 t_cam,
                 cam_cov_0,
@@ -219,11 +218,11 @@ class tab_fiducial:
             x_axis_label='Time [s]',
             y_axis_label='Angle Covariance [m]',
             title='Camera Angle Covariance')
-        for mskcf_df in self.fiducial_dfs:
-            t_cam = mskcf_df['time']
-            cam_cov_3 = mskcf_df['cam_cov_3']
[…]
