### 2025-01-09 e25dc84
First pass at single detection fiducials
