+++ src/ekf/update/fiducial_updater.hpp
@@ -56,12 +56,8 @@ public:
   /// @param ekf EKF pointer
   /// @param time Time of update
   /// @param board_track Board track to be used for state update
-  /// @param pos_error Standard deviation of the position error
-  /// @param ang_error Standard deviation of the angle error
   ///
-  void UpdateEKF(
-    std::shared_ptr<EKF> ekf,
-    double time, BoardTrack board_track, double pos_error, double ang_error);
+  void UpdateEKF(std::shared_ptr<EKF> ekf, double time, BoardDetection board_detection);
 
 private:
   bool m_is_cam_extrinsic;
