+++ eval/tab_fiducial.py
@@ -252,100 +252,30 @@ class tab_fiducial:
             y_axis_label='Position [m]',
             title='Fiducial Position in Camera Frame')
 
-        err_px = collections.defaultdict(list)
-        err_py = collections.defaultdict(list)
-        err_pz = collections.defaultdict(list)
-
-        for fiducial_df, true_df in zip(self.fiducial_dfs, self.board_truth_dfs):
+        for fiducial_df in self.fiducial_dfs:
             time = fiducial_df['time']
-            board = fiducial_df['board']
             board_px = fiducial_df['board_pos_0']
             board_py = fiducial_df['board_pos_1']
             board_pz = fiducial_df['board_pos_2']
 
-            true_px = true_df['pos_x']
-            true_py = true_df['pos_y']
-            true_pz = true_df['pos_z']
-
-            for (t, b, px, py, pz) in zip(time, board, board_px, board_py, board_pz):
-                err_px[t].append(px - true_px[int(b)])
-                err_py[t].append(py - true_py[int(b)])
-                err_pz[t].append(pz - true_pz[int(b)])
-        times = []
-        mean_px = []
-        mean_py = []
-        mean_pz = []
-        std_px = []
-        std_py = []
-        std_pz = []
-
-        for time in err_px:
-            times.append(time)
-            mean_px.append(np.mean(err_px[time]))
-            mean_py.append(np.mean(err_py[time]))
-            mean_pz.append(np.mean(err_pz[time]))
-            std_px.append(np.std(err_px[time]))
-            std_py.append(np.std(err_py[time]))
-            std_pz.append(np.std(err_pz[time]))
-
-        times = np.array(times)
-        mean_px = np.array(mean_px)
-        mean_py = np.array(mean_py)
-        mean_pz = np.array(mean_pz)
-        std_px = np.array(std_px)
-        std_py = np.array(std_py)
-        std_pz = np.array(std_pz)
-
-        t_indices = times.argsort()
-        times = times[t_indices]
-        mean_px = mean_px[t_indices]
-        mean_py = mean_py[t_indices]
-        mean_pz = mean_pz[t_indices]
-        std_px = std_px[t_indices]
-        std_py = std_py[t_indices]
-        std_pz = std_pz[t_indices]
-
-        fig.line(times, mean_px, color=self.colors[0])
-        fig.line(times, mean_py, color=self.colors[1])
-        fig.line(times, mean_pz, color=self.colors[2])
-
-        cds_x = ColumnDataSource({'base': times,
-                                  'lower': mean_px - std_px,
-                                  'upper': mean_px + std_px})
-        cds_y = ColumnDataSource({'base': times,
-                                  'lower': mean_py - std_py,
-                                  'upper': mean_py + std_py})
-        cds_z = ColumnDataSource({'base': times,
-                                  'lower': mean_pz - std_pz,
-                                  'upper': mean_pz + std_pz})
-
-        fig.add_layout(
-            Band(
-                base='base',
-                lower='lower',
-                upper='upper',
-                fill_alpha=0.3,
-                source=cds_x,
-                fill_color=self.colors[0],
-                line_color=self.colors[0]))
-        fig.add_layout(
-            Band(
-                base='base',
-                lower='lower',
-                upper='upper',
-                fill_alpha=0.3,
-                source=cds_y,
-                fill_color=self.colors[1],
-                line_color=self.colors[1]))
-        fig.add_layout(
-            Band(
-                base='base',
-                lower='lower',
-                upper='upper',
-                fill_alpha=0.3,
-                source=cds_z,
-                fill_color=self.colors[2],
-                line_color=self.colors[2]))
+            fig.line(
+                time,
+                board_px,
+                alpha=self.alpha,
[…]

[…]
