### 2025-01-10 ecc1a45
Make plot labels consistent
