### 2025-01-10 7fbc365
Improvements to fiducial tracking
