### 2025-01-11 4ae5dd5
Add out of bounds covariance checks
