### 2025-01-12 1cad0b4
log_directory naming convention
