+++ src/sensors/gps.cpp
@@ -32,7 +32,7 @@ GPS::GPS(GPS::Parameters params)
   m_gps_updater(
     m_id,
     params.is_extrinsic,
-    params.output_directory,
+    params.log_directory,
     params.data_log_rate,
     params.logger)
 {
