+++ src/application/sim/simulation.cpp
@@ -67,7 +67,7 @@ void LoadSensorParams(
   Sensor::Parameters & params,
   YAML::Node node,
   std::string name,
-  std::string out_dir,
+  std::string log_directory,
   std::shared_ptr<EKF> ekf,
   std::shared_ptr<DebugLogger> debug_logger
 )
@@ -76,7 +76,7 @@ void LoadSensorParams(
   params.rate = node["rate"].as<double>(1.0);
   params.data_log_rate = node["data_log_rate"].as<double>(0.0);
   params.name = name;
-  params.output_directory = out_dir;
+  params.log_directory = log_directory;
   params.ekf = ekf;
   params.logger = debug_logger;
 }
@@ -92,13 +92,13 @@ void LoadSimSensorParams(
 void LoadTrackerParams(
   Tracker::Parameters & params,
   std::string name,
-  std::string out_dir,
+  std::string log_directory,
   std::shared_ptr<EKF> ekf,
   std::shared_ptr<DebugLogger> debug_logger
 )
 {
   params.name = name;
-  params.output_directory = out_dir;
+  params.log_directory = log_directory;
   params.ekf = ekf;
   params.logger = debug_logger;
 }
@@ -126,20 +126,20 @@ Eigen::VectorXd LoadProcessNoise(YAML::Node process_noise_node)
 int main(int argc, char * argv[])
 {
   const cv::String keys =
-    "{@config        | | Input YAML configuration file }"
-    "{@out_dir       | | Output directory for logs     }"
-    "{help h usage ? | | print this help message       }"
+    "{@config        | <none> | Input YAML configuration file }"
+    "{@log_directory | ~/log/ | Output directory for logs     }"
+    "{help h usage ? |        | print this help message       }"
   ;
 
   cv::CommandLineParser parser(argc, argv, keys);
   parser.about("ekf_cal Simulation: " + std::string(EKF_CAL_VERSION));
-  if (parser.has("help") || (!parser.has("@config") || !parser.has("@out_dir"))) {
+  if (parser.has("help") || (!parser.has("@config") || !parser.has("@log_directory"))) {
     parser.printMessage();
     return 0;
   }
 
   std::string config = parser.get<std::string>("@config");
-  std::string out_dir = parser.get<std::string>("@out_dir");
+  std::string log_directory = parser.get<std::string>("@log_directory");
 
   // Define sensors to use (load config from yaml)
   YAML::Node root = YAML::LoadFile(config);
@@ -161,14 +161,14 @@ int main(int argc, char * argv[])
   EKF::Parameters ekf_params;
   YAML::Node ros_params = root["/EkfCalNode"]["ros__parameters"];
   unsigned int debug_log_level = ros_params["debug_log_level"].as<unsigned int>(0U);
-  auto debug_logger = std::make_shared<DebugLogger>(debug_log_level, out_dir);
+  auto debug_logger = std::make_shared<DebugLogger>(debug_log_level, log_directory);
   debug_logger->Log(LogLevel::INFO, "EKF CAL Version: " + std::string(EKF_CAL_VERSION));
 
   // EKF parameters
   double data_log_rate = ros_params["data_log_rate"].as<double>(0.0);
   ekf_params.debug_logger = debug_logger;
   ekf_params.data_log_rate = data_log_rate;
-  ekf_params.log_directory = out_dir;
+  ekf_params.log_directory = log_directory;
   ekf_params.augmenting_type =
     static_cast<AugmentationType>(ros_params["augmenting_type"].as<unsigned int>(1));
[…]
