### 2025-01-12 a6613e2
Pass by value for opencv matrices
