+++ src/trackers/feature_tracker.hpp
@@ -110,7 +110,7 @@ public:
   /// @param img_in Input frame
   /// @param img_out Output frame with drawn track lines
   ///
-  void Track(double time, int frame_id, cv::Mat & img_in, cv::Mat & img_out);
+  void Track(double time, int frame_id, cv::Mat img_in, cv::Mat img_out);
 
 
   ///
