### 2025-01-12 e683b5b
Save output image as RGB
