### 2025-01-12 23ac869
Use directory for default log location
