### 2025-01-12 87316b5
Use message casting instead of manual conversions
