### 2025-01-12 928f3b2
Assume incoming images are greyscale
