+++ src/ekf/update/imu_updater.cpp
@@ -69,10 +69,12 @@ void ImuUpdater::UpdateEKF(
   Eigen::Matrix3d acceleration_covariance, Eigen::Vector3d angular_rate,
   Eigen::Matrix3d angular_rate_covariance)
 {
+  double local_time = ekf->CalculateLocalTime(time);
+
   // Check for zero velocity
   if (ZeroAccelerationUpdate(
       ekf,
-      time,
+      local_time,
       acceleration,
       acceleration_covariance,
       angular_rate,
@@ -114,7 +116,7 @@ void ImuUpdater::UpdateEKF(
     KalmanUpdate(ekf, H, resid, R);
   }
 
-  ekf->PredictModel(time);
+  ekf->PredictModel(local_time);
 
   auto t_end = std::chrono::high_resolution_clock::now();
   auto t_execution = std::chrono::duration_cast<std::chrono::microseconds>(t_end - t_start);
@@ -126,7 +128,7 @@ void ImuUpdater::UpdateEKF(
   resid.segment<3>(0) = acceleration - ekf->m_state.body_state.acc_b_in_l;
   resid.segment<3>(3) = angular_rate - ekf->m_state.body_state.ang_vel_b_in_l;
 
-  msg << time;
+  msg << local_time;
   msg << ",0," << ekf->GetMotionDetectionChiSquared();
   msg << VectorToCommaString(ekf->m_state.imu_states[m_id].pos_i_in_b);
   msg << QuaternionToCommaString(ekf->m_state.imu_states[m_id].ang_i_to_b);
@@ -146,7 +148,7 @@ void ImuUpdater::UpdateEKF(
   msg << VectorToCommaString(angular_rate);
   msg << VectorToCommaString(resid);
   msg << "," << t_execution.count();
-  m_data_logger.RateLimitedLog(msg.str(), time);
+  m_data_logger.RateLimitedLog(msg.str(), local_time);
 
   ekf->LogBodyStateIfNeeded(t_execution.count());
 }
@@ -180,7 +182,7 @@ Eigen::MatrixXd ImuUpdater::GetZeroAccelerationJacobian(std::shared_ptr<EKF> ekf
 
 bool ImuUpdater::ZeroAccelerationUpdate(
   std::shared_ptr<EKF> ekf,
-  double time,
+  double local_time,
   Eigen::Vector3d acceleration,
   Eigen::Matrix3d acceleration_covariance,
   Eigen::Vector3d angular_rate,
@@ -229,7 +231,7 @@ bool ImuUpdater::ZeroAccelerationUpdate(
   /// @todo Test stationary rotation
   // AngularUpdate(ekf, angular_rate, angular_rate_covariance);
 
-  ekf->PredictModel(time);
+  ekf->PredictModel(local_time);
 
   // Update Jacobian
   H = GetZeroAccelerationJacobian(ekf);
@@ -260,7 +262,7 @@ bool ImuUpdater::ZeroAccelerationUpdate(
   // Write outputs
   std::stringstream msg;
 
-  msg << time;
+  msg << local_time;
   msg << ",1," << score;
   msg << VectorToCommaString(ekf->m_state.imu_states[m_id].pos_i_in_b);
   msg << QuaternionToCommaString(ekf->m_state.imu_states[m_id].ang_i_to_b);
@@ -288,7 +290,7 @@ bool ImuUpdater::ZeroAccelerationUpdate(
   }
 
   msg << "," << t_execution.count();
-  m_data_logger.RateLimitedLog(msg.str(), time);
+  m_data_logger.RateLimitedLog(msg.str(), local_time);
 
   ekf->LogBodyStateIfNeeded(t_execution.count());
 
