+++ src/ekf/update/msckf_updater.hpp
@@ -54,13 +54,13 @@ public:
   ///
   /// @brief Triangulate feature seen from multiple camera frames
   /// @param ekf EKF pointer
-  /// @param time Time of update
+  /// @param local_time Measurement in EKF time
   /// @param feature_track Single feature track
   /// @param pos_f_in_l Output estimate of feature position in camera frame given observations
   /// @return If triangulation was successful
   ///
   bool TriangulateFeature(
-    double time,
+    double local_time,
     std::shared_ptr<EKF> ekf,
     FeatureTrack & feature_track,
     Eigen::Vector3d & pos_f_in_l);
