+++ src/sensors/camera.cpp
@@ -62,7 +62,8 @@ void Camera::Callback(std::shared_ptr<CameraMessage> camera_message)
 
   if (!camera_message->image.empty()) {
     if (!m_trackers.empty() || !m_fiducials.empty()) {
-      m_ekf->PredictModel(camera_message->time);
+      double local_time = m_ekf->CalculateLocalTime(camera_message->time);
+      m_ekf->PredictModel(local_time);
 
       unsigned int frameID = GenerateFrameID();
 
