### 2025-01-14 b048596
Update keypoint and match tests
