+++ src/trackers/feature_tracker.cpp
@@ -75,7 +75,7 @@ cv::Ptr<cv::FeatureDetector> FeatureTracker::InitFeatureDetector(
       break;
     case Detector::ORB:
       feature_detector =
-        cv::ORB::create(1000, 1.2f, 8, 31, 0, 2, cv::ORB::HARRIS_SCORE, 31, threshold);
+        cv::ORB::create(200, 1.2f, 8, 31, 0, 2, cv::ORB::HARRIS_SCORE, 31, threshold);
       break;
     case Detector::SIFT:
       feature_detector = cv::SIFT::create();
@@ -143,7 +143,7 @@ std::vector<cv::KeyPoint> FeatureTracker::GridFeatures(
   cv::Mat grid_2d = cv::Mat::zeros(size, CV_8UC1);
 
   std::vector<cv::KeyPoint> grid_key_points;
-  for (size_t i = 0; i < key_points.size(); i++) {
+  for (unsigned int i = 0; i < key_points.size(); i++) {
     // Get current left keypoint, check that it is in bounds
     cv::KeyPoint kpt = key_points.at(i);
     int x = static_cast<int>(kpt.pt.x);
@@ -188,65 +188,51 @@ void FeatureTracker::Track(double time, int frame_id, cv::Mat img_in, cv::Mat im
   /// @todo create occupancy grid of key_points using minimal pixel distance
   curr_key_points = GridFeatures(curr_key_points, img_down.rows, img_down.cols);
 
-  double threshold_dist =
-    0.1 * sqrt(
-    static_cast<double>(
-      down_sample_size.height * down_sample_size.height +
-      down_sample_size.width * down_sample_size.width
-    ));
+  // double threshold_dist =
+  //   0.1 * sqrt(
+  //   static_cast<double>(
+  //     down_sample_size.height * down_sample_size.height +
+  //     down_sample_size.width * down_sample_size.width
+  //   ));
 
   cv::Mat curr_descriptors;
   m_descriptor_extractor->compute(img_down, curr_key_points, curr_descriptors);
   curr_descriptors.convertTo(curr_descriptors, CV_32F);
-  cv::drawKeypoints(img_down, curr_key_points, img_out);
 
   m_logger->Log(LogLevel::DEBUG, "Called Tracker for frame ID: " + std::to_string(frame_id));
 
   if (m_prev_descriptors.rows > 0 && curr_descriptors.rows > 0) {
     std::vector<std::vector<cv::DMatch>> matches_forward, matches_backward;
-    std::vector<cv::DMatch> matches_good;
 
     /// @todo Mask using maximum distance from predicted IMU rotations
     m_descriptor_matcher->knnMatch(m_prev_descriptors, curr_descriptors, matches_forward, 2);
     m_descriptor_matcher->knnMatch(curr_descriptors, m_prev_descriptors, matches_backward, 2);
 
-    ratio_test(matches_forward);
-    ratio_test(matches_backward);
+    // Perform test on matches
+    RatioTest(matches_forward);
+    RatioTest(matches_backward);
+    std::vector<cv::DMatch> matches_symmetry, matches_ransac, matches_final;
+    SymmetryTest(matches_forward, matches_backward, matches_symmetry);
+    RANSAC(matches_symmetry, curr_key_points, matches_ransac);
+    DistanceTest(matches_ransac, curr_key_points, matches_final);
 
-    /// @todo: Symmetry testing?
-
-    matches_good.reserve(matches_forward.size());
-    for (unsigned int i = 0; i < matches_forward.size(); ++i) {
-      for (unsigned int j = 0; j < matches_forward[i].size(); j++) {
-        cv::Point2f point_old = m_prev_key_points[matches_forward[i][j].queryIdx].pt;
[…]
