+++ src/infrastructure/sim/truth_engine_smoother.cpp
@@ -1,0 +1,243 @@
+// Copyright 2024 Jacob Hartzer
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+#include "infrastructure/sim/truth_engine_smoother.hpp"
+
+#include <algorithm>
+#include <memory>
+#include <vector>
+
+#include "utility/sim/sim_rng.hpp"
+#include "utility/type_helper.hpp"
+
+
+Eigen::Vector3d TruthEngineSmoother::GetInterpolatedValue(
+  double time,
+  std::vector<Eigen::Vector3d> & values
+)
+{
+  double whole, alpha;
+  alpha = std::modf(time * m_rate, &whole);
+  auto index = static_cast<unsigned int>(whole);
+  return values[index] * (1 - alpha) + values[index + 1] * alpha;
+}
+
+Eigen::Vector3d TruthEngineSmoother::GetBodyPosition(double time)
+{
+  double relative_time = time - m_stationary_time;
+  if (IsTimeInvalid(relative_time) ) {
+    return Eigen::Vector3d::Zero(3);
+  } else {
+    return GetInterpolatedValue(time, m_pos);
+  }
+}
+
+Eigen::Vector3d TruthEngineSmoother::GetBodyVelocity(double time)
+{
+  double relative_time = time - m_stationary_time;
+  if (IsTimeInvalid(relative_time) ) {
+    return Eigen::Vector3d::Zero(3);
+  } else {
+    return GetInterpolatedValue(time, m_vel);
+  }
+}
+
+Eigen::Vector3d TruthEngineSmoother::GetBodyAcceleration(double time)
+{
+  double relative_time = time - m_stationary_time;
+  if (IsTimeInvalid(relative_time) ) {
+    return Eigen::Vector3d::Zero(3);
+  } else {
+    return GetInterpolatedValue(time, m_acc);
+  }
+}
+
+Eigen::Quaterniond TruthEngineSmoother::GetBodyAngularPosition(double time)
+{
+  double relative_time = time - m_stationary_time;
+  if (IsTimeInvalid(relative_time)) {
+    return Eigen::Quaterniond{1.0, 0.0, 0.0, 0.0};
+  } else {
+    Eigen::Vector3d euler_angles_in = GetInterpolatedValue(time, m_ang);
+    Eigen::Quaterniond angular_position = EigVecToQuat(euler_angles_in);
+    return angular_position;
+  }
+}
+
+Eigen::Vector3d TruthEngineSmoother::GetBodyAngularRate(double time)
+{
+  double relative_time = time - m_stationary_time;
+  if (IsTimeInvalid(relative_time)) {
+    return Eigen::Vector3d::Zero(3);
+  } else {
+    return GetInterpolatedValue(time, m_ang_vel);
+  }
+}
+
+Eigen::Vector3d TruthEngineSmoother::GetBodyAngularAcceleration(double time)
+{
+  double relative_time = time - m_stationary_time;
[…]
