+++ src/infrastructure/sim/truth_engine_smoother.hpp
@@ -1,0 +1,118 @@
+// Copyright 2024 Jacob Hartzer
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+#ifndef INFRASTRUCTURE__SIM__TRUTH_ENGINE_SMOOTHER_HPP_
+#define INFRASTRUCTURE__SIM__TRUTH_ENGINE_SMOOTHER_HPP_
+
+#include <eigen3/Eigen/Eigen>
+
+#include <memory>
+#include <vector>
+
+#include "infrastructure/sim/truth_engine.hpp"
+#include "utility/sim/sim_rng.hpp"
+
+///
+/// @class TruthEngineSmoother
+/// @brief Truth for simulation
+///
+class TruthEngineSmoother : public TruthEngine
+{
+public:
+  ///
+  /// @brief Smoother-based truth engine
+  /// @param times Times for control points
+  /// @param positions Position control points
+  /// @param angles Angular position control points
+  /// @param position_errors Standard deviations of position control points
+  /// @param angle_errors Standard deviations of angular control points
+  /// @param stationary_time Time to be stationary before beginning motion
+  /// @param max_time Maximum simulation time
+  /// @param max_rate Maximum sensor rate
+  /// @param logger Debug logger pointer
+  /// @param rng Random number generator
+  ///
+  TruthEngineSmoother(
+    std::vector<double> times,
+    std::vector<double> positions,
+    std::vector<double> angles,
+    std::vector<double> position_errors,
+    std::vector<double> angle_errors,
+    double stationary_time,
+    double max_time,
+    double max_rate,
+    std::shared_ptr<DebugLogger> logger,
+    SimRNG rng
+  );
+
+  ///
+  /// @brief True body position getter
+  /// @param time Simulation time
+  ///
+  Eigen::Vector3d GetBodyPosition(double time);
+
+  ///
+  /// @brief True body velocity getter
+  /// @param time Simulation time
+  ///
+  Eigen::Vector3d GetBodyVelocity(double time);
+
+  ///
+  /// @brief True body acceleration getter
+  /// @param time Simulation time
+  ///
+  Eigen::Vector3d GetBodyAcceleration(double time);
+
+  ///
+  /// @brief True body orientation quaternion getter
+  /// @param time Simulation time
+  ///
+  Eigen::Quaterniond GetBodyAngularPosition(double time);
+
+  ///
+  /// @brief True body angular rate getter
+  /// @param time Simulation time
+  ///
+  Eigen::Vector3d GetBodyAngularRate(double time);
+
+  ///
+  /// @brief True body angular acceleration getter
+  /// @param time Simulation time
+  ///
+  Eigen::Vector3d GetBodyAngularAcceleration(double time);
+
+private:
+  Eigen::Vector3d GetInterpolatedValue(
+    double time,
+    std::vector<Eigen::Vector3d> & values
+  );
+
+  bool IsTimeInvalid(double time);
+
+
+  double m_stationary_time {0.1};
[…]

[…]
