+++ src/infrastructure/sim/test/truth_engine_smoother_test.cpp
@@ -23,133 +23,6 @@
 #include "utility/sim/sim_rng.hpp"
 #include "utility/type_helper.hpp"
 
-
-// TEST(test_TruthEngineSmoother, Constructor) {
-//   double stationary_time {0.0};
-//   double max_time {1.0};
-//   std::vector<double> times{0, 1, 2, 3};
-//   std::vector<double> positions;
-//   std::vector<double> angles;
-
-//   for (unsigned int i = 0; i < 4 * 3; ++i) {
-//     positions.push_back(0.0);
-//     angles.push_back(0.0);
-//   }
-
-//   auto pos_errs = std::vector<double>{0.0, 0.0, 0.0};
-//   auto ang_errs = std::vector<double>{0.0, 0.0, 0.0};
-
-//   auto logger = std::make_shared<DebugLogger>(LogLevel::DEBUG, "");
-
-//   SimRNG rng;
-
-//   TruthEngineSmoother truth_engine_smoother(
-//     times, positions, angles, pos_errs, ang_errs, stationary_time, max_time, 100.0, logger, rng);
-// }
-
-
-// TEST(test_TruthEngineSmoother, Constant_Velocity) {
-//   double stationary_time {0.0};
-//   double max_time {3.0};
-//   std::vector<double> times {0, 1, 2, 3, 4, 5};
-//   std::vector<double> positions {0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 2.0, 2.0, 2.0, 3.0, 3.0, 3.0};
-//   std::vector<double> angles {0.0, 0.0, 0.0, 0.1, 0.1, 0.1, 0.2, 0.2, 0.2, 0.3, 0.3, 0.3};
-
-//   auto pos_errs = std::vector<double>{0.0, 0.0, 0.0};
-//   auto ang_errs = std::vector<double>{0.0, 0.0, 0.0};
-
-//   auto logger = std::make_shared<DebugLogger>(LogLevel::DEBUG, "");
-
-//   SimRNG rng;
-
-//   TruthEngineSmoother truth_engine_spline(
-//     times, positions, angles, pos_errs, ang_errs, stationary_time, max_time, 100.0, logger, rng);
-
-//   Eigen::Vector3d pos_n = truth_engine_spline.GetBodyPosition(-1.0);
-//   Eigen::Vector3d pos_0 = truth_engine_spline.GetBodyPosition(0.0);
-//   Eigen::Vector3d pos_1 = truth_engine_spline.GetBodyPosition(1.0);
-//   Eigen::Vector3d pos_2 = truth_engine_spline.GetBodyPosition(2.0);
-//   Eigen::Vector3d pos_3 = truth_engine_spline.GetBodyPosition(3.0);
-//   Eigen::Vector3d pos_4 = truth_engine_spline.GetBodyPosition(4.0);
-
-//   EXPECT_TRUE(EXPECT_EIGEN_NEAR(pos_n, Eigen::Vector3d{0, 0, 0}, 1e-6));
-//   EXPECT_NEAR(pos_0[0], positions[0], 1e-6);
-//   EXPECT_NEAR(pos_1[0], positions[3], 1e-6);
-//   EXPECT_NEAR(pos_2[0], positions[6], 1e-6);
[…]
