### 2025-01-20 2218cf9
Calculate addition body errors
