### 2025-01-24 #7292 @arfon
@AKuederle @stevemacenski @abougouffa @mhu48 @Mechazo11 :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is _EKF_CAL: Extended Kalman Filter-based Calibration and Localization_

Many thanks
Arfon
