### 2025-02-02 0afcf01
Calculate IMU Durations
