### 2025-02-02 6d719f5
Add smoother to regression set
