### 2025-02-04 #7292 @arfon
@gautam-sharma1 @gareth-cross @sea-bass 👋 would any of you be willing to review this submission for JOSS? The submission under consideration is EKF_CAL: Extended Kalman Filter-based Calibration and Localization

Many thanks
Arfon
