### 2025-02-12 82c298f
Add intrinsics to IMU jacobian
