### 2025-02-12 0428513
Plotting updates
