+++ eval/tab_imu.py
@@ -20,7 +20,8 @@ from bokeh.models import Range1d, TabPanel
 from bokeh.plotting import figure
 import numpy as np
 
-from utilities import calculate_alpha, get_colors, interpolate_quat_error, plot_update_timing
+from utilities import calculate_alpha, get_colors, interpolate_quat_error, interpolate_error, \
+    plot_update_timing
 
 
 class tab_imu:
@@ -131,21 +132,21 @@ class tab_imu:
             y_axis_label='Angular Rate [rad/s]',
             title='Angular Rate Residuals')
         for imu_df in self.imu_dfs:
-            t_imu = imu_df['time']
+            time = imu_df['time']
             fig.line(
-                t_imu,
+                time,
                 imu_df['residual_3'],
                 alpha=self.alpha,
                 color=self.colors[0],
                 legend_label='X')
             fig.line(
-                t_imu,
+                time,
                 imu_df['residual_4'],
                 alpha=self.alpha,
                 color=self.colors[1],
                 legend_label='Y')
             fig.line(
-                t_imu,
+                time,
                 imu_df['residual_5'],
                 alpha=self.alpha,
                 color=self.colors[2],
@@ -167,59 +168,14 @@ class tab_imu:
         pos_2 = np.zeros([len(self.imu_dfs), n])
 
         for i in range(len(self.imu_dfs)):
-            t_imu[i, :] = self.imu_dfs[i]['time'][:n]
+            time = self.imu_dfs[i]['time'][:n]
             pos_0[i, :] = np.array(self.imu_dfs[i]['imu_pos_0'][:n])
             pos_1[i, :] = np.array(self.imu_dfs[i]['imu_pos_1'][:n])
             pos_2[i, :] = np.array(self.imu_dfs[i]['imu_pos_2'][:n])
-            fig.line(
-                t_imu[i, :], pos_0[i, :],
-                alpha=self.alpha,
-                color=self.colors[0],
-                legend_label='X')
-            fig.line(
-                t_imu[i, :], pos_1[i, :],
-                alpha=self.alpha,
-                color=self.colors[1],
-                legend_label='Y')
-            fig.line(
-                t_imu[i, :], pos_2[i, :],
-                alpha=self.alpha,
-                color=self.colors[2],
-                legend_label='Z')
+            fig.line(time, pos_0[i, :], alpha=self.alpha, color=self.colors[0], legend_label='X')
+            fig.line(time, pos_1[i, :], alpha=self.alpha, color=self.colors[1], legend_label='Y')
+            fig.line(time, pos_2[i, :], alpha=self.alpha, color=self.colors[2], legend_label='Z')
 
-        pos_0_std = np.std(pos_0, axis=0)
-        pos_1_std = np.std(pos_1, axis=0)
-        pos_2_std = np.std(pos_2, axis=0)
-        fig.line(
-            t_imu[0, :], +3.0 * pos_0_std,
-            line_dash='dashed',
-            color='red',
-            alpha=self.alpha)
-        fig.line(
-            t_imu[0, :], +3.0 * pos_1_std,
-            line_dash='dashed',
-            color='red',
-            alpha=self.alpha)
-        fig.line(
-            t_imu[0, :], +3.0 * pos_2_std,
-            line_dash='dashed',
-            color='red',
-            alpha=self.alpha)
-        fig.line(
-            t_imu[0, :], -3.0 * pos_0_std,
-            line_dash='dashed',
-            color='red',
[…]
