### 2025-02-13 6f2495c
Fix GPS log rate
