### 2025-02-13 a9bfaab
Simplify regression set
