+++ config/regression/one_imu_cam.yaml
@@ -1,22 +1,21 @@
 /EkfCalNode:
     ros__parameters:
         debug_log_level: 2
-        data_log_rate: 10.0
+        data_log_rate: 100.0
         process_noise:
             pos: 1.0e-0
             vel: 1.0e-1
-            ang_pos: 1.0e-0
+            acc: 1.0e-2
+            ang_pos: 1.0e-1
         sim_params:
             seed: 1.0
             max_time: 50.0
-            stationary_time: 0.1
+            stationary_time: 0.5
             truth_type: "cyclic"
             pos_frequency: [0.3, 0.5, 0.7]
             ang_frequency: [0.2, 0.3, 0.4]
             ang_amplitude: 0.3
             number_of_runs: 10
-            pos_b_in_l_error: [0.5, 0.5, 0.5]
-            ang_b_to_l_error: [0.3, 0.3, 0.3]
             run_number: 0
 
         imu_list:
@@ -36,23 +35,17 @@
             imu_1:
                 is_extrinsic: false
                 is_intrinsic: false
-                rate: 1000.0
+                rate: 100.0
                 topic: "/imu_1"
-                variance: [0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01]
+                variance: [0.01, 0.01, 0.01, 0.01, 0.01, 0.01,
+                           1e-6, 1e-6, 1e-6,
+                           1e-6, 1e-6, 1e-6]
                 pos_i_in_b: [0.0, 0.0, 0.0]
                 ang_i_to_b: [1.0, 0.0, 0.0, 0.0]
-                acc_bias: [0.0, 0.0, 0.0]
-                omg_bias: [0.0, 0.0, 0.0]
-                data_log_rate: 10.0
+                data_log_rate: 100.0
                 sim_params:
-                    time_bias_error: 0.0
-                    time_error: 1.0e-6
-                    pos_error: [0.0, 0.0, 0.0]
-                    ang_error: [0.0, 0.0, 0.0]
-                    acc_error: [1.0e-2, 1.0e-2, 1.0e-2]
+                    acc_error: [1.0e-1, 1.0e-1, 1.0e-1]
                     omg_error: [1.0e-3, 1.0e-3, 1.0e-3]
-                    acc_bias_error: [0.0, 0.0, 0.0]
-                    omg_bias_error: [0.0, 0.0, 0.0]
 
         camera:
             cam_1:
