+++ config/regression/one_imu_cam_translation.yaml
@@ -8,14 +8,12 @@
             ang_pos: 1.0e-0
         sim_params:
             seed: 1.0
-            max_time: 60.0
+            max_time: 10.0
             stationary_time: 0.1
             truth_type: "cyclic"
             pos_frequency: [0.3, 0.5, 0.7]
             ang_amplitude: 0.3
-            number_of_runs: 10
-            pos_b_in_l_error: [0.5, 0.5, 0.5]
-            ang_b_to_l_error: [0.3, 0.3, 0.3]
+            number_of_runs: 3
             run_number: 0
 
         imu_list:
@@ -54,9 +52,10 @@
                 ang_c_to_b: [-0.5, 0.5, -0.5, 0.5]
                 variance: [0.1, 0.1, 0.1, 0.1, 0.1, 0.1]
                 tracker: "orb"
-                is_extrinsic: false
+                is_extrinsic: true
+                pos_stability: 1e-6
                 sim_params:
-                    pos_error: [0.0, 0.0, 0.0]
+                    pos_error: [0.01, 0.01, 0.01]
                     ang_error: [0.0, 0.0, 0.0]
 
         tracker:
