+++ eval/tab_msckf.py
@@ -47,25 +47,13 @@ class tab_msckf:
             y_axis_label='Position [m]',
             title='Camera Position')
         for msckf_df in self.msckf_dfs:
-            t_cam = msckf_df['time']
-            fig.line(
-                t_cam,
-                msckf_df['cam_pos_0'],
-                alpha=self.alpha,
-                color=self.colors[0],
-                legend_label='X')
-            fig.line(
-                t_cam,
-                msckf_df['cam_pos_1'],
-                alpha=self.alpha,
-                color=self.colors[1],
-                legend_label='Y')
-            fig.line(
-                t_cam,
-                msckf_df['cam_pos_2'],
-                alpha=self.alpha,
-                color=self.colors[2],
-                legend_label='Z')
+            time = msckf_df['time']
+            pos_x = msckf_df['cam_pos_0']
+            pos_y = msckf_df['cam_pos_1']
+            pos_z = msckf_df['cam_pos_2']
+            fig.line(time, pos_x, alpha=self.alpha, color=self.colors[0], legend_label='X')
+            fig.line(time, pos_y, alpha=self.alpha, color=self.colors[1], legend_label='Y')
+            fig.line(time, pos_z, alpha=self.alpha, color=self.colors[2], legend_label='Z')
         return fig
 
     def plot_cam_ang(self):
@@ -77,25 +65,13 @@ class tab_msckf:
             y_axis_label='Orientation',
             title='Camera Orientation')
         for msckf_df in self.msckf_dfs:
-            t_cam = msckf_df['time']
-            fig.line(
-                t_cam,
-                msckf_df['cam_ang_pos_0'],
-                alpha=self.alpha,
-                color=self.colors[0],
-                legend_label='X')
-            fig.line(
-                t_cam,
-                msckf_df['cam_ang_pos_1'],
-                alpha=self.alpha,
-                color=self.colors[1],
-                legend_label='Y')
-            fig.line(
-                t_cam,
-                msckf_df['cam_ang_pos_2'],
-                alpha=self.alpha,
-                color=self.colors[2],
-                legend_label='Z')
+            time = msckf_df['time']
+            ang_x = msckf_df['cam_ang_pos_0']
+            ang_y = msckf_df['cam_ang_pos_1']
+            ang_z = msckf_df['cam_ang_pos_2']
+            fig.line(time, ang_x, alpha=self.alpha, color=self.colors[0], legend_label='X')
+            fig.line(time, ang_y, alpha=self.alpha, color=self.colors[1], legend_label='Y')
+            fig.line(time, ang_z, alpha=self.alpha, color=self.colors[2], legend_label='Z')
         return fig
 
     def plot_cam_pos_err(self):
@@ -112,33 +88,18 @@ class tab_msckf:
             true_p1 = body_truth[f"cam_pos_{msckf_df.attrs['id']}_1"]
             true_p2 = body_truth[f"cam_pos_{msckf_df.attrs['id']}_2"]
 
-            t_gps = msckf_df['time']
+            time = msckf_df['time']
             est_p0 = msckf_df['cam_pos_0']
             est_p1 = msckf_df['cam_pos_1']
             est_p2 = msckf_df['cam_pos_2']
 
-            err_pos_0 = np.array(interpolate_error(true_t, true_p0, t_gps, est_p0)) * 1e3
-            err_pos_1 = np.array(interpolate_error(true_t, true_p1, t_gps, est_p1)) * 1e3
-            err_pos_2 = np.array(interpolate_error(true_t, true_p2, t_gps, est_p2)) * 1e3
+            err_pos_0 = np.array(interpolate_error(true_t, true_p0, time, est_p0)) * 1e3
+            err_pos_1 = np.array(interpolate_error(true_t, true_p1, time, est_p1)) * 1e3
[…]
