### 2025-03-09 unmannedlab/ekf_cal#1 @gareth-cross
Build instructions in documentation compile in debug mode
This issue is part of the JOSS review for ekf_cal: https://github.com/openjournals/joss-reviews/issues/7793

The current documentation recommends building with:
```bash
colcon build --symlink-install --packages-select ekf_cal \
 --event-handlers console_cohesion+ \
 --cmake-args -DCMAKE_C_FLAGS='--coverage' -DCMAKE_CXX_FLAGS='--coverage'
```

By default, this leaves `CMAKE_BUILD_TYPE` unspecified, which produced a debug build (https://github.com/colcon/colcon-cmake/issues/96) when I ran it in the provided dev container. As a result, the example test script `eval/run.py` took longer to run than necessary.

This is just a suggestion, but I expect the configuration that would be most useful to the average user would be:
- Build in either `RelWithDebInfo` or `Release` mode.
- Do not include code coverage by default. (Code coverage is fantastic, but most consumers will not need that information).

So the suggestion would be to update the docs to:
```bash
colcon build --symlink-install --packages-select ekf_cal \
 --event-handlers console_cohesion+ \
 --cmake-args -DCMAKE_BUILD_TYPE=Release
```
And provide the current instructions as an alternative for those users that intend to actively develop the package.
