### 2025-03-09 unmannedlab/ekf_cal#2 @gareth-cross
Dev container is missing `bokeh` and `scipy` (required for generating example plots)
This issue is part of the JOSS review for ekf_cal: https://github.com/openjournals/joss-reviews/issues/7793

This is a suggestion to improve the experience for users trying out the dev container. The python packages `bokeh` and `scipy` are required by `eval/report.py`, but are not specified in the docker file.
