### 2025-03-17 #7793 @gareth-cross
@JHartzer

Regarding the following checklist item:
> References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

I suggest adding citations for the following works that are directly referenced or utilized in `ekf_cal`:

- The MSCKF [paper](https://intra.ece.ucr.edu/~mourikis/papers/MourikisRoumeliotis-ICRA07.pdf).

- OpenCV ([Citing OpenCV](https://github.com/opencv/opencv/wiki/CiteOpenCV)).
