### 2025-03-17 #7793 @gareth-cross
@arfon, I have completed my review and highlighted one issue above.

---

@JHartzer Congratulations on completing this project. The following are some suggestions regarding how you might improve `ekf_cal`. These are not directly required to the JOSS review process, but might be useful for your future projects.

### Plots

I definitely appreciate the inclusion of plotting tools with the framework. One plot that is absent is [Normalized Estimation Error Squared](https://kalman-filter.com/normalized-estimation-error-squared/). NEES is a measure of filter consistency, and is useful for assessing whether the filter produces uncertainty estimates that reflect the true underlying error. In a paper focused on filter design, I would typically expect to see a plot of NEES ([an example work](https://link.springer.com/content/pdf/10.1007/s11263-011-0492-5.pdf) that compares different filter parameterizations for a particular problem domain).

Fortunately, `ekf_cal` already produces the prerequisite quantities for calculating NEES:
- Marginal uncertainty of the filter estimates, and
- Error with respect to ground truth over many trials.

So it would not be too difficult to add in future.

### Documentation

The inclusion of an example configuration file is very helpful. One way to improve on this would be to add a "Getting Started" section to the documentation. I would aim to answer the following questions:
- How can I configure `ekf_cal` to match my sensor configuration?
- What parameters might I need to adjust to run the filter successfully?
- What outputs should I check when running the ROS node in order to validate that it is running as expected?

This would also help to clarify how you expect users to interact with the framework. Would the typical consumer run this as a self-contained ROS node, or are they more likely to directly link to the C++ library itself?

### Jacobian test coverage

One property I look for in a framework like this is the level of test coverage on core mathematical components. This type of code is tricky to get right - and easy to subtly break later. A pattern worth adopting is the use of numerical differentiation to validate prediction/measurement Jacobians. For example, GTSAM uses [numerical differentiation to unit test many of the built-in handwritten Jacobians](https://github.com/borglab/gtsam/blob/f401a9056ed79d0a96a1b64f4cd01252db33afc8/gtsam/geometry/tests/testCalibratedCamera.cpp#L108C18-L108C39).

### Code quality

Overall I am heartened to see that you adopted a style guide and formatted code in a consistent fashion throughout the project. I did notice a couple of patterns in the C++ that could be improved upon.

**Argument passing style:**

Many arguments in the C++ code are presently passed by value, including dynamically allocated types such as `std::string` or `Eigen::MatrixXd`. For example:
```cpp
void RegisterCamera(unsigned int cam_id, CamState cam_state, Eigen::MatrixXd covariance)
```
Or,
```cpp
void RateLimitedLog(std::string message, double time)
```
This pattern will force additional allocations and copies in places where they could be avoided. In keeping with the C++ core guidelines, I suggest [passing such types by const-reference](https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rf-in).

**`shared_ptr<T>` as an argument type:**

Generally speaking, `shared_ptr<T>` indicates shared ownership of an underlying resource. Using `shared_ptr` as an argument type suggests that the function will take ownership of the argument (thereby incrementing the reference count). Functions that do not need to do so should [avoid accepting `shared_ptr`](https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rr-sharedptrparam-owner).

For example, the `KalmanUpdater` function:
```cpp
void KalmanUpdate(
    std::shared_ptr<EKF> ekf,
    const Eigen::MatrixXd & jacobian,
    const Eigen::VectorXd & residual,
    Eigen::MatrixXd & measurement_noise)
```
will mutate the argument `ekf` by way of the `shared_ptr`, but at no point does it copy the shared pointer. Instead, you can accept `EKF&` directly here which achieves two things:
- It is clearer from the function signature that `KalmanUpdate` does not take ownership of `ekf`.
- We eliminate the possibility of passing an empty `shared_ptr<EKF>`, thus avoiding a bug.
