### 2025-03-21 ddd5862
Minor paper updates
