+++ docs/paper/paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 The increasing complexity of autonomous systems operating in challenging environments demands robust and accurate calibration methods. Modern autonomous systems use multiple types of sensors to handle a wide variety of environments. To improve system robustness in these challenging environments, redundant sensors are often used to protect against erroneous sensor measurements or outright sensor failures. However, the underlying navigation system must be designed in such a way as to handle these duplicate measurement streams and the addition of each sensor introduces parameters that must be properly calibrated before these sensors' measurements can be effectively used. Moreover, these underlying calibration parameters can change, and often must be estimated *online*. Kalman filters are a common and near-optimal method to estimate a system's state given a series of measurements and transitions through time. Improper tuning can lead to filter instability and poor performance . As such, it is important to verify a filters performance and stability across a large set of runs in the desired operating environment with various initial errors in state estimates.
 
-This software package, `ekf_cal`, seeks to provide a framework for testing mutli-sensor Kalman filters using Monte Carlo techniques. By running a given sensor configuration through a large sample of runs, each with varying initial conditions and errors, stability can be verified. By looking and the resulting error distributions in any of the system states, biases can be eliminated and the expected system error can be estimated. In addition to the simulation running framework, additional scripts are provided that calculate the distribution of errors for a Monte Carlo and generate a report is showing state values and errors across the run to aid in analysis.
+This software package, `ekf_cal`, seeks to provide a framework for testing multi-sensor Kalman filters using Monte Carlo techniques. By running a given sensor configuration through a large sample of runs, each with varying initial conditions and errors, stability is verified. By looking at the resulting error distributions in any of the system states, the expected system bias and error are estimated. In addition to the simulation running framework, additional scripts are provided that calculate the distribution of errors for a Monte Carlo and generate a report showing state values and errors across the run to aid in analysis.
 
 Additionally, this software package provides example implementations of state-of-the-art Kalman filter-based calibration routines and is currently used for ongoing research [@2022_Multi_Cam] and [@2023_Multi_IMU].
 
@@ -33,7 +33,7 @@ Additionally, this software package provides example implementations of state-of
 
 Currently, there are not many software packages for prototyping and evaluating Kalman Filter-based calibration and localization algorithms. The most popular, which also includes a simulation architecture, is Open-VINS [@Geneva2020ICRA], which makes use of interpolated B-splines to define true motion data. Inertial measurements are sampled from the second derivative of the spline path, and bearing camera measurements are generated using global fixed features which are projected onto the camera sensor frame. The primary limitation of Open-VINS is the limited types of sensors available and the constraint to a single IMU for filtering.
 
-There exist other batch-based (offline) calibration packages, such as the Matlab Camera Calibration Toolbox [@matlab-cam-calib] for mono and stereo camera setups, where the primary limitations are the limited sensor types and closed-source codebase. Additionally, there are batch-based optimizers such as Kalibr [@Rehder] or OpenCalib [@opencalib] that provide tools for performing calibration of multi-sensor systems, but do not provide these calibrations online. These software applications are therefore susceptible to changing calibration parameters, such as drifts or shifts in mounting.
[…]
