+++ config/example.yaml
@@ -131,7 +131,7 @@
                     0.1, 0.1, 0.1                   # Orientation
                     ]
                 tracker: "orb"                      # Tracker to use
-                fiducial: "charuco"                 # Fiducial to use
+                # fiducial: "charuco"                 # Fiducial to use
                 pos_stability: 1.0e-9               # Position stability
                 ang_stability: 1.0e-9               # Orientation stability
                 data_log_rate: 10.0                 # Data logging rate
