### 2025-03-21 106c528
Default readme to compile release: Fix #1
