### 2025-03-21 unmannedlab/ekf-cal#1 @JHartzer
Howdy Gareth! Thanks for the suggestion and explanation! I completely agree, especially with how much of a difference building release vs. debug makes due to the matrix operations. I'll update the primary instructions to match what you've provided
