### 2025-03-21 41b5271
Add missing python packages to dockerfile: Fix #2
