### 2025-03-21 unmannedlab/ekf-cal#2 @JHartzer
Howdy Gareth! Yes it looks like there are a few other missing packages as well (e.g. `numpy` and `pandas`). I'll get those added to the dockerfile. Thanks for the catch!
