### 2025-03-23 unmannedlab/ekf_cal#3 @JHartzer
Add citations to external packages and methods used in ekf_cal in paper
This issue was brought up by @gareth-cross during JOSS review for ekf_cal: https://github.com/openjournals/joss-reviews/issues/7793

>
> Regarding the following checklist item:
>
> > References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?
>
> I suggest adding citations for the following works that are directly referenced or utilized in ekf_cal:
>
> - The MSCKF [paper](https://intra.ece.ucr.edu/~mourikis/papers/MourikisRoumeliotis-ICRA07.pdf).
>
> - OpenCV ([Citing OpenCV](https://github.com/opencv/opencv/wiki/CiteOpenCV)).

This is a good point to do a scrub of the referenced works and ensure they are all cited.
