### 2025-03-23 unmannedlab/ekf_cal#4 @JHartzer
Include Normalized Estimation Error Squared Plots
This issue was brought up by @gareth-cross during JOSS review for ekf_cal: https://github.com/openjournals/joss-reviews/issues/7793

> One plot that is absent is [Normalized Estimation Error Squared](https://kalman-filter.com/normalized-estimation-error-squared/). NEES is a measure of filter consistency, and is useful for assessing whether the filter produces uncertainty estimates that reflect the true underlying error. In a paper focused on filter design, I would typically expect to see a plot of NEES ([an example work](https://link.springer.com/content/pdf/10.1007/s11263-011-0492-5.pdf) that compares different filter parameterizations for a particular problem domain).
>
> Fortunately, ekf_cal already produces the prerequisite quantities for calculating NEES:
>
> - Marginal uncertainty of the filter estimates, and
> - Error with respect to ground truth over many trials.
>
> So it would not be too difficult to add in future.

These plots would be very useful in `ekf_cal` and therefore should be added
