### 2025-03-23 unmannedlab/ekf_cal#5 @JHartzer
Documentation Updates for External Users
This issue was brought up by @gareth-cross during JOSS review for ekf_cal: https://github.com/openjournals/joss-reviews/issues/7793

> The inclusion of an example configuration file is very helpful. One way to improve on this would be to add a "Getting Started" section to the documentation. I would aim to answer the following questions:
>
> - How can I configure ekf_cal to match my sensor configuration?
> - What parameters might I need to adjust to run the filter successfully?
> - What outputs should I check when running the ROS node in order to validate that it is running as expected?
>
> This would also help to clarify how you expect users to interact with the framework. Would the typical consumer run this as a self-contained ROS node, or are they more likely to directly link to the C++ library itself?

These questions definitely should be addressed by the documentation.
