### 2025-03-23 unmannedlab/ekf_cal#6 @JHartzer
Improve Jacobian Test Quality
This issue was brought up by @gareth-cross during JOSS review for ekf_cal: https://github.com/openjournals/joss-reviews/issues/7793

> One property I look for in a framework like this is the level of test coverage on core mathematical components. This type of code is tricky to get right - and easy to subtly break later. A pattern worth adopting is the use of numerical differentiation to validate prediction/measurement Jacobians. For example, GTSAM uses [numerical differentiation to unit test many of the built-in handwritten Jacobians](https://github.com/borglab/gtsam/blob/f401a9056ed79d0a96a1b64f4cd01252db33afc8/gtsam/geometry/tests/testCalibratedCamera.cpp#L108C18-L108C39).

This is definitely a salient point as most of the tests currently test basic functionality and essentially are checking that the code doesn't segfault instead of the actual underlying math. The idea of using a numerical differentiation to verify the Jacobian is excellent and will be much more robust.
