### 2025-03-23 unmannedlab/ekf_cal#7 @JHartzer
Pass by const reference, where Appropriate
This issue was brought up by @gareth-cross during JOSS review for ekf_cal: https://github.com/openjournals/joss-reviews/issues/7793

> Many arguments in the C++ code are presently passed by value, including dynamically allocated types such as std::string or Eigen::MatrixXd. For example:
>
> ```c++
> void RegisterCamera(unsigned int cam_id, CamState cam_state, Eigen::MatrixXd covariance)
> ```
>
> Or,
>
> ```c++
> void RateLimitedLog(std::string message, double time)
> ```
>
> This pattern will force additional allocations and copies in places where they could be avoided. In keeping with the C++ core guidelines, I suggest [passing such types by const-reference](https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rf-in).

Overall this is a great suggestion and can greatly reduce the number of copies, especially of `Eigen::MatrixXd` which can be quite large data types.
