### 2025-03-23 unmannedlab/ekf_cal#8 @JHartzer
Pass EKF by reference instead of shared pointers to helper functions
This issue was brought up by @gareth-cross during JOSS review for ekf_cal: https://github.com/openjournals/joss-reviews/issues/7793

> Generally speaking, `shared_ptr<T>` indicates shared ownership of an underlying resource. Using `shared_ptr` as an argument type suggests that the function will take ownership of the argument (thereby incrementing the reference count). Functions that do not need to do so should [avoid accepting shared_ptr](https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#Rr-sharedptrparam-owner).
>
> For example, the `KalmanUpdater` function:
>
> ```c++
> void KalmanUpdate(
>     std::shared_ptr<EKF> ekf,
>     const Eigen::MatrixXd & jacobian,
>     const Eigen::VectorXd & residual,
>     Eigen::MatrixXd & measurement_noise)
> ```
>
> will mutate the argument `ekf` by way of the `shared_ptr,` but at no point does it copy the shared pointer. Instead, you can accept `EKF&` directly here which achieves two things:
>
> - It is clearer from the function signature that `KalmanUpdate` does not take ownership of `ekf`.
> - We eliminate the possibility of passing an empty `shared_ptr<EKF>`, thus avoiding a bug.

This is an excellent point and should be implemented across any function that accepts an `EKF` object as input.
