### 2025-03-23 #7793 @JHartzer
Thank you so much @gareth-cross for your review and suggestions! I've created issues to track progress on the paper updates and software improvements you brought up. I'll work the paper updates first.
