### 2025-03-23 1cfc8e9
Add additional paper citations: Fixes #3
