+++ docs/paper/paper.md
@@ -25,9 +25,7 @@ bibliography: paper.bib
 
 The increasing complexity of autonomous systems operating in challenging environments demands robust and accurate calibration methods. Modern autonomous systems use multiple types of sensors to handle a wide variety of environments. To improve system robustness in these challenging environments, redundant sensors are often used to protect against erroneous sensor measurements or outright sensor failures. However, the underlying navigation system must be designed in such a way as to handle these duplicate measurement streams and the addition of each sensor introduces parameters that must be properly calibrated before these sensors' measurements can be effectively used. Moreover, these underlying calibration parameters can change, and often must be estimated *online*. Kalman filters are a common and near-optimal method to estimate a system's state given a series of measurements and transitions through time. Improper tuning can lead to filter instability and poor performance . As such, it is important to verify a filters performance and stability across a large set of runs in the desired operating environment with various initial errors in state estimates.
 
-This software package, `ekf_cal`, seeks to provide a framework for testing multi-sensor Kalman filters using Monte Carlo techniques. By running a given sensor configuration through a large sample of runs, each with varying initial conditions and errors, stability is verified. By looking at the resulting error distributions in any of the system states, the expected system bias and error are estimated. In addition to the simulation running framework, additional scripts are provided that calculate the distribution of errors for a Monte Carlo and generate a report showing state values and errors across the run to aid in analysis.
-
-Additionally, this software package provides example implementations of state-of-the-art Kalman filter-based calibration routines and is currently used for ongoing research [@2022_Multi_Cam] and [@2023_Multi_IMU].
+This software package, `ekf_cal`, seeks to provide a framework for testing multi-sensor Kalman filters using Monte Carlo techniques. By running a given sensor configuration through a large sample of runs, each with varying initial conditions and errors, stability is verified. By looking at the resulting error distributions in any of the system states, the expected system bias and error are estimated. In addition to the simulation running framework, additional scripts are provided that calculate the distribution of errors for a Monte Carlo and generate a report showing state values and errors across the run to aid in analysis. The Multi-State Constraint Kalman Filter (MSCKF) formulation used in this work was originally proposed proposed in [@MSCKF]. Additionally, this work heavily utilizes the Eigen [@eigen] and OpenCV libraries [@opencv]. Lastly, this software package provides example implementations of state-of-the-art Kalman filter-based calibration routines and is currently used for ongoing research [@2022_Multi_Cam] and [@2023_Multi_IMU].
 
 # Statement of need
 
@@ -39,11 +37,12 @@ The `ekf_cal` package, in contrast, seeks to provide additional avenues for test
 
 In summary, this package
 
-- Provides examples of the filtering techniques outlined in [@2022_Multi_Cam] and [@2023_Multi_IMU]
+- Provides examples of the filtering techniques outlined in @2022_Multi_Cam and @2023_Multi_IMU
 - Provides a Monte Carlo architecture for testing filter-based calibration techniques
 - Provides analysis tools for plotting results and evaluating performance statistics
 
 # Capabilities
+
 The `ekf_cal` package supports any number or combination of the sensors listed in the following sections. Errors in sensor measurements and calibrations are varied across Monte Carlo runs, which allow for more robust testing of calibration and localization algorithms. The key parameters supported by all sensors are
 
 | Sensor Parameter   | Description                   |
@@ -56,6 +55,7 @@ The `ekf_cal` package supports any number or combination of the sensors listed i
 | `output_directory` | Data logging output directory |
 
 ### IMU
+
[…]
