+++ docs/paper/paper.bib
@@ -14,7 +14,14 @@
   year      = {2023},
   doi       = {10.1109/SDF-MFI59545.2023.10361310},
 }
-  
+
+@misc{eigen,
+  author       = {Ga\"{e}l Guennebaud and Beno\^{i}t Jacob and others},
+  title        = {Eigen v3},
+  howpublished = {http://eigen.tuxfamily.org},
+  year         = {2010},
+}
+
 @inproceedings{Geneva2020ICRA,
   title     = {{OpenVINS}: A Research Platform for Visual-Inertial Estimation},
   author    = {Patrick Geneva and Kevin Eckenhoff and Woosik Lee and Yulin Yang and Guoquan Huang},
@@ -34,12 +41,35 @@
   title     = {Computer Vision Toolbox version: 24.1 (R2022b)},
 }
 
+@inproceedings{MSCKF,
+  author    = {Mourikis, Anastasios I. and Roumeliotis, Stergios I.},
+  booktitle = {Proceedings 2007 IEEE International Conference on Robotics and Automation},
+  title     = {A Multi-State Constraint Kalman Filter for Vision-aided Inertial Navigation},
+  year      = {2007},
+  volume    = {},
+  number    = {},
+  pages     = {3565-3572},
+  keywords  = {Inertial navigation;Cameras;Simultaneous localization and mapping;Computational complexity;Solid modeling;Position measurement;Vectors;Large-scale systems;Motion measurement;Motion estimation},
+  doi       = {10.1109/ROBOT.2007.364024},
+}
+
 @article{opencalib,
   title   = {OpenCalib: A Multi-sensor Calibration Toolbox for Autonomous Driving},
   author  = {Yan, Guohang and Liu, Zhuochun and Wang, Chengjie and Shi, Chunlei and Wei, Pengjin and Cai, Xinyu and Ma, Tao and Liu, Zhizheng and Zhong, Zebin and Liu, Yuqian and Zhao, Ming and Ma, Zheng and Li, Yikang},
   journal = {arXiv preprint arXiv:2205.14087},
   year    = {2022},
   doi     = {10.48550/arXiv.2205.14087},
+}
+
+@article{opencv,
+  author               = {Bradski, G.},
+  citeulike-article-id = {2236121},
+  journal              = {Dr. Dobb's Journal of Software Tools},
+  keywords             = {bibtex-import},
+  posted-at            = {2008-01-15 19:21:54},
+  priority             = {4},
+  title                = {{The OpenCV Library}},
+  year                 = {2000},
 }
 
 @inproceedings{Rehder,
