### 2025-04-02 5f8ea56
Pass EKF by reference for functions: Fix #8
