### 2025-04-02 ff258ac
Pass strings by const ref
