+++ src/infrastructure/ros/ros_debug_logger.cpp
@@ -45,7 +45,7 @@ void DebugLogger::SetLogLevel(unsigned int level)
   m_log_level = static_cast<LogLevel>(level);
 }
 
-void DebugLogger::Log(LogLevel level, std::string message)
+void DebugLogger::Log(LogLevel level, const std::string & message)
 {
   const char * message_c_str = message.c_str();
   if (m_log_level >= level) {
