+++ src/ekf/update/msckf_updater.hpp
@@ -45,7 +45,7 @@ public:
   explicit MsckfUpdater(
     int cam_id,
     bool is_cam_extrinsic,
-    std::string log_file_directory,
+    const std::string & log_file_directory,
     double data_log_rate,
     double min_feat_dist,
     std::shared_ptr<DebugLogger> logger
