### 2025-04-02 bff27bc
Pass Eigen objects by reference
