### 2025-04-02 8c9263c
Pass custom types by reference: Fixes #7
