### 2025-04-08 1824476
Update ros version and python requirements
