+++ requirements.txt
@@ -4,3 +4,4 @@ numpy>=1.24
 pandas>=2.2
 PyYAML>=6.0
 scipy>=1.8
+setuptools>=78.1
