+++ CMakeLists.txt
@@ -22,7 +22,7 @@ endif()
 
 # @TODO Should this compile without ROS? Just make EKF library?
 if(NOT DEFINED ${ROS_DISTRO})
-    set(ROS_DISTRO iron)
+    set(ROS_DISTRO jazzy)
 endif()
 
 list(APPEND CMAKE_PREFIX_PATH /opt/ros/${ROS_DISTRO}/)
