### 2025-04-08 2c40738
Catching a few more pass by values
