### 2025-04-15 58887cd
Test GPS updater using numerical method
