+++ src/ekf/update/gps_updater.cpp
@@ -59,11 +59,16 @@ GpsUpdater::GpsUpdater(
 
 Eigen::MatrixXd GpsUpdater::GetMeasurementJacobian(EKF & ekf)
 {
+  if (!ekf.GetUseFirstEstimateJacobian() || m_is_first_estimate) {
+    m_pos_a_in_b = ekf.m_state.gps_states[m_id].pos_a_in_b;
+    m_is_first_estimate = false;
+  }
+
   Eigen::Quaterniond ang_b_to_l = ekf.m_state.body_state.ang_b_to_l;
 
   Eigen::MatrixXd measurement_jacobian = Eigen::MatrixXd::Zero(3, ekf.GetStateSize());
   measurement_jacobian.block<3, 3>(0, 0) = Eigen::Matrix3d::Identity(3, 3);
-  measurement_jacobian.block<3, 3>(0, 6) = -ang_b_to_l.toRotationMatrix() *
+  measurement_jacobian.block<3, 3>(0, 9) = -ang_b_to_l.toRotationMatrix() *
     SkewSymmetric(m_pos_a_in_b) * quaternion_jacobian(ang_b_to_l);
 
   if (m_is_extrinsic) {
@@ -73,17 +78,22 @@ Eigen::MatrixXd GpsUpdater::GetMeasurementJacobian(EKF & ekf)
   return measurement_jacobian;
 }
 
+Eigen::Vector3d GpsUpdater::PredictMeasurement(EKF & ekf)
+{
+  if (!ekf.GetUseFirstEstimateJacobian() || m_is_first_estimate) {
+    m_pos_a_in_b = ekf.m_state.gps_states[m_id].pos_a_in_b;
+    m_is_first_estimate = false;
+  }
+
+  return ekf.m_state.body_state.pos_b_in_l + ekf.m_state.body_state.ang_b_to_l * m_pos_a_in_b;
+}
+
 void GpsUpdater::UpdateEKF(
   EKF & ekf, double time, const Eigen::Vector3d & gps_lla, const Eigen::MatrixXd & pos_covariance)
 {
   auto t_start = std::chrono::high_resolution_clock::now();
   double local_time = ekf.CalculateLocalTime(time);
   ekf.PredictModel(local_time);
-
-  if (!ekf.GetUseFirstEstimateJacobian() || m_is_first_estimate) {
-    m_pos_a_in_b = ekf.m_state.gps_states[m_id].pos_a_in_b;
-    m_is_first_estimate = false;
-  }
 
   Eigen::Vector3d pos_a_in_l = Eigen::Vector3d::Zero();
   Eigen::Vector3d residual = Eigen::Vector3d::Zero();
@@ -101,10 +111,7 @@ void GpsUpdater::UpdateEKF(
     Eigen::Vector3d gps_enu = lla_to_enu(gps_lla, pos_e_in_g);
     pos_a_in_l = enu_to_local(gps_enu, ang_l_to_e);
 
-    Eigen::Vector3d pos_b_in_l = ekf.m_state.body_state.pos_b_in_l;
-    Eigen::Quaterniond ang_b_to_l = ekf.m_state.body_state.ang_b_to_l;
-
-    Eigen::Vector3d pos_a_in_l_hat = pos_b_in_l + ang_b_to_l * m_pos_a_in_b;
+    Eigen::Vector3d pos_a_in_l_hat = PredictMeasurement(ekf);
     residual = pos_a_in_l - pos_a_in_l_hat;
     Eigen::MatrixXd jacobian = GetMeasurementJacobian(ekf);
     KalmanUpdate(ekf, jacobian, residual, pos_covariance);
