+++ src/ekf/update/gps_updater.hpp
@@ -50,14 +50,21 @@ public:
 
   ///
   /// @brief Measurement Jacobian method
-  /// @param ekf EKF pointer
+  /// @param ekf EKF address
   /// @return Measurement Jacobian matrix
   ///
   Eigen::MatrixXd GetMeasurementJacobian(EKF & ekf);
 
   ///
+  /// @brief Predict GPS measurement
+  /// @param ekf EKF address
+  /// @return Predicted GPS measurement
+  ///
+  Eigen::Vector3d PredictMeasurement(EKF & ekf);
+
+  ///
   /// @brief EKF update method for GPS measurements
-  /// @param ekf EKF pointer
+  /// @param ekf EKF address
   /// @param time Measurement time
   /// @param gps_lla GPS measured lat-lon-alt
   /// @param pos_covariance GPS measurement covariance
@@ -70,7 +77,7 @@ public:
 
   ///
   /// @brief Update/marginalize EKF using GPS measurements used to initialize local frame
-  /// @param ekf EKF pointer
+  /// @param ekf EKF address
   ///
   void MultiUpdateEKF(EKF & ekf);
 

[…]
