### 2025-04-21 #7793 @abougouffa
Hi everyone,

I've finished the review check list, I'm very sorry for taking so much time.

First, congratulation of the work you provided, the package is original and really useful (I will certainly use it in the future). The code is well written, uses a persistent coding style and formatting, the same applies to the paper too.

Also, providing a Dev Container & Docker configs made it easy to set up a test environment and get `ekf_cal` to work, thanks for that!

I have only a few minor remarks:

1. I had the same comments on passing some parameters as const references, I saw that @gareth-cross mentioned them and @JHartzer already addressed some of them in https://github.com/unmannedlab/ekf_cal/issues/7. However, I think there are a few other instances where const references/shared pointers can be used. For example, in `GPS::Callback`, `IMU::Callback`, etc., the shared pointer of the message can be marked as const since it isn't modified inside the callback.

```cpp
// ----------------------
//                      |
//                      v
void GPS::Callback(/* const */  std::shared_ptr<GpsMessage> gps_message)
{
  ...
}
```

2. There are a few implicit conversions from `float` to `double` and vice versa, from `unsigned int` to `int` (which can be problematic in some cases), and so on. It would be good to have explicit casts when comparing or combining variables of different types. I suggest enabling `-Wconversion -Wsign-conversion -Wuseless-cast -Wdouble-promotion` in GCC/LLVM to catch and fix them.

3. In some functions, you are shadowing variable names inside nested scopes, namely, the `i` variable, for example, in `src/infrastructure/sim/truth_engine.cpp`. The inner loops are using the same variable name `i`, shadowing the outer loop's one, which works, but it is a bit confusing! You catch this kind of issues by enabling `-Wshadow` in GCC/LLVM.

```cpp
  for (unsigned int i = 0; i < num_measurements; ++i) {
    ....
    for (unsigned int i = 0; i < m_imu_pos.size(); ++i) {
      ++sensor_count;
      msg << VectorToCommaString(GetImuPosition(sensor_count));
      ...
    }
    for (unsigned int i = 0; i < m_cam_pos.size(); ++i) {
      ++sensor_count;
      msg << VectorToCommaString(GetCameraPosition(sensor_count));
      ...
    }
    for (unsigned int i = 0; i < m_gps_pos.size(); ++i) {
      ++sensor_count;
      msg << VectorToCommaString(GetGpsPosition(sensor_count));
    }
    ...

    truth_logger.Log(msg.str());
  }
```

4. I agree with @gareth-cross on the documentation remark, it is a bit unclear how to add custom sensors, although it can be deduced from studying the code base.

I reiterate my apologies for taking so long to reply.
