### 2025-04-23 unmannedlab/ekf_cal#9 @JHartzer
Additional Const References/Shared Pointers
This issue was brought up by @abougouffa during JOSS review for ekf_cal: https://github.com/openjournals/joss-reviews/issues/7793

> I had the same comments on passing some parameters as const references, I saw that @gareth-cross mentioned them and @JHartzer already addressed some of them in https://github.com/unmannedlab/ekf_cal/issues/7. However, I think there are a few other instances where const references/shared pointers can be used. For example, in GPS::Callback, IMU::Callback, etc., the shared pointer of the message can be marked as const since it isn't modified inside the callback.
>
> ```c++
> // ----------------------
> //                      |
> //                      v
> void GPS::Callback(/* const */  std::shared_ptr<GpsMessage> gps_message)
> {
>   ...
> }
> ```

It's definitely worth another scrub of the code to ensure everything that is unchanging should be marked const for clarity.
