### 2025-04-23 unmannedlab/ekf_cal#10 @JHartzer
Make Conversions Explicit
This issue was brought up by @abougouffa during JOSS review for ekf_cal: https://github.com/openjournals/joss-reviews/issues/7793

> There are a few implicit conversions from float to double and vice versa, from unsigned int to int (which can be problematic in some cases), and so on. It would be good to have explicit casts when comparing or combining variables of different types. I suggest enabling `-Wconversion -Wsign-conversion -Wuseless-cast -Wdouble-promotion` in GCC/LLVM to catch and fix them.

This was a good learning moment for me! These types of conversions should be explicit at least for clarity but also for correctness.
