### 2025-04-23 unmannedlab/ekf_cal#11 @JHartzer
Rename Shadowed Variables
This issue was brought up by @abougouffa during JOSS review for ekf_cal: https://github.com/openjournals/joss-reviews/issues/7793

> In some functions, you are shadowing variable names inside nested scopes, namely, the `i` variable, for example, in `src/infrastructure/sim/truth_engine.cpp`. The inner loops are using the same variable name `i`, shadowing the outer loop's one, which works, but it is a bit confusing! You catch this kind of issues by enabling `-Wshadow` in GCC/LLVM.
>
> ```c++
>   for (unsigned int i = 0; i < num_measurements; ++i) {
>     ....
>     for (unsigned int i = 0; i < m_imu_pos.size(); ++i) {
>       ++sensor_count;
>       msg << VectorToCommaString(GetImuPosition(sensor_count));
>       ...
>     }
>     for (unsigned int i = 0; i < m_cam_pos.size(); ++i) {
>       ++sensor_count;
>       msg << VectorToCommaString(GetCameraPosition(sensor_count));
>       ...
>     }
>     for (unsigned int i = 0; i < m_gps_pos.size(); ++i) {
>       ++sensor_count;
>       msg << VectorToCommaString(GetGpsPosition(sensor_count));
>     }
>     ...
>
>     truth_logger.Log(msg.str());
>   }
> ```

Another good learning moment for me! The code should be scrubbed for shadowed variables.
