### 2025-04-23 095ef59
Rename shadowed variables. Fix #11
