+++ src/infrastructure/sim/truth_engine.cpp
@@ -171,7 +171,7 @@ void TruthEngine::GenerateGridFeatures()
 
 std::vector<cv::Point3d> TruthEngine::GenerateVisibleFeatures(
   double time,
-  int camera_id,
+  unsigned int camera_id,
   unsigned int new_feature_count,
   SimRNG rng
 )
@@ -268,19 +268,19 @@ void TruthEngine::WriteTruthData(
     msg << QuaternionToCommaString(GetBodyAngularPosition(time));
     msg << VectorToCommaString(GetBodyAngularRate(time));
     msg << VectorToCommaString(GetBodyAngularAcceleration(time));
-    for (unsigned int i = 0; i < m_imu_pos.size(); ++i) {
+    for (unsigned int j = 0; j < m_imu_pos.size(); ++j) {
       ++sensor_count;
       msg << VectorToCommaString(GetImuPosition(sensor_count));
       msg << QuaternionToCommaString(GetImuAngularPosition(sensor_count));
       msg << VectorToCommaString(GetImuAccelerometerBias(sensor_count));
       msg << VectorToCommaString(GetImuGyroscopeBias(sensor_count));
     }
-    for (unsigned int i = 0; i < m_cam_pos.size(); ++i) {
+    for (unsigned int j = 0; j < m_cam_pos.size(); ++j) {
       ++sensor_count;
       msg << VectorToCommaString(GetCameraPosition(sensor_count));
       msg << QuaternionToCommaString(GetCameraAngularPosition(sensor_count));
     }
-    for (unsigned int i = 0; i < m_gps_pos.size(); ++i) {
+    for (unsigned int j = 0; j < m_gps_pos.size(); ++j) {
       ++sensor_count;
       msg << VectorToCommaString(GetGpsPosition(sensor_count));
     }
