### 2025-04-23 59ff8fb
Don't implicitly convert float to double
