+++ src/ekf/update/fiducial_updater.hpp
@@ -44,7 +44,7 @@ public:
   ///
   explicit FiducialUpdater(
     int fiducial_id,
-    int camera_id,
+    unsigned int camera_id,
     bool is_cam_extrinsic,
     const std::string & log_file_directory,
     double data_log_rate,
