### 2025-04-23 610f16b
Resolve some of the implicit type conversions
