+++ src/sensors/sim/test/sim_camera_test.cpp
@@ -49,7 +49,7 @@ TEST(test_SimCamera, feature_track) {
   );
 
   SimRNG rng;
-  rng.SetSeed(1.0);
+  rng.SetSeed(1);
 
   Intrinsics intrinsics;
   intrinsics.f_x = 0.01;
@@ -120,7 +120,7 @@ TEST(test_SimCamera, fiducial_track) {
   );
 
   SimRNG rng;
-  rng.SetSeed(1.0);
+  rng.SetSeed(1);
 
   Intrinsics intrinsics;
   intrinsics.f_x = 0.01;
