### 2025-04-25 d4c2151
Make conversions explicit: Fix #10
