+++ src/ekf/ekf.cpp
@@ -181,7 +181,7 @@ void EKF::PredictModel(double local_time)
   auto t_end = std::chrono::high_resolution_clock::now();
   auto t_execution = std::chrono::duration_cast<std::chrono::microseconds>(t_end - t_start);
 
-  LogBodyStateIfNeeded(t_execution.count());
+  LogBodyStateIfNeeded(static_cast<int>(t_execution.count()));
 }
 
 unsigned int EKF::GetStateSize()
@@ -206,7 +206,7 @@ CamState EKF::GetCamState(unsigned int cam_id)
 
 unsigned int EKF::GetImuCount()
 {
-  return m_state.imu_states.size();
+  return static_cast<unsigned int>(m_state.imu_states.size());
 }
 
 unsigned int EKF::GetImuStateSize()
@@ -216,7 +216,7 @@ unsigned int EKF::GetImuStateSize()
 
 unsigned int EKF::GetGpsCount()
 {
-  return m_state.gps_states.size();
+  return static_cast<unsigned int>(m_state.gps_states.size());
 }
 
 unsigned int EKF::GetGpsStateSize()
@@ -231,7 +231,7 @@ unsigned int EKF::GetCamStateSize()
 
 unsigned int EKF::GetCamCount()
 {
-  return m_state.cam_states.size();
+  return static_cast<unsigned int>(m_state.cam_states.size());
 }
 
 unsigned int EKF::GetAugStateSize()
@@ -375,10 +375,10 @@ void EKF::RegisterFiducial(const FidState & fid_state, const Eigen::MatrixXd & c
 
 Eigen::MatrixXd EKF::AugmentCovariance(const Eigen::MatrixXd & in_cov, unsigned int index)
 {
-  unsigned int in_rows = in_cov.rows();
-  unsigned int in_cols = in_cov.cols();
-  unsigned int out_rows = in_cov.rows() + g_aug_state_size;
-  unsigned int out_cols = in_cov.cols() + g_aug_state_size;
+  auto in_rows = static_cast<unsigned int>(in_cov.rows());
+  auto in_cols = static_cast<unsigned int>(in_cov.cols());
+  auto out_rows = static_cast<unsigned int>(in_cov.rows() + g_aug_state_size);
+  auto out_cols = static_cast<unsigned int>(in_cov.cols() + g_aug_state_size);
 
   Eigen::MatrixXd out_cov = Eigen::MatrixXd::Zero(out_rows, out_cols);
 
@@ -474,15 +474,15 @@ void EKF::AugmentStateIfNeeded()
   }
 
   // Prune old states
-  for (int i = m_state.aug_states[0].size() - 2; i >= 0; --i) {
+  for (int i = static_cast<int>(m_state.aug_states[0].size()) - 2; i >= 0; --i) {
     // Check if any states are too old
     if (
-      ((m_current_time - m_state.aug_states[0][i + 1].time) > m_max_track_duration) &&
-      (static_cast<unsigned int>(i) < m_state.aug_states[0].size()))
+      ((m_current_time - m_state.aug_states[0][static_cast<unsigned int>(i + 1)].time) >
+      m_max_track_duration) && (static_cast<unsigned int>(i) < m_state.aug_states[0].size()))
     {
       m_state_size -= g_aug_state_size;
       m_aug_state_size -= g_aug_state_size;
-      unsigned int aug_index = m_state.aug_states[0][i].index;
+      unsigned int aug_index = m_state.aug_states[0][static_cast<unsigned int>(i)].index;
 
       // Prune old states
       m_state.aug_states[0].erase(m_state.aug_states[0].begin() + i);
[…]
