### 2025-04-25 97d9e6d
Only use smart pointers for ownership, otherwise refs
