+++ src/sensors/imu.cpp
@@ -59,17 +59,17 @@ IMU::IMU(IMU::Parameters params)
   m_ekf->RegisterIMU(m_id, imu_state, cov);
 }
 
-void IMU::Callback(std::shared_ptr<ImuMessage> imu_message)
+void IMU::Callback(ImuMessage & imu_message)
 {
   m_logger->Log(
     LogLevel::DEBUG,
-    "IMU \"" + m_name + "\" callback at time " + std::to_string(imu_message->time));
+    "IMU \"" + m_name + "\" callback at time " + std::to_string(imu_message.time));
   m_imu_updater.UpdateEKF(
     *m_ekf,
-    imu_message->time,
-    imu_message->acceleration,
-    imu_message->acceleration_covariance,
-    imu_message->angular_rate,
-    imu_message->angular_rate_covariance);
+    imu_message.time,
+    imu_message.acceleration,
+    imu_message.acceleration_covariance,
+    imu_message.angular_rate,
+    imu_message.angular_rate_covariance);
   m_logger->Log(LogLevel::DEBUG, "IMU \"" + m_name + "\" callback complete");
 }
