+++ src/sensors/camera.cpp
@@ -54,26 +54,26 @@ Camera::Camera(Camera::Parameters cam_params)
   m_ekf->RegisterCamera(m_id, cam_state, cov);
 }
 
-void Camera::Callback(std::shared_ptr<CameraMessage> camera_message)
+void Camera::Callback(CameraMessage & camera_message)
 {
   m_logger->Log(
-    LogLevel::DEBUG, "Camera " + std::to_string(camera_message->sensor_id) +
-    " callback called at time = " + std::to_string(camera_message->time));
+    LogLevel::DEBUG, "Camera " + std::to_string(camera_message.sensor_id) +
+    " callback called at time = " + std::to_string(camera_message.time));
 
-  if (!camera_message->image.empty()) {
+  if (!camera_message.image.empty()) {
     if (!m_trackers.empty() || !m_fiducials.empty()) {
-      double local_time = m_ekf->CalculateLocalTime(camera_message->time);
+      double local_time = m_ekf->CalculateLocalTime(camera_message.time);
       m_ekf->PredictModel(local_time);
 
       unsigned int frameID = GenerateFrameID();
 
       m_ekf->AugmentStateIfNeeded(m_id, frameID);
-      cv::cvtColor(camera_message->image, m_out_img, cv::COLOR_GRAY2RGB);
+      cv::cvtColor(camera_message.image, m_out_img, cv::COLOR_GRAY2RGB);
 
       if (!m_trackers.empty()) {
         for (auto const & track_iter : m_trackers) {
           m_trackers[track_iter.first]->Track(
-            camera_message->time, frameID, camera_message->image, m_out_img);
+            camera_message.time, frameID, camera_message.image, m_out_img);
           /// @todo Undistort points post track?
           // cv::undistortPoints();
         }
@@ -82,7 +82,7 @@ void Camera::Callback(std::shared_ptr<CameraMessage> camera_message)
       if (!m_fiducials.empty()) {
         for (auto const & fiducial_iter : m_fiducials) {
           m_fiducials[fiducial_iter.first]->Track(
-            camera_message->time, frameID, camera_message->image, m_out_img);
+            camera_message.time, frameID, camera_message.image, m_out_img);
         }
       }
     } else {
