### 2025-04-25 2c2667d
Static analysis fixes
