### 2025-04-25 ea7e5d4
Make callback and ekf updater parameters constant where appropriate
