+++ src/sensors/ros/ros_camera.hpp
@@ -41,7 +41,7 @@ public:
   /// @brief RosCamera callback method
   /// @param ros_camera_message ROS image message
   ///
-  void Callback(RosCameraMessage & ros_camera_message);
+  void Callback(const RosCameraMessage & ros_camera_message);
 
   ///
   /// @brief Camera output ROS image getter method
