+++ src/trackers/sim/sim_fiducial_tracker.hpp
@@ -74,14 +74,14 @@ public:
   /// @brief Return currently visible keypoints
   /// @param time Current time
   ///
-  bool IsBoardVisible(double time);
+  bool IsBoardVisible(const double time) const;
 
   ///
   /// @brief Callback for feature tracker
   /// @param time Message time
   /// @param msg Feature track message
   ///
-  void Callback(double time, SimFiducialTrackerMessage & msg);
+  void Callback(const double time, const SimFiducialTrackerMessage & msg);
 
 private:
   SimRNG m_rng;
