### 2025-04-25 a1cfc6c
Utilize const where appropriate: Fixes #9
