+++ src/infrastructure/sim/truth_engine_cyclic.hpp
@@ -99,7 +99,7 @@ public:
   /// @brief Calculate transformation matrix to convert RPY derivatives into final frame
   /// @param rpy_vector Roll-Pitch-Yaw vector
   ///
-  Eigen::Matrix3d EulerDerivativeTransform(const Eigen::Vector3d & rpy_vector);
+  Eigen::Matrix3d EulerDerivativeTransform(const Eigen::Vector3d & rpy_vector) const;
 
   ///
   /// @brief True body angular rate getter
