### 2025-04-27 baa8e47
Automate numerical jacobian in gps test
