### 2025-04-27 5ee9390
Resolve Doxygen warnings
