+++ src/infrastructure/sim/truth_engine.hpp
@@ -249,6 +249,13 @@ public:
   ///
   void GenerateGridFeatures();
 
+  ///
+  /// @brief Generate new features that are visible
+  /// @param time Current time
+  /// @param camera_id Camera ID
+  /// @param new_feature_count Number of new features to generate
+  /// @param rng Random number generator
+  ///
   std::vector<cv::Point3d> GenerateVisibleFeatures(
     double time,
     unsigned int camera_id,
