### 2025-04-27 e4dbcf4
Use integers for seeds
