+++ config/regression/fiducial.yaml
@@ -10,7 +10,7 @@
             ang_vel: 1.0e-1
             ang_acc: 1.0e-2
         sim_params:
-            seed: 1.0
+            seed: 1
             max_time: 300.0
             number_of_runs: 10
             run_number: 0
