+++ config/regression/gps.yaml
@@ -13,7 +13,7 @@
         gps_init_baseline_dist: 100.0
 
         sim_params:
-            seed: 1.0
+            seed: 1
             max_time: 300.0
             number_of_runs: 10
             run_number: 0
@@ -87,7 +87,7 @@
                 variance: [1.0, 1.0, 1.0]
                 data_log_rate: 1.0
                 pos_stability: 1.0e-4
-                is_extrinsic: true
+                is_extrinsic: false
                 sim_params:
                     lla_error: [1.0e-2, 1.0e-2, 1.0e-2]
                     pos_a_in_b_err: [1.0e-0, 1.0e-0, 1.0e-0]
