+++ config/regression/smoother.yaml
@@ -7,7 +7,7 @@
             vel: 2.0e-0
             ang_pos: 1.0e-0
         sim_params:
-            seed: 1.0
+            seed: 1
             max_time: 60.0
             number_of_runs: 20
             run_number: 0
