+++ config/regression/one_imu_cam_translation.yaml
@@ -7,7 +7,7 @@
             vel: 1.0e-1
             ang_pos: 1.0e-0
         sim_params:
-            seed: 1.0
+            seed: 1
             max_time: 10.0
             stationary_time: 0.1
             truth_type: "cyclic"
