### 2025-04-27 5eb1087
Use empty method instead of size checks
