### 2025-04-27 7c74618
Definition and declaration consistency
