+++ src/infrastructure/sim/truth_engine_cyclic.cpp
@@ -46,9 +46,10 @@ TruthEngineCyclic::TruthEngineCyclic(
 
 Eigen::Vector3d TruthEngineCyclic::GetBodyPosition(double time)
 {
+  Eigen::Vector3d body_pos;
   double relative_time = time - m_stationary_time;
   if (relative_time < 0.0) {
-    return m_pos_offset;
+    body_pos = m_pos_offset;
   } else {
     Eigen::Vector3d position;
     position[0] = m_pos_amplitude / 2.0 *
@@ -58,15 +59,17 @@ Eigen::Vector3d TruthEngineCyclic::GetBodyPosition(double time)
     position[2] = m_pos_amplitude / 2.0 *
       (1 - std::cos(m_pos_frequency[2] * 2 * M_PI * relative_time));
     position += m_pos_offset;
-    return position;
+    body_pos = position;
   }
+  return body_pos;
 }
 
 Eigen::Vector3d TruthEngineCyclic::GetBodyVelocity(double time)
 {
+  Eigen::Vector3d body_vel;
   double relative_time = time - m_stationary_time;
   if (relative_time < 0.0) {
-    return Eigen::Vector3d::Zero(3);
+    body_vel = Eigen::Vector3d::Zero(3);
   } else {
     Eigen::Vector3d velocity;
     velocity[0] = m_pos_amplitude * m_pos_frequency[0] *
@@ -75,15 +78,17 @@ Eigen::Vector3d TruthEngineCyclic::GetBodyVelocity(double time)
       M_PI * std::sin(m_pos_frequency[1] * 2 * M_PI * relative_time);
     velocity[2] = m_pos_amplitude * m_pos_frequency[2] *
       M_PI * std::sin(m_pos_frequency[2] * 2 * M_PI * relative_time);
-    return velocity;
+    body_vel = velocity;
   }
+  return body_vel;
 }
 
 Eigen::Vector3d TruthEngineCyclic::GetBodyAcceleration(double time)
 {
+  Eigen::Vector3d body_acc;
   double relative_time = time - m_stationary_time;
   if (relative_time < 0.0) {
-    return Eigen::Vector3d::Zero(3);
+    body_acc = Eigen::Vector3d::Zero(3);
   } else {
     Eigen::Vector3d acceleration;
     acceleration[0] = m_pos_amplitude * 2 * M_PI * M_PI * m_pos_frequency[0] *
@@ -92,15 +97,17 @@ Eigen::Vector3d TruthEngineCyclic::GetBodyAcceleration(double time)
       m_pos_frequency[1] * std::cos(m_pos_frequency[1] * 2 * M_PI * relative_time);
     acceleration[2] = m_pos_amplitude * 2 * M_PI * M_PI * m_pos_frequency[2] *
       m_pos_frequency[2] * std::cos(m_pos_frequency[2] * 2 * M_PI * relative_time);
-    return acceleration;
+    body_acc = acceleration;
   }
+  return body_acc;
 }
 
 Eigen::Vector3d TruthEngineCyclic::GetBodyRollPitchYaw(double time)
 {
+  Eigen::Vector3d body_rpy;
   double relative_time = time - m_stationary_time;
   if (relative_time < 0.0) {
-    return Eigen::Vector3d::Zero(3);
+    body_rpy = Eigen::Vector3d::Zero(3);
   } else {
     Eigen::Vector3d rpy_vector;
     rpy_vector[0] = m_ang_amplitude / 2.0 *
@@ -110,8 +117,9 @@ Eigen::Vector3d TruthEngineCyclic::GetBodyRollPitchYaw(double time)
[…]
