### 2025-04-28 2365c42
Make RNG static
