+++ src/application/sim/simulation.cpp
@@ -202,9 +202,8 @@ int main(int argc, char * argv[])
   unsigned int rng_seed = sim_params["seed"].as<unsigned int>(1);
   double max_time = sim_params["max_time"].as<double>(10.0);
 
-  SimRNG rng;
   if (rng_seed > 0) {
-    rng.SetSeed(rng_seed);
+    SimRNG::SetSeed(rng_seed);
   }
 
   // Truth parameters
@@ -238,7 +237,7 @@ int main(int argc, char * argv[])
     auto ang_errs = sim_params["ang_errors"].as<std::vector<double>>(def_vec);
     auto truth_engine_spline = std::make_shared<TruthEngineSmoother>(
       times, positions, angles, pos_errs, ang_errs,
-      stationary_time, max_time, 1000.0, debug_logger, rng);
+      stationary_time, max_time, 1000.0, debug_logger);
     truth_engine = std::static_pointer_cast<TruthEngine>(truth_engine_spline);
   } else if (truth_type == "spline") {
     auto positions = sim_params["positions"].as<std::vector<double>>(def_vec);
@@ -246,7 +245,7 @@ int main(int argc, char * argv[])
     auto pos_errs = sim_params["pos_errors"].as<std::vector<double>>(def_vec);
     auto ang_errs = sim_params["ang_errors"].as<std::vector<double>>(def_vec);
     auto truth_engine_spline = std::make_shared<TruthEngineSpline>(
-      positions, angles, pos_errs, ang_errs, stationary_time, max_time, debug_logger, rng);
+      positions, angles, pos_errs, ang_errs, stationary_time, max_time, debug_logger);
     truth_engine = std::static_pointer_cast<TruthEngine>(truth_engine_spline);
   } else {
     std::stringstream msg;
@@ -260,10 +259,10 @@ int main(int argc, char * argv[])
   auto ang_b_to_l_err =
     StdToEigVec(sim_params["ang_b_to_l_error"].as<std::vector<double>>(def_vec));
   BodyState initial_state;
-  initial_state.pos_b_in_l = rng.VecNormRand(ekf_params.pos_b_in_l, pos_b_in_l_err);
+  initial_state.pos_b_in_l = SimRNG::VecNormRand(ekf_params.pos_b_in_l, pos_b_in_l_err);
   initial_state.ang_b_to_l =
-    Eigen::AngleAxisd(rng.NormRand(0, ang_b_to_l_err[0]), Eigen::Vector3d::UnitX()) *
-    Eigen::AngleAxisd(rng.NormRand(0, ang_b_to_l_err[1]), Eigen::Vector3d::UnitY()) *
+    Eigen::AngleAxisd(SimRNG::NormRand(0, ang_b_to_l_err[0]), Eigen::Vector3d::UnitX()) *
+    Eigen::AngleAxisd(SimRNG::NormRand(0, ang_b_to_l_err[1]), Eigen::Vector3d::UnitY()) *
     ekf_params.ang_b_to_l;
 
   ekf->Initialize(0.0, initial_state);
@@ -274,10 +273,12 @@ int main(int argc, char * argv[])
   auto ang_l_to_e_err = sim_params["ang_l_to_e_error"].as<double>(0.0);
 
   Eigen::Vector3d pos_e_in_g_true;
-  pos_e_in_g_true(0) = rng.NormRand(ekf_params.pos_e_in_g(0), wgs84_m_to_deg(pos_e_in_g_err(0)));
-  pos_e_in_g_true(1) = rng.NormRand(ekf_params.pos_e_in_g(1), wgs84_m_to_deg(pos_e_in_g_err(1)));
[…]
