### 2025-04-29 0673336
Numerical method for fiducial updater test
