+++ src/ekf/update/fiducial_updater.cpp
@@ -63,6 +63,79 @@ FiducialUpdater::FiducialUpdater(
   if (data_log_rate != 0.0) {m_fiducial_logger.EnableLogging();}
 }
 
+void FiducialUpdater::RefreshStates(EKF & ekf)
+{
+  if (!ekf.GetUseFirstEstimateJacobian() || m_is_first_estimate) {
+    m_pos_c_in_b = ekf.m_state.cam_states[m_camera_id].pos_c_in_b;
+    m_ang_c_to_b = ekf.m_state.cam_states[m_camera_id].ang_c_to_b;
+    m_pos_f_in_l = ekf.m_state.fid_states[m_id].pos_f_in_l;
+    m_ang_f_to_l = ekf.m_state.fid_states[m_id].ang_f_to_l;
+    m_is_first_estimate = false;
+  }
+}
+
+Eigen::MatrixXd FiducialUpdater::GetMeasurementJacobian(EKF & ekf)
+{
+  RefreshStates(ekf);
+  unsigned int state_size = ekf.GetStateSize();
+  unsigned int cam_index = ekf.m_state.cam_states[m_camera_id].index;
+
+  Eigen::Vector3d pos_b_in_l = ekf.m_state.body_state.pos_b_in_l;
+  Eigen::Quaterniond ang_b_to_l = ekf.m_state.body_state.ang_b_to_l;
+  Eigen::Matrix3d rot_b_to_c = m_ang_c_to_b.toRotationMatrix().transpose();
+  Eigen::Matrix3d rot_f_to_l = m_ang_f_to_l.toRotationMatrix();
+  Eigen::Matrix3d rot_l_to_b = ang_b_to_l.toRotationMatrix().transpose();
+  Eigen::Matrix3d rot_l_to_c = rot_b_to_c * rot_l_to_b;
+
+  Eigen::MatrixXd jacobian = Eigen::MatrixXd::Zero(g_fid_measurement_size, state_size);
+
+  jacobian.block<3, 3>(0, 0) = -rot_l_to_c;
+
+  jacobian.block<3, 3>(0, 9) = rot_l_to_c *
+    SkewSymmetric(m_pos_f_in_l - pos_b_in_l) *
+    quaternion_jacobian(ang_b_to_l).transpose();
+
+  jacobian.block<3, 3>(3, 9) = -rot_l_to_c * rot_f_to_l *
+    quaternion_jacobian(ang_b_to_l).transpose();
+
+  /// @todo Test camera calibration jacobians
+  if (ekf.m_state.cam_states[m_camera_id].GetIsExtrinsic()) {
+    jacobian.block<3, 3>(0, cam_index + 0) = -rot_b_to_c;
+
+    jacobian.block<3, 3>(0, cam_index + 3) = rot_b_to_c *
+      SkewSymmetric(rot_l_to_b * (m_pos_f_in_l - pos_b_in_l) - m_pos_c_in_b) *
+      quaternion_jacobian(m_ang_c_to_b).transpose();
+
+    jacobian.block<3, 3>(3, cam_index + 3) = -rot_b_to_c *
+      quaternion_jacobian(m_ang_c_to_b).transpose() * rot_l_to_b * rot_f_to_l;
+  }
+
+  /// @todo Fiducial calibration jacobians
+
+  return jacobian;
+}
+
+Eigen::VectorXd FiducialUpdater::PredictMeasurement(EKF & ekf)
+{
+  RefreshStates(ekf);
+  Eigen::VectorXd predicted_measurement = Eigen::VectorXd::Zero(6);
+  Eigen::Quaterniond ang_b_to_l = ekf.m_state.body_state.ang_b_to_l;
+  Eigen::Vector3d pos_b_in_l = ekf.m_state.body_state.pos_b_in_l;
[…]
