### 2025-04-29 #7793 @JHartzer
Just a quick update: I believe I've addressed the good comments brought up by @abougouffa , but I will need a bit more time to finish up the numerical Jacobian and NEES plotting updates. I'm traveling the rest of this week but will try to address them next week
