### 2025-05-05 3ea8d25
Test fiducial extrinsics
