+++ src/ekf/types.cpp
@@ -96,53 +96,53 @@ State & operator+=(State & l_state, const Eigen::VectorXd & r_vector)
   Eigen::VectorXd r_body_state = r_vector.segment<g_body_state_size>(0);
   l_state.body_state += r_body_state;
 
-  unsigned int index = g_body_state_size;
+  unsigned int idx = g_body_state_size;
   for (auto & imu_iter : l_state.imu_states) {
     if (imu_iter.second.GetIsExtrinsic()) {
-      imu_iter.second.pos_i_in_b += r_vector.segment<3>(index + 0);
+      imu_iter.second.pos_i_in_b += r_vector.segment<3>(idx + 0);
       imu_iter.second.ang_i_to_b =
-        imu_iter.second.ang_i_to_b * RotVecToQuat(r_vector.segment<3>(index + 3));
-      index += g_imu_extrinsic_state_size;
+        imu_iter.second.ang_i_to_b * RotVecToQuat(r_vector.segment<3>(idx + 3));
+      idx += g_imu_extrinsic_state_size;
     }
     if (imu_iter.second.GetIsIntrinsic()) {
-      imu_iter.second.acc_bias += r_vector.segment<3>(index + 0);
-      imu_iter.second.omg_bias += r_vector.segment<3>(index + 3);
-      index += g_imu_intrinsic_state_size;
+      imu_iter.second.acc_bias += r_vector.segment<3>(idx + 0);
+      imu_iter.second.omg_bias += r_vector.segment<3>(idx + 3);
+      idx += g_imu_intrinsic_state_size;
     }
   }
 
   for (auto & gps_iter : l_state.gps_states) {
     if (gps_iter.second.GetIsExtrinsic()) {
-      gps_iter.second.pos_a_in_b += r_vector.segment<3>(index);
-      index += g_gps_extrinsic_state_size;
+      gps_iter.second.pos_a_in_b += r_vector.segment<3>(idx);
+      idx += g_gps_extrinsic_state_size;
     }
   }
 
   for (auto & cam_iter : l_state.cam_states) {
     if (cam_iter.second.GetIsExtrinsic()) {
-      cam_iter.second.pos_c_in_b += r_vector.segment<3>(index + 0);
+      cam_iter.second.pos_c_in_b += r_vector.segment<3>(idx + 0);
       cam_iter.second.ang_c_to_b =
-        cam_iter.second.ang_c_to_b * RotVecToQuat(r_vector.segment<3>(index + 3));
-      index += g_cam_extrinsic_state_size;
+        cam_iter.second.ang_c_to_b * RotVecToQuat(r_vector.segment<3>(idx + 3));
+      idx += g_cam_extrinsic_state_size;
     }
   }
 
   for (auto & fid_iter : l_state.fid_states) {
     if (fid_iter.second.GetIsExtrinsic()) {
-      fid_iter.second.pos_f_in_l += r_vector.segment<3>(index);
+      fid_iter.second.pos_f_in_l += r_vector.segment<3>(idx);
       fid_iter.second.ang_f_to_l =
-        fid_iter.second.ang_f_to_l * RotVecToQuat(r_vector.segment<3>(index + 3));
-      index += g_fid_extrinsic_state_size;
+        fid_iter.second.ang_f_to_l * RotVecToQuat(r_vector.segment<3>(idx + 3));
+      idx += g_fid_extrinsic_state_size;
     }
   }
 
   for (auto & aug_iter : l_state.aug_states) {
     unsigned int aug_id = aug_iter.first;
     for (unsigned int i = 0; i < l_state.aug_states[aug_id].size(); ++i) {
[…]
