### 2025-05-05 894ed4b
Numerical method for IMU updater test
