+++ src/ekf/update/test/imu_updater_test.cpp
@@ -23,6 +23,7 @@
 #include "ekf/ekf.hpp"
 #include "ekf/types.hpp"
 #include "ekf/update/imu_updater.hpp"
+#include "utility/custom_assertions.hpp"
 
 TEST(test_imu_updater, update) {
   EKF::Parameters ekf_params;
@@ -223,4 +224,41 @@ TEST(test_imu_updater, non_initialized_time) {
   EXPECT_EQ(state.body_state.pos_b_in_l[0], 0);
   EXPECT_EQ(state.body_state.pos_b_in_l[1], 0);
   EXPECT_EQ(state.body_state.pos_b_in_l[2], 0);
+}
+
+TEST(test_imu_updater, jacobian) {
+  unsigned int imu_id{0};
+  bool is_extrinsic{true};
+  bool is_intrinsic{true};
+  ImuState imu_state;
+  imu_state.pos_i_in_b = Eigen::Vector3d{1, 2, 3};
+  imu_state.SetIsExtrinsic(is_extrinsic);
+  imu_state.SetIsIntrinsic(is_intrinsic);
+
+  Eigen::Matrix3d covariance = Eigen::Matrix3d::Identity();
+
+  auto debug_logger = std::make_shared<DebugLogger>(LogLevel::DEBUG, "");
+  EKF::Parameters ekf_params;
+  ekf_params.debug_logger = debug_logger;
+  EKF ekf(ekf_params);
+  ekf.RegisterIMU(imu_id, imu_state, covariance);
+
+  auto imu_updater =
+    ImuUpdater(imu_id, is_extrinsic, is_intrinsic, "log_file_directory", 0.0, debug_logger);
+
+  Eigen::VectorXd base_state = ekf.m_state.ToVector();
+  Eigen::MatrixXd jac_analytical = imu_updater.GetMeasurementJacobian(ekf);
+  Eigen::VectorXd base_meas = imu_updater.PredictMeasurement(ekf);
+
+  double delta = 1.0e-6;
+  unsigned int jac_size = base_state.size();
+  Eigen::MatrixXd jac_numerical = Eigen::MatrixXd::Zero(6, jac_size);
+  for (unsigned int i = 0; i < jac_size; ++i) {
+    Eigen::VectorXd delta_vec = base_state;
+    delta_vec[i] += delta;
+    ekf.m_state.SetState(delta_vec);
+    jac_numerical.block<6, 1>(0, i) = (imu_updater.PredictMeasurement(ekf) - base_meas) / delta;
+  }
+
+  EXPECT_TRUE(EXPECT_EIGEN_NEAR(jac_analytical, jac_numerical, 1e-3));
 }
