### 2025-05-06 31a40d0
Add numerical testing for MSCKF jacobians: Fixes #6
