### 2025-05-08 f2ecaa6
Add NEES plot to body states
