+++ eval/tab_body.py
@@ -17,10 +17,11 @@
 
 
 from bokeh.layouts import layout
-from bokeh.models import TabPanel
+from bokeh.models import Range1d, Spacer, TabPanel
 from bokeh.plotting import figure
 import numpy as np
 from scipy.spatial.transform import Rotation
+from scipy.stats.distributions import chi2
 from utilities import calculate_alpha, calculate_rotation_errors, get_colors, interpolate_error, \
     lists_to_rot, plot_update_timing
 
@@ -378,6 +379,125 @@ class tab_body:
             fig.line(time, cov_z, alpha=self.alpha, color=self.colors[2], legend_label='Z')
         return fig
 
+    def plot_body_nees(self):
+        """Plot body normalized estimation error squared."""
+        fig = figure(width=800, height=300, x_axis_label='Time [s]',
+                     y_axis_label='NEES', title='Normalized Estimation Error Squared')
+        for body_state, body_truth in zip(self.body_state_dfs, self.body_truth_dfs):
+            xt  = body_state['time']
+            x00 = body_state['body_pos_0']
+            x01 = body_state['body_pos_1']
+            x02 = body_state['body_pos_2']
+            x03 = body_state['body_vel_0']
+            x04 = body_state['body_vel_1']
+            x05 = body_state['body_vel_2']
+            x06 = body_state['body_acc_0']
+            x07 = body_state['body_acc_1']
+            x08 = body_state['body_acc_2']
+            xw  = body_state['body_ang_pos_0']
+            xx  = body_state['body_ang_pos_1']
+            xy  = body_state['body_ang_pos_2']
+            xz  = body_state['body_ang_pos_3']
+            x12 = body_state['body_ang_vel_0']
+            x13 = body_state['body_ang_vel_1']
+            x14 = body_state['body_ang_vel_2']
+            x15 = body_state['body_ang_acc_0']
+            x16 = body_state['body_ang_acc_1']
+            x17 = body_state['body_ang_acc_2']
+
+            c00 = body_state['body_cov_0']
+            c01 = body_state['body_cov_1']
+            c02 = body_state['body_cov_2']
+            c03 = body_state['body_cov_3']
+            c04 = body_state['body_cov_4']
+            c05 = body_state['body_cov_5']
+            c06 = body_state['body_cov_6']
+            c07 = body_state['body_cov_7']
+            c08 = body_state['body_cov_8']
+            c09 = body_state['body_cov_9']
+            c10 = body_state['body_cov_10']
+            c11 = body_state['body_cov_11']
+            c12 = body_state['body_cov_12']
+            c13 = body_state['body_cov_13']
+            c14 = body_state['body_cov_14']
+            c15 = body_state['body_cov_15']
+            c16 = body_state['body_cov_16']
+            c17 = body_state['body_cov_17']
+
+            tt  = body_truth['time']
+            t00 = body_truth['body_pos_0']
+            t01 = body_truth['body_pos_1']
+            t02 = body_truth['body_pos_2']
+            t03 = body_truth['body_vel_0']
+            t04 = body_truth['body_vel_1']
+            t05 = body_truth['body_vel_2']
+            t06 = body_truth['body_acc_0']
+            t07 = body_truth['body_acc_1']
+            t08 = body_truth['body_acc_2']
+            tw  = body_truth['body_ang_pos_0']
+            tx  = body_truth['body_ang_pos_1']
+            ty  = body_truth['body_ang_pos_2']
+            tz  = body_truth['body_ang_pos_3']
[…]
