### 2025-05-08 20d69fd
Utilize common function for quaternion interpolation
