+++ eval/tab_imu.py
@@ -16,12 +16,12 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 from bokeh.layouts import layout
-from bokeh.models import Range1d, TabPanel
+from bokeh.models import Range1d, Spacer, TabPanel
 from bokeh.plotting import figure
 import numpy as np
-
-from utilities import calculate_alpha, get_colors, interpolate_error, interpolate_quat_error, \
-    plot_update_timing
+from scipy.stats.distributions import chi2
+from utilities import calculate_alpha, calculate_rotation_errors, get_colors, interpolate_error, \
+    interpolate_quat_error, lists_to_rot, plot_update_timing
 
 
 class tab_imu:
@@ -29,7 +29,8 @@ class tab_imu:
     def __init__(self, imu_dfs, body_truth_dfs, args):
         self.imu_dfs = imu_dfs
         self.body_truth_dfs = body_truth_dfs
-
+        self.is_extrinsic = 'imu_ext_cov_0' in self.imu_dfs[0].keys()
+        self.is_intrinsic = 'imu_int_cov_0' in self.imu_dfs[0].keys()
         self.alpha = calculate_alpha(len(self.imu_dfs))
         self.colors = get_colors(args)
 
@@ -333,21 +334,116 @@ class tab_imu:
         fig.y_range = Range1d(0, 1)
         return fig
 
+    def plot_body_nees(self):
+        """Plot body normalized estimation error squared."""
+        fig = figure(width=800, height=300, x_axis_label='Time [s]',
+                     y_axis_label='NEES', title='Normalized Estimation Error Squared')
+        for imu_df, body_truth in zip(self.imu_dfs, self.body_truth_dfs):
+            xt  = imu_df['time']
+            tt  = body_truth['time']
+            nees = np.zeros(len(xt))
+
+            if self.is_extrinsic:
+                x00 = imu_df['imu_pos_0']
+                x01 = imu_df['imu_pos_1']
+                x02 = imu_df['imu_pos_2']
+                xw  = imu_df['imu_ang_pos_0']
+                xx  = imu_df['imu_ang_pos_1']
+                xy  = imu_df['imu_ang_pos_2']
+                xz  = imu_df['imu_ang_pos_3']
+
+                c00 = imu_df['imu_ext_cov_0']
+                c01 = imu_df['imu_ext_cov_1']
+                c02 = imu_df['imu_ext_cov_2']
+                c03 = imu_df['imu_ext_cov_3']
+                c04 = imu_df['imu_ext_cov_4']
+                c05 = imu_df['imu_ext_cov_5']
+
+                t00 = body_truth[f'imu_pos_{self.imu_dfs[0].attrs['id']}_0']
+                t01 = body_truth[f'imu_pos_{self.imu_dfs[0].attrs['id']}_1']
+                t02 = body_truth[f'imu_pos_{self.imu_dfs[0].attrs['id']}_2']
+                tw  = body_truth[f'imu_ang_pos_{self.imu_dfs[0].attrs['id']}_0']
+                tx  = body_truth[f'imu_ang_pos_{self.imu_dfs[0].attrs['id']}_1']
+                ty  = body_truth[f'imu_ang_pos_{self.imu_dfs[0].attrs['id']}_2']
+                tz  = body_truth[f'imu_ang_pos_{self.imu_dfs[0].attrs['id']}_3']
+
+                e00 = interpolate_error(tt, t00, xt, x00)
+                e01 = interpolate_error(tt, t01, xt, x01)
+                e02 = interpolate_error(tt, t02, xt, x02)
+                e03, e04, e05 = interpolate_quat_error(tt, tw, tx, ty, tz, xt, xw, xx, xy, xz)
+
[…]
