### 2025-05-09 26d670a
Add MSCKF NEES plot
