### 2025-05-09 24c994b
Static analysis updates
