### 2025-05-09 6fa173f
Add Fiducial NEES plot
