### 2025-05-10 bf7de8f
Additional details on configuration: Fixes #5
