+++ eval/tab_body.py
@@ -27,8 +27,10 @@ from utilities import calculate_alpha, calculate_rotation_errors, get_colors, in
 
 
 class tab_body:
+    """Class for plotting body data."""
 
     def __init__(self, body_state_dfs, aug_state_dfs, body_truth_dfs, args):
+        """Initializes the tab_body class for plotting body state information."""
         self.body_state_dfs = body_state_dfs
         self.aug_state_dfs = aug_state_dfs
         self.body_truth_dfs = body_truth_dfs
@@ -586,6 +588,7 @@ class tab_body:
         return fig
 
     def get_tab(self):
+        """Generates the Bokeh TabPanel containing all body state plots."""
         layout_plots = [
             [
                 self.plot_body_pos(),
