### 2025-05-13 e2370b4
Make descripions PEP compliant
