### 2025-05-26 #7793 @JHartzer
Excellent! I have reviewed the paper and had no further updates.
I released version `0.7` of the software and have made an archive on Zenodo at the DOI: [10.5281/zenodo.15520959](https://doi.org/10.5281/zenodo.15520959)
The title and authorship match this submission.

Let me know if you need anything else!
