### 2025-05-29 #7793 @JHartzer
Thanks for the fix @arfon! Should be merged now
